% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dax_operations.R
\name{dax_create_subnet_group}
\alias{dax_create_subnet_group}
\title{Creates a new subnet group}
\usage{
dax_create_subnet_group(SubnetGroupName, Description, SubnetIds)
}
\arguments{
\item{SubnetGroupName}{[required] A name for the subnet group. This value is stored as a lowercase string.}

\item{Description}{A description for the subnet group}

\item{SubnetIds}{[required] A list of VPC subnet IDs for the subnet group.}
}
\description{
Creates a new subnet group.
}
\section{Request syntax}{
\preformatted{svc$create_subnet_group(
  SubnetGroupName = "string",
  Description = "string",
  SubnetIds = list(
    "string"
  )
)
}
}

\keyword{internal}
