% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dax_operations.R
\name{dax_delete_cluster}
\alias{dax_delete_cluster}
\title{Deletes a previously provisioned DAX cluster}
\usage{
dax_delete_cluster(ClusterName)
}
\arguments{
\item{ClusterName}{[required] The name of the cluster to be deleted.}
}
\description{
Deletes a previously provisioned DAX cluster. \emph{DeleteCluster} deletes all associated nodes, node endpoints and the DAX cluster itself. When you receive a successful response from this action, DAX immediately begins deleting the cluster; you cannot cancel or revert this action.

See \url{https://paws-r.github.io/docs/dax/delete_cluster.html} for full documentation.
}
\keyword{internal}
