% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dax_operations.R
\name{dax_delete_parameter_group}
\alias{dax_delete_parameter_group}
\title{Deletes the specified parameter group}
\usage{
dax_delete_parameter_group(ParameterGroupName)
}
\arguments{
\item{ParameterGroupName}{[required] The name of the parameter group to delete.}
}
\description{
Deletes the specified parameter group. You cannot delete a parameter group if it is associated with any DAX clusters.

See \url{https://www.paws-r-sdk.com/docs/dax_delete_parameter_group/} for full documentation.
}
\keyword{internal}
