% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dax_operations.R
\name{dax_delete_subnet_group}
\alias{dax_delete_subnet_group}
\title{Deletes a subnet group}
\usage{
dax_delete_subnet_group(SubnetGroupName)
}
\arguments{
\item{SubnetGroupName}{[required] The name of the subnet group to delete.}
}
\description{
Deletes a subnet group.

See \url{https://www.paws-r-sdk.com/docs/dax_delete_subnet_group/} for full documentation.
}
\keyword{internal}
