% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dax_operations.R
\name{dax_describe_parameter_groups}
\alias{dax_describe_parameter_groups}
\title{Returns a list of parameter group descriptions}
\usage{
dax_describe_parameter_groups(ParameterGroupNames, MaxResults,
  NextToken)
}
\arguments{
\item{ParameterGroupNames}{The names of the parameter groups.}

\item{MaxResults}{The maximum number of results to include in the response. If more
results exist than the specified \code{MaxResults} value, a token is included
in the response so that the remaining results can be retrieved.

The value for \code{MaxResults} must be between 20 and 100.}

\item{NextToken}{An optional token returned from a prior request. Use this token for
pagination of results from this action. If this parameter is specified,
the response includes only results beyond the token, up to the value
specified by \code{MaxResults}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  NextToken = "string",
  ParameterGroups = list(
    list(
      ParameterGroupName = "string",
      Description = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns a list of parameter group descriptions. If a parameter group
name is specified, the list will contain only the descriptions for that
group.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_parameter_groups(
  ParameterGroupNames = list(
    "string"
  ),
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
