% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dax_operations.R
\name{dax_increase_replication_factor}
\alias{dax_increase_replication_factor}
\title{Adds one or more nodes to a DAX cluster}
\usage{
dax_increase_replication_factor(
  ClusterName,
  NewReplicationFactor,
  AvailabilityZones = NULL
)
}
\arguments{
\item{ClusterName}{[required] The name of the DAX cluster that will receive additional nodes.}

\item{NewReplicationFactor}{[required] The new number of nodes for the DAX cluster.}

\item{AvailabilityZones}{The Availability Zones (AZs) in which the cluster nodes will be created.
All nodes belonging to the cluster are placed in these Availability
Zones. Use this parameter if you want to distribute the nodes across
multiple AZs.}
}
\description{
Adds one or more nodes to a DAX cluster.

See \url{https://paws-r.github.io/docs/dax/increase_replication_factor.html} for full documentation.
}
\keyword{internal}
