% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dax_operations.R
\name{dax_update_parameter_group}
\alias{dax_update_parameter_group}
\title{Modifies the parameters of a parameter group}
\usage{
dax_update_parameter_group(ParameterGroupName, ParameterNameValues)
}
\arguments{
\item{ParameterGroupName}{[required] The name of the parameter group.}

\item{ParameterNameValues}{[required] An array of name-value pairs for the parameters in the group. Each
element in the array represents a single parameter.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ParameterGroup = list(
    ParameterGroupName = "string",
    Description = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Modifies the parameters of a parameter group. You can modify up to 20
parameters in a single request by submitting a list parameter name and
value pairs.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_parameter_group(
  ParameterGroupName = "string",
  ParameterNameValues = list(
    list(
      ParameterName = "string",
      ParameterValue = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
