% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_add_tags_to_resource}
\alias{docdb_add_tags_to_resource}
\title{Adds metadata tags to an Amazon DocumentDB resource}
\usage{
docdb_add_tags_to_resource(ResourceName, Tags)
}
\arguments{
\item{ResourceName}{[required] The Amazon DocumentDB resource that the tags are added to. This value is
an Amazon Resource Name .}

\item{Tags}{[required] The tags to be assigned to the Amazon DocumentDB resource.}
}
\description{
Adds metadata tags to an Amazon DocumentDB resource. You can use these tags with cost allocation reporting to track costs that are associated with Amazon DocumentDB resources or in a \code{Condition} statement in an Identity and Access Management (IAM) policy for Amazon DocumentDB.

See \url{https://paws-r.github.io/docs/docdb/add_tags_to_resource.html} for full documentation.
}
\keyword{internal}
