% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_create_db_cluster_parameter_group}
\alias{docdb_create_db_cluster_parameter_group}
\title{Creates a new cluster parameter group}
\usage{
docdb_create_db_cluster_parameter_group(
  DBClusterParameterGroupName,
  DBParameterGroupFamily,
  Description,
  Tags = NULL
)
}
\arguments{
\item{DBClusterParameterGroupName}{[required] The name of the cluster parameter group.

Constraints:
\itemize{
\item Must not match the name of an existing \code{DBClusterParameterGroup}.
}

This value is stored as a lowercase string.}

\item{DBParameterGroupFamily}{[required] The cluster parameter group family name.}

\item{Description}{[required] The description for the cluster parameter group.}

\item{Tags}{The tags to be assigned to the cluster parameter group.}
}
\description{
Creates a new cluster parameter group.

See \url{https://paws-r.github.io/docs/docdb/create_db_cluster_parameter_group.html} for full documentation.
}
\keyword{internal}
