% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_create_db_subnet_group}
\alias{docdb_create_db_subnet_group}
\title{Creates a new subnet group}
\usage{
docdb_create_db_subnet_group(
  DBSubnetGroupName,
  DBSubnetGroupDescription,
  SubnetIds,
  Tags = NULL
)
}
\arguments{
\item{DBSubnetGroupName}{[required] The name for the subnet group. This value is stored as a lowercase
string.

Constraints: Must contain no more than 255 letters, numbers, periods,
underscores, spaces, or hyphens. Must not be default.

Example: \code{mySubnetgroup}}

\item{DBSubnetGroupDescription}{[required] The description for the subnet group.}

\item{SubnetIds}{[required] The Amazon EC2 subnet IDs for the subnet group.}

\item{Tags}{The tags to be assigned to the subnet group.}
}
\description{
Creates a new subnet group. subnet groups must contain at least one subnet in at least two Availability Zones in the Amazon Web Services Region.

See \url{https://paws-r.github.io/docs/docdb/create_db_subnet_group.html} for full documentation.
}
\keyword{internal}
