% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_create_event_subscription}
\alias{docdb_create_event_subscription}
\title{Creates an Amazon DocumentDB event notification subscription}
\usage{
docdb_create_event_subscription(SubscriptionName, SnsTopicArn,
  SourceType, EventCategories, SourceIds, Enabled, Tags)
}
\arguments{
\item{SubscriptionName}{[required] The name of the subscription.

Constraints: The name must be fewer than 255 characters.}

\item{SnsTopicArn}{[required] The Amazon Resource Name (ARN) of the SNS topic created for event
notification. Amazon SNS creates the ARN when you create a topic and
subscribe to it.}

\item{SourceType}{The type of source that is generating the events. For example, if you
want to be notified of events generated by an instance, you would set
this parameter to \code{db-instance}. If this value is not specified, all
events are returned.

Valid values: \code{db-instance}, \code{db-cluster}, \code{db-parameter-group},
\code{db-security-group}, \code{db-snapshot}, \code{db-cluster-snapshot}}

\item{EventCategories}{A list of event categories for a \code{SourceType} that you want to subscribe
to.}

\item{SourceIds}{The list of identifiers of the event sources for which events are
returned. If not specified, then all sources are included in the
response. An identifier must begin with a letter and must contain only
ASCII letters, digits, and hyphens; it can't end with a hyphen or
contain two consecutive hyphens.

Constraints:
\itemize{
\item If \code{SourceIds} are provided, \code{SourceType} must also be provided.
\item If the source type is an instance, a \code{DBInstanceIdentifier} must be
provided.
\item If the source type is a security group, a \code{DBSecurityGroupName} must
be provided.
\item If the source type is a parameter group, a \code{DBParameterGroupName}
must be provided.
\item If the source type is a snapshot, a \code{DBSnapshotIdentifier} must be
provided.
}}

\item{Enabled}{A Boolean value; set to \code{true} to activate the subscription, set to
\code{false} to create the subscription but not active it.}

\item{Tags}{The tags to be assigned to the event subscription.}
}
\value{
A list with the following syntax:\preformatted{list(
  EventSubscription = list(
    CustomerAwsId = "string",
    CustSubscriptionId = "string",
    SnsTopicArn = "string",
    Status = "string",
    SubscriptionCreationTime = "string",
    SourceType = "string",
    SourceIdsList = list(
      "string"
    ),
    EventCategoriesList = list(
      "string"
    ),
    Enabled = TRUE|FALSE,
    EventSubscriptionArn = "string"
  )
)
}
}
\description{
Creates an Amazon DocumentDB event notification subscription. This
action requires a topic Amazon Resource Name (ARN) created by using the
Amazon DocumentDB console, the Amazon SNS console, or the Amazon SNS
API. To obtain an ARN with Amazon SNS, you must create a topic in Amazon
SNS and subscribe to the topic. The ARN is displayed in the Amazon SNS
console.

You can specify the type of source (\code{SourceType}) that you want to be
notified of. You can also provide a list of Amazon DocumentDB sources
(\code{SourceIds}) that trigger the events, and you can provide a list of
event categories (\code{EventCategories}) for events that you want to be
notified of. For example, you can specify \code{SourceType = db-instance},
\verb{SourceIds = mydbinstance1, mydbinstance2} and
\verb{EventCategories = Availability, Backup}.

If you specify both the \code{SourceType} and \code{SourceIds} (such as
\code{SourceType = db-instance} and \code{SourceIdentifier = myDBInstance1}), you
are notified of all the \code{db-instance} events for the specified source.
If you specify a \code{SourceType} but do not specify a \code{SourceIdentifier},
you receive notice of the events for that source type for all your
Amazon DocumentDB sources. If you do not specify either the \code{SourceType}
or the \code{SourceIdentifier}, you are notified of events generated from all
Amazon DocumentDB sources belonging to your customer account.
}
\section{Request syntax}{
\preformatted{svc$create_event_subscription(
  SubscriptionName = "string",
  SnsTopicArn = "string",
  SourceType = "string",
  EventCategories = list(
    "string"
  ),
  SourceIds = list(
    "string"
  ),
  Enabled = TRUE|FALSE,
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
