% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_delete_db_instance}
\alias{docdb_delete_db_instance}
\title{Deletes a previously provisioned instance}
\usage{
docdb_delete_db_instance(DBInstanceIdentifier)
}
\arguments{
\item{DBInstanceIdentifier}{[required] The instance identifier for the instance to be deleted. This parameter
isn't case sensitive.

Constraints:
\itemize{
\item Must match the name of an existing instance.
}}
}
\description{
Deletes a previously provisioned instance.

See \url{https://www.paws-r-sdk.com/docs/docdb_delete_db_instance/} for full documentation.
}
\keyword{internal}
