% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_delete_db_subnet_group}
\alias{docdb_delete_db_subnet_group}
\title{Deletes a subnet group}
\usage{
docdb_delete_db_subnet_group(DBSubnetGroupName)
}
\arguments{
\item{DBSubnetGroupName}{[required] The name of the database subnet group to delete.

You can't delete the default subnet group.

Constraints:

Must match the name of an existing \code{DBSubnetGroup}. Must not be default.

Example: \code{mySubnetgroup}}
}
\description{
Deletes a subnet group.

See \url{https://paws-r.github.io/docs/docdb/delete_db_subnet_group.html} for full documentation.
}
\keyword{internal}
