% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_delete_event_subscription}
\alias{docdb_delete_event_subscription}
\title{Deletes an Amazon DocumentDB event notification subscription}
\usage{
docdb_delete_event_subscription(SubscriptionName)
}
\arguments{
\item{SubscriptionName}{[required] The name of the Amazon DocumentDB event notification subscription that
you want to delete.}
}
\value{
A list with the following syntax:\preformatted{list(
  EventSubscription = list(
    CustomerAwsId = "string",
    CustSubscriptionId = "string",
    SnsTopicArn = "string",
    Status = "string",
    SubscriptionCreationTime = "string",
    SourceType = "string",
    SourceIdsList = list(
      "string"
    ),
    EventCategoriesList = list(
      "string"
    ),
    Enabled = TRUE|FALSE,
    EventSubscriptionArn = "string"
  )
)
}
}
\description{
Deletes an Amazon DocumentDB event notification subscription.
}
\section{Request syntax}{
\preformatted{svc$delete_event_subscription(
  SubscriptionName = "string"
)
}
}

\keyword{internal}
