% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_delete_global_cluster}
\alias{docdb_delete_global_cluster}
\title{Deletes a global cluster}
\usage{
docdb_delete_global_cluster(GlobalClusterIdentifier)
}
\arguments{
\item{GlobalClusterIdentifier}{[required] The cluster identifier of the global cluster being deleted.}
}
\description{
Deletes a global cluster. The primary and secondary clusters must already be detached or deleted before attempting to delete a global cluster.

See \url{https://www.paws-r-sdk.com/docs/docdb_delete_global_cluster/} for full documentation.
}
\keyword{internal}
