% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_describe_db_cluster_snapshot_attributes}
\alias{docdb_describe_db_cluster_snapshot_attributes}
\title{Returns a list of cluster snapshot attribute names and values for a
manual DB cluster snapshot}
\usage{
docdb_describe_db_cluster_snapshot_attributes(DBClusterSnapshotIdentifier)
}
\arguments{
\item{DBClusterSnapshotIdentifier}{[required] The identifier for the cluster snapshot to describe the attributes for.}
}
\description{
Returns a list of cluster snapshot attribute names and values for a manual DB cluster snapshot.

See \url{https://www.paws-r-sdk.com/docs/docdb_describe_db_cluster_snapshot_attributes/} for full documentation.
}
\keyword{internal}
