% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_describe_db_subnet_groups}
\alias{docdb_describe_db_subnet_groups}
\title{Returns a list of DBSubnetGroup descriptions}
\usage{
docdb_describe_db_subnet_groups(
  DBSubnetGroupName = NULL,
  Filters = NULL,
  MaxRecords = NULL,
  Marker = NULL
)
}
\arguments{
\item{DBSubnetGroupName}{The name of the subnet group to return details for.}

\item{Filters}{This parameter is not currently supported.}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
(marker) is included in the response so that the remaining results can
be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}

\item{Marker}{An optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}
}
\description{
Returns a list of \code{DBSubnetGroup} descriptions. If a \code{DBSubnetGroupName} is specified, the list will contain only the descriptions of the specified \code{DBSubnetGroup}.

See \url{https://paws-r.github.io/docs/docdb/describe_db_subnet_groups.html} for full documentation.
}
\keyword{internal}
