% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_describe_pending_maintenance_actions}
\alias{docdb_describe_pending_maintenance_actions}
\title{Returns a list of resources (for example, instances) that have at least
one pending maintenance action}
\usage{
docdb_describe_pending_maintenance_actions(ResourceIdentifier, Filters,
  Marker, MaxRecords)
}
\arguments{
\item{ResourceIdentifier}{The ARN of a resource to return pending maintenance actions for.}

\item{Filters}{A filter that specifies one or more resources to return pending
maintenance actions for.

Supported filters:
\itemize{
\item \code{db-cluster-id} - Accepts cluster identifiers and cluster Amazon
Resource Names (ARNs). The results list includes only pending
maintenance actions for the clusters identified by these ARNs.
\item \code{db-instance-id} - Accepts instance identifiers and instance ARNs.
The results list includes only pending maintenance actions for the
DB instances identified by these ARNs.
}}

\item{Marker}{An optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
(marker) is included in the response so that the remaining results can
be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  PendingMaintenanceActions = list(
    list(
      ResourceIdentifier = "string",
      PendingMaintenanceActionDetails = list(
        list(
          Action = "string",
          AutoAppliedAfterDate = as.POSIXct(
            "2015-01-01"
          ),
          ForcedApplyDate = as.POSIXct(
            "2015-01-01"
          ),
          OptInStatus = "string",
          CurrentApplyDate = as.POSIXct(
            "2015-01-01"
          ),
          Description = "string"
        )
      )
    )
  ),
  Marker = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns a list of resources (for example, instances) that have at least
one pending maintenance action.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_pending_maintenance_actions(
  ResourceIdentifier = "string",
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  Marker = "string",
  MaxRecords = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
