% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_failover_db_cluster}
\alias{docdb_failover_db_cluster}
\title{Forces a failover for a cluster}
\usage{
docdb_failover_db_cluster(DBClusterIdentifier,
  TargetDBInstanceIdentifier)
}
\arguments{
\item{DBClusterIdentifier}{A cluster identifier to force a failover for. This parameter is not case
sensitive.

Constraints:
\itemize{
\item Must match the identifier of an existing \code{DBCluster}.
}}

\item{TargetDBInstanceIdentifier}{The name of the instance to promote to the primary instance.

You must specify the instance identifier for an Amazon DocumentDB
replica in the cluster. For example, \code{mydbcluster-replica1}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DBCluster = list(
    AvailabilityZones = list(
      "string"
    ),
    BackupRetentionPeriod = 123,
    DBClusterIdentifier = "string",
    DBClusterParameterGroup = "string",
    DBSubnetGroup = "string",
    Status = "string",
    PercentProgress = "string",
    EarliestRestorableTime = as.POSIXct(
      "2015-01-01"
    ),
    Endpoint = "string",
    ReaderEndpoint = "string",
    MultiAZ = TRUE|FALSE,
    Engine = "string",
    EngineVersion = "string",
    LatestRestorableTime = as.POSIXct(
      "2015-01-01"
    ),
    Port = 123,
    MasterUsername = "string",
    PreferredBackupWindow = "string",
    PreferredMaintenanceWindow = "string",
    DBClusterMembers = list(
      list(
        DBInstanceIdentifier = "string",
        IsClusterWriter = TRUE|FALSE,
        DBClusterParameterGroupStatus = "string",
        PromotionTier = 123
      )
    ),
    VpcSecurityGroups = list(
      list(
        VpcSecurityGroupId = "string",
        Status = "string"
      )
    ),
    HostedZoneId = "string",
    StorageEncrypted = TRUE|FALSE,
    KmsKeyId = "string",
    DbClusterResourceId = "string",
    DBClusterArn = "string",
    AssociatedRoles = list(
      list(
        RoleArn = "string",
        Status = "string"
      )
    ),
    ClusterCreateTime = as.POSIXct(
      "2015-01-01"
    ),
    EnabledCloudwatchLogsExports = list(
      "string"
    ),
    DeletionProtection = TRUE|FALSE
  )
)
}\if{html}{\out{</div>}}
}
\description{
Forces a failover for a cluster.

A failover for a cluster promotes one of the Amazon DocumentDB replicas
(read-only instances) in the cluster to be the primary instance (the
cluster writer).

If the primary instance fails, Amazon DocumentDB automatically fails
over to an Amazon DocumentDB replica, if one exists. You can force a
failover when you want to simulate a failure of a primary instance for
testing.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$failover_db_cluster(
  DBClusterIdentifier = "string",
  TargetDBInstanceIdentifier = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
