% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_list_tags_for_resource}
\alias{docdb_list_tags_for_resource}
\title{Lists all tags on an Amazon DocumentDB resource}
\usage{
docdb_list_tags_for_resource(ResourceName, Filters = NULL)
}
\arguments{
\item{ResourceName}{[required] The Amazon DocumentDB resource with tags to be listed. This value is an
Amazon Resource Name (ARN).}

\item{Filters}{This parameter is not currently supported.}
}
\description{
Lists all tags on an Amazon DocumentDB resource.

See \url{https://paws-r.github.io/docs/docdb/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
