% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_reboot_db_instance}
\alias{docdb_reboot_db_instance}
\title{You might need to reboot your instance, usually for maintenance reasons}
\usage{
docdb_reboot_db_instance(DBInstanceIdentifier, ForceFailover = NULL)
}
\arguments{
\item{DBInstanceIdentifier}{[required] The instance identifier. This parameter is stored as a lowercase string.

Constraints:
\itemize{
\item Must match the identifier of an existing \code{DBInstance}.
}}

\item{ForceFailover}{When \code{true}, the reboot is conducted through a Multi-AZ failover.

Constraint: You can't specify \code{true} if the instance is not configured
for Multi-AZ.}
}
\description{
You might need to reboot your instance, usually for maintenance reasons. For example, if you make certain changes, or if you change the cluster parameter group that is associated with the instance, you must reboot the instance for the changes to take effect.

See \url{https://paws-r.github.io/docs/docdb/reboot_db_instance.html} for full documentation.
}
\keyword{internal}
