% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_remove_source_identifier_from_subscription}
\alias{docdb_remove_source_identifier_from_subscription}
\title{Removes a source identifier from an existing Amazon DocumentDB event
notification subscription}
\usage{
docdb_remove_source_identifier_from_subscription(
  SubscriptionName,
  SourceIdentifier
)
}
\arguments{
\item{SubscriptionName}{[required] The name of the Amazon DocumentDB event notification subscription that
you want to remove a source identifier from.}

\item{SourceIdentifier}{[required] The source identifier to be removed from the subscription, such as the
instance identifier for an instance, or the name of a security group.}
}
\description{
Removes a source identifier from an existing Amazon DocumentDB event notification subscription.

See \url{https://www.paws-r-sdk.com/docs/docdb_remove_source_identifier_from_subscription/} for full documentation.
}
\keyword{internal}
