% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_stop_db_cluster}
\alias{docdb_stop_db_cluster}
\title{Stops the running cluster that is specified by DBClusterIdentifier}
\usage{
docdb_stop_db_cluster(DBClusterIdentifier)
}
\arguments{
\item{DBClusterIdentifier}{[required] The identifier of the cluster to stop. Example:
\code{docdb-2019-05-28-15-24-52}}
}
\description{
Stops the running cluster that is specified by \code{DBClusterIdentifier}.
The cluster must be in the \emph{available} state. For more information, see
\href{https://docs.aws.amazon.com/documentdb/latest/developerguide/db-cluster-stop-start.html}{Stopping and Starting an Amazon DocumentDB Cluster}.
}
\section{Request syntax}{
\preformatted{svc$stop_db_cluster(
  DBClusterIdentifier = "string"
)
}
}

\keyword{internal}
