% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdbelastic_operations.R
\name{docdbelastic_get_cluster}
\alias{docdbelastic_get_cluster}
\title{Returns information about a specific Elastic DocumentDB cluster}
\usage{
docdbelastic_get_cluster(clusterArn)
}
\arguments{
\item{clusterArn}{[required] The arn of the Elastic DocumentDB cluster.}
}
\description{
Returns information about a specific Elastic DocumentDB cluster.

See \url{https://www.paws-r-sdk.com/docs/docdbelastic_get_cluster/} for full documentation.
}
\keyword{internal}
