% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdbelastic_operations.R
\name{docdbelastic_get_cluster_snapshot}
\alias{docdbelastic_get_cluster_snapshot}
\title{Returns information about a specific elastic cluster snapshot}
\usage{
docdbelastic_get_cluster_snapshot(snapshotArn)
}
\arguments{
\item{snapshotArn}{[required] The ARN identifier of the elastic cluster snapshot.}
}
\description{
Returns information about a specific elastic cluster snapshot

See \url{https://www.paws-r-sdk.com/docs/docdbelastic_get_cluster_snapshot/} for full documentation.
}
\keyword{internal}
