% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdbelastic_operations.R
\name{docdbelastic_get_pending_maintenance_action}
\alias{docdbelastic_get_pending_maintenance_action}
\title{Retrieves all maintenance actions that are pending}
\usage{
docdbelastic_get_pending_maintenance_action(resourceArn)
}
\arguments{
\item{resourceArn}{[required] Retrieves pending maintenance actions for a specific Amazon Resource
Name (ARN).}
}
\description{
Retrieves all maintenance actions that are pending.

See \url{https://www.paws-r-sdk.com/docs/docdbelastic_get_pending_maintenance_action/} for full documentation.
}
\keyword{internal}
