% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdbelastic_operations.R
\name{docdbelastic_list_cluster_snapshots}
\alias{docdbelastic_list_cluster_snapshots}
\title{Returns information about Elastic DocumentDB snapshots for a specified
cluster}
\usage{
docdbelastic_list_cluster_snapshots(
  clusterArn = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{clusterArn}{The arn of the Elastic DocumentDB cluster.}

\item{maxResults}{The maximum number of entries to recieve in the response.}

\item{nextToken}{The nextToken which is used the get the next page of data.}
}
\description{
Returns information about Elastic DocumentDB snapshots for a specified cluster.

See \url{https://www.paws-r-sdk.com/docs/docdbelastic_list_cluster_snapshots/} for full documentation.
}
\keyword{internal}
