% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdbelastic_operations.R
\name{docdbelastic_list_clusters}
\alias{docdbelastic_list_clusters}
\title{Returns information about provisioned Elastic DocumentDB clusters}
\usage{
docdbelastic_list_clusters(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of entries to recieve in the response.}

\item{nextToken}{The nextToken which is used the get the next page of data.}
}
\description{
Returns information about provisioned Elastic DocumentDB clusters.

See \url{https://www.paws-r-sdk.com/docs/docdbelastic_list_clusters/} for full documentation.
}
\keyword{internal}
