% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdbelastic_operations.R
\name{docdbelastic_list_pending_maintenance_actions}
\alias{docdbelastic_list_pending_maintenance_actions}
\title{Retrieves a list of all maintenance actions that are pending}
\usage{
docdbelastic_list_pending_maintenance_actions(
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of results to include in the response. If more
records exist than the specified \code{maxResults} value, a pagination token
(marker) is included in the response so that the remaining results can
be retrieved.}

\item{nextToken}{An optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{maxResults}.}
}
\description{
Retrieves a list of all maintenance actions that are pending.

See \url{https://www.paws-r-sdk.com/docs/docdbelastic_list_pending_maintenance_actions/} for full documentation.
}
\keyword{internal}
