% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdbelastic_operations.R
\name{docdbelastic_list_tags_for_resource}
\alias{docdbelastic_list_tags_for_resource}
\title{Lists all tags on a Elastic DocumentDB resource}
\usage{
docdbelastic_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The arn of the Elastic DocumentDB resource.}
}
\description{
Lists all tags on a Elastic DocumentDB resource

See \url{https://www.paws-r-sdk.com/docs/docdbelastic_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
