% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdbelastic_operations.R
\name{docdbelastic_start_cluster}
\alias{docdbelastic_start_cluster}
\title{Restarts the stopped elastic cluster that is specified by clusterARN}
\usage{
docdbelastic_start_cluster(clusterArn)
}
\arguments{
\item{clusterArn}{[required] The ARN identifier of the elastic cluster.}
}
\description{
Restarts the stopped elastic cluster that is specified by \code{clusterARN}.

See \url{https://www.paws-r-sdk.com/docs/docdbelastic_start_cluster/} for full documentation.
}
\keyword{internal}
