% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdbelastic_operations.R
\name{docdbelastic_stop_cluster}
\alias{docdbelastic_stop_cluster}
\title{Stops the running elastic cluster that is specified by clusterArn}
\usage{
docdbelastic_stop_cluster(clusterArn)
}
\arguments{
\item{clusterArn}{[required] The ARN identifier of the elastic cluster.}
}
\description{
Stops the running elastic cluster that is specified by \code{clusterArn}. The elastic cluster must be in the \emph{available} state.

See \url{https://www.paws-r-sdk.com/docs/docdbelastic_stop_cluster/} for full documentation.
}
\keyword{internal}
