% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdbelastic_operations.R
\name{docdbelastic_tag_resource}
\alias{docdbelastic_tag_resource}
\title{Adds metadata tags to a Elastic DocumentDB resource}
\usage{
docdbelastic_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The arn of the Elastic DocumentDB resource.}

\item{tags}{[required] The tags to be assigned to the Elastic DocumentDB resource.}
}
\description{
Adds metadata tags to a Elastic DocumentDB resource

See \url{https://www.paws-r-sdk.com/docs/docdbelastic_tag_resource/} for full documentation.
}
\keyword{internal}
