% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_delete_backup}
\alias{dynamodb_delete_backup}
\title{Deletes an existing backup of a table}
\usage{
dynamodb_delete_backup(BackupArn)
}
\arguments{
\item{BackupArn}{[required] The ARN associated with the backup.}
}
\description{
Deletes an existing backup of a table.

See \url{https://www.paws-r-sdk.com/docs/dynamodb_delete_backup/} for full documentation.
}
\keyword{internal}
