% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_describe_backup}
\alias{dynamodb_describe_backup}
\title{Describes an existing backup of a table}
\usage{
dynamodb_describe_backup(BackupArn)
}
\arguments{
\item{BackupArn}{[required] The Amazon Resource Name (ARN) associated with the backup.}
}
\description{
Describes an existing backup of a table.

You can call \code{\link[=dynamodb_describe_backup]{describe_backup}} at a maximum
rate of 10 times per second.
}
\section{Request syntax}{
\preformatted{svc$describe_backup(
  BackupArn = "string"
)
}
}

\keyword{internal}
