% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_describe_endpoints}
\alias{dynamodb_describe_endpoints}
\title{Returns the regional endpoint information}
\usage{
dynamodb_describe_endpoints()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Endpoints = list(
    list(
      Address = "string",
      CachePeriodInMinutes = 123
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns the regional endpoint information.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_endpoints()
}\if{html}{\out{</div>}}
}

\keyword{internal}
