% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_describe_global_table}
\alias{dynamodb_describe_global_table}
\title{Returns information about the specified global table}
\usage{
dynamodb_describe_global_table(GlobalTableName)
}
\arguments{
\item{GlobalTableName}{[required] The name of the global table.}
}
\description{
Returns information about the specified global table.

See \url{https://paws-r.github.io/docs/dynamodb/describe_global_table.html} for full documentation.
}
\keyword{internal}
