% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_describe_import}
\alias{dynamodb_describe_import}
\title{Represents the properties of the import}
\usage{
dynamodb_describe_import(ImportArn)
}
\arguments{
\item{ImportArn}{[required] The Amazon Resource Name (ARN) associated with the table you're
importing to.}
}
\description{
Represents the properties of the import.

See \url{https://www.paws-r-sdk.com/docs/dynamodb_describe_import/} for full documentation.
}
\keyword{internal}
