% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_describe_kinesis_streaming_destination}
\alias{dynamodb_describe_kinesis_streaming_destination}
\title{Returns information about the status of Kinesis streaming}
\usage{
dynamodb_describe_kinesis_streaming_destination(TableName)
}
\arguments{
\item{TableName}{[required] The name of the table being described.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  TableName = "string",
  KinesisDataStreamDestinations = list(
    list(
      StreamArn = "string",
      DestinationStatus = "ENABLING"|"ACTIVE"|"DISABLING"|"DISABLED"|"ENABLE_FAILED",
      DestinationStatusDescription = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns information about the status of Kinesis streaming.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_kinesis_streaming_destination(
  TableName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
