% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_describe_table}
\alias{dynamodb_describe_table}
\title{Returns information about the table, including the current status of the
table, when it was created, the primary key schema, and any indexes on
the table}
\usage{
dynamodb_describe_table(TableName)
}
\arguments{
\item{TableName}{[required] The name of the table to describe.}
}
\description{
Returns information about the table, including the current status of the table, when it was created, the primary key schema, and any indexes on the table.

See \url{https://www.paws-r-sdk.com/docs/dynamodb_describe_table/} for full documentation.
}
\keyword{internal}
