% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_describe_time_to_live}
\alias{dynamodb_describe_time_to_live}
\title{Gives a description of the Time to Live (TTL) status on the specified
table}
\usage{
dynamodb_describe_time_to_live(TableName)
}
\arguments{
\item{TableName}{[required] The name of the table to be described.}
}
\value{
A list with the following syntax:\preformatted{list(
  TimeToLiveDescription = list(
    TimeToLiveStatus = "ENABLING"|"DISABLING"|"ENABLED"|"DISABLED",
    AttributeName = "string"
  )
)
}
}
\description{
Gives a description of the Time to Live (TTL) status on the specified
table.
}
\section{Request syntax}{
\preformatted{svc$describe_time_to_live(
  TableName = "string"
)
}
}

\keyword{internal}
