% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_get_resource_policy}
\alias{dynamodb_get_resource_policy}
\title{Returns the resource-based policy document attached to the resource,
which can be a table or stream, in JSON format}
\usage{
dynamodb_get_resource_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the DynamoDB resource to which the
policy is attached. The resources you can specify include tables and
streams.}
}
\description{
Returns the resource-based policy document attached to the resource, which can be a table or stream, in JSON format.

See \url{https://www.paws-r-sdk.com/docs/dynamodb_get_resource_policy/} for full documentation.
}
\keyword{internal}
