% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_put_item}
\alias{dynamodb_put_item}
\title{Creates a new item, or replaces an old item with a new item}
\usage{
dynamodb_put_item(
  TableName,
  Item,
  Expected = NULL,
  ReturnValues = NULL,
  ReturnConsumedCapacity = NULL,
  ReturnItemCollectionMetrics = NULL,
  ConditionalOperator = NULL,
  ConditionExpression = NULL,
  ExpressionAttributeNames = NULL,
  ExpressionAttributeValues = NULL,
  ReturnValuesOnConditionCheckFailure = NULL
)
}
\arguments{
\item{TableName}{[required] The name of the table to contain the item. You can also provide the Amazon Resource Name (ARN) of the table in this parameter.}

\item{Item}{[required] A map of attribute name/value pairs, one for each attribute. Only the primary key attributes are required; you can optionally provide other attribute name-value pairs for the item.

You must provide all of the attributes for the primary key. For example, with a simple primary key, you only need to provide a value for the partition key. For a composite primary key, you must provide both values for both the partition key and the sort key.

If you specify any attributes that are part of an index key, then the data types for those attributes must match those of the schema in the table's attribute definition.

If the table has vector indexes, the following validations apply to write operations. A violation of any of these constraints results in a \code{ValidationException}:
\itemize{
\item The vector attribute must be a list of numbers with dimensions matching the index configuration.
\item Vector values must fit in 32-bit IEEE-754 floating point format (f32).
\item Partition key and inline filter attributes defined in the search schema must have data types matching the index schema definition.
}

Empty String and Binary attribute values are allowed. Attribute values of type String and Binary must have a length greater than zero if the attribute is used as a key attribute for a table or index.

For more information about primary keys, see \href{https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/HowItWorks.CoreComponents.html#HowItWorks.CoreComponents.PrimaryKey}{Primary Key} in the \emph{Amazon DynamoDB Developer Guide}.

Each element in the \code{Item} map is an \code{AttributeValue} object.}

\item{Expected}{This is a legacy parameter. Use \code{ConditionExpression} instead. For more information, see \href{https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/LegacyConditionalParameters.Expected.html}{Expected} in the \emph{Amazon DynamoDB Developer Guide}.}

\item{ReturnValues}{Use \code{ReturnValues} if you want to get the item attributes as they appeared before they were updated with the \code{\link[=dynamodb_put_item]{put_item}} request. For \code{\link[=dynamodb_put_item]{put_item}}, the valid values are:
\itemize{
\item \code{NONE} - If \code{ReturnValues} is not specified, or if its value is \code{NONE}, then nothing is returned. (This setting is the default for \code{ReturnValues}.)
\item \code{ALL_OLD} - If \code{\link[=dynamodb_put_item]{put_item}} overwrote an attribute name-value pair, then the content of the old item is returned.
}

The values returned are strongly consistent.

There is no additional cost associated with requesting a return value aside from the small network and processing overhead of receiving a larger response. No read capacity units are consumed.

The \code{ReturnValues} parameter is used by several DynamoDB operations; however, \code{\link[=dynamodb_put_item]{put_item}} does not recognize any values other than \code{NONE} or \code{ALL_OLD}.}

\item{ReturnConsumedCapacity}{Determines the level of detail about either provisioned or on-demand throughput consumption that is returned in the response:
\itemize{
\item \code{INDEXES} - The response includes the aggregate \code{ConsumedCapacity} for the operation, together with \code{ConsumedCapacity} for each table and secondary index that was accessed.

Note that some operations, such as \code{\link[=dynamodb_get_item]{get_item}} and \code{\link[=dynamodb_batch_get_item]{batch_get_item}}, do not access any indexes at all. In these cases, specifying \code{INDEXES} will only return \code{ConsumedCapacity} information for table(s).
\item \code{TOTAL} - The response includes only the aggregate \code{ConsumedCapacity} for the operation.
\item \code{NONE} - No \code{ConsumedCapacity} details are included in the response.
}}

\item{ReturnItemCollectionMetrics}{Determines whether item collection metrics are returned. If set to \code{SIZE}, the response includes statistics about item collections, if any, that were modified during the operation are returned in the response. If set to \code{NONE} (the default), no statistics are returned.}

\item{ConditionalOperator}{This is a legacy parameter. Use \code{ConditionExpression} instead. For more information, see \href{https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/LegacyConditionalParameters.ConditionalOperator.html}{ConditionalOperator} in the \emph{Amazon DynamoDB Developer Guide}.}

\item{ConditionExpression}{A condition that must be satisfied in order for a conditional \code{\link[=dynamodb_put_item]{put_item}} operation to succeed.

An expression can contain any of the following:
\itemize{
\item Functions: \code{attribute_exists | attribute_not_exists | attribute_type | contains | begins_with | size}

These function names are case-sensitive.
\item Comparison operators: \verb{= | <> | < | > | <= | >= | BETWEEN | IN }
\item Logical operators: \code{AND | OR | NOT}
}

For more information on condition expressions, see \href{https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/Expressions.ConditionExpressions.html}{Condition Expressions} in the \emph{Amazon DynamoDB Developer Guide}.}

\item{ExpressionAttributeNames}{One or more substitution tokens for attribute names in an expression. The following are some use cases for using \code{ExpressionAttributeNames}:
\itemize{
\item To access an attribute whose name conflicts with a DynamoDB reserved word.
\item To create a placeholder for repeating occurrences of an attribute name in an expression.
\item To prevent special characters in an attribute name from being misinterpreted in an expression.
}

Use the \strong{#} character in an expression to dereference an attribute name. For example, consider the following attribute name:
\itemize{
\item \code{Percentile}
}

The name of this attribute conflicts with a reserved word, so it cannot be used directly in an expression. (For the complete list of reserved words, see \href{https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/ReservedWords.html}{Reserved Words} in the \emph{Amazon DynamoDB Developer Guide}). To work around this, you could specify the following for \code{ExpressionAttributeNames}:
\itemize{
\item \code{{"#P":"Percentile"}}
}

You could then use this substitution in an expression, as in this example:
\itemize{
\item \verb{#P = :val}
}

Tokens that begin with the \strong{:} character are \emph{expression attribute values}, which are placeholders for the actual value at runtime.

For more information on expression attribute names, see \href{https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/Expressions.Attributes.html}{Specifying Item Attributes} in the \emph{Amazon DynamoDB Developer Guide}.}

\item{ExpressionAttributeValues}{One or more values that can be substituted in an expression.

Use the \strong{:} (colon) character in an expression to dereference an attribute value. For example, suppose that you wanted to check whether the value of the \emph{ProductStatus} attribute was one of the following:

\code{Available | Backordered | Discontinued}

You would first need to specify \code{ExpressionAttributeValues} as follows:

\verb{\{ ":avail":\{"S":"Available"\}, ":back":\{"S":"Backordered"\}, ":disc":\{"S":"Discontinued"\} \}}

You could then use these values in an expression, such as this:

\verb{ProductStatus IN (:avail, :back, :disc)}

For more information on expression attribute values, see \href{https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/Expressions.ConditionExpressions.html}{Condition Expressions} in the \emph{Amazon DynamoDB Developer Guide}.}

\item{ReturnValuesOnConditionCheckFailure}{An optional parameter that returns the item attributes for a \code{\link[=dynamodb_put_item]{put_item}} operation that failed a condition check.

There is no additional cost associated with requesting a return value aside from the small network and processing overhead of receiving a larger response. No read capacity units are consumed.}
}
\description{
Creates a new item, or replaces an old item with a new item. If an item that has the same primary key as the new item already exists in the specified table, the new item completely replaces the existing item. You can perform a conditional put operation (add a new item if one with the specified primary key doesn't exist), or replace an existing item if it has certain attribute values. You can return the item's attribute values in the same operation, using the \code{ReturnValues} parameter.

See \url{https://www.paws-r-sdk.com/docs/dynamodb_put_item/} for full documentation.
}
\keyword{internal}
