% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_update_continuous_backups}
\alias{dynamodb_update_continuous_backups}
\title{UpdateContinuousBackups enables or disables point in time recovery for
the specified table}
\usage{
dynamodb_update_continuous_backups(TableName, PointInTimeRecoverySpecification)
}
\arguments{
\item{TableName}{[required] The name of the table. You can also provide the Amazon Resource Name
(ARN) of the table in this parameter.}

\item{PointInTimeRecoverySpecification}{[required] Represents the settings used to enable point in time recovery.}
}
\description{
\code{\link[=dynamodb_update_continuous_backups]{update_continuous_backups}} enables or disables point in time recovery for the specified table. A successful \code{\link[=dynamodb_update_continuous_backups]{update_continuous_backups}} call returns the current \code{ContinuousBackupsDescription}. Continuous backups are \code{ENABLED} on all tables at table creation. If point in time recovery is enabled, \code{PointInTimeRecoveryStatus} will be set to ENABLED.

See \url{https://www.paws-r-sdk.com/docs/dynamodb_update_continuous_backups/} for full documentation.
}
\keyword{internal}
