% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_update_contributor_insights}
\alias{dynamodb_update_contributor_insights}
\title{Updates the status for contributor insights for a specific table or
index}
\usage{
dynamodb_update_contributor_insights(TableName, IndexName,
  ContributorInsightsAction)
}
\arguments{
\item{TableName}{[required] The name of the table.}

\item{IndexName}{The global secondary index name, if applicable.}

\item{ContributorInsightsAction}{[required] Represents the contributor insights action.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  TableName = "string",
  IndexName = "string",
  ContributorInsightsStatus = "ENABLING"|"ENABLED"|"DISABLING"|"DISABLED"|"FAILED"
)
}\if{html}{\out{</div>}}
}
\description{
Updates the status for contributor insights for a specific table or
index.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_contributor_insights(
  TableName = "string",
  IndexName = "string",
  ContributorInsightsAction = "ENABLE"|"DISABLE"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
