% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_update_time_to_live}
\alias{dynamodb_update_time_to_live}
\title{The UpdateTimeToLive method enables or disables Time to Live (TTL) for
the specified table}
\usage{
dynamodb_update_time_to_live(TableName, TimeToLiveSpecification)
}
\arguments{
\item{TableName}{[required] The name of the table to be configured.}

\item{TimeToLiveSpecification}{[required] Represents the settings used to enable or disable Time to Live for the
specified table.}
}
\description{
The \code{\link[=dynamodb_update_time_to_live]{update_time_to_live}} method enables or disables Time to Live (TTL) for the specified table. A successful \code{\link[=dynamodb_update_time_to_live]{update_time_to_live}} call returns the current \code{TimeToLiveSpecification}. It can take up to one hour for the change to fully process. Any additional \code{\link[=dynamodb_update_time_to_live]{update_time_to_live}} calls for the same table during this one hour duration result in a \code{ValidationException}.

See \url{https://www.paws-r-sdk.com/docs/dynamodb_update_time_to_live/} for full documentation.
}
\keyword{internal}
