% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodbstreams_operations.R
\name{dynamodbstreams_get_records}
\alias{dynamodbstreams_get_records}
\title{Retrieves the stream records from a given shard}
\usage{
dynamodbstreams_get_records(ShardIterator, Limit = NULL)
}
\arguments{
\item{ShardIterator}{[required] A shard iterator that was retrieved from a previous GetShardIterator
operation. This iterator can be used to access the stream records in
this shard.}

\item{Limit}{The maximum number of records to return from the shard. The upper limit
is 1000.}
}
\description{
Retrieves the stream records from a given shard.

See \url{https://paws-r.github.io/docs/dynamodbstreams/get_records.html} for full documentation.
}
\keyword{internal}
