% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_copy_serverless_cache_snapshot}
\alias{elasticache_copy_serverless_cache_snapshot}
\title{Creates a copy of an existing serverless cache’s snapshot}
\usage{
elasticache_copy_serverless_cache_snapshot(
  SourceServerlessCacheSnapshotName,
  TargetServerlessCacheSnapshotName,
  KmsKeyId = NULL,
  Tags = NULL
)
}
\arguments{
\item{SourceServerlessCacheSnapshotName}{[required] The identifier of the existing serverless cache’s snapshot to be copied.
Available for Valkey, Redis OSS and Serverless Memcached only.}

\item{TargetServerlessCacheSnapshotName}{[required] The identifier for the snapshot to be created. Available for Valkey,
Redis OSS and Serverless Memcached only.}

\item{KmsKeyId}{The identifier of the KMS key used to encrypt the target snapshot.
Available for Valkey, Redis OSS and Serverless Memcached only.}

\item{Tags}{A list of tags to be added to the target snapshot resource. A tag is a
key-value pair. Available for Valkey, Redis OSS and Serverless Memcached
only. Default: NULL}
}
\description{
Creates a copy of an existing serverless cache’s snapshot. Available for Valkey, Redis OSS and Serverless Memcached only.

See \url{https://www.paws-r-sdk.com/docs/elasticache_copy_serverless_cache_snapshot/} for full documentation.
}
\keyword{internal}
