% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_create_cache_subnet_group}
\alias{elasticache_create_cache_subnet_group}
\title{Creates a new cache subnet group}
\usage{
elasticache_create_cache_subnet_group(
  CacheSubnetGroupName,
  CacheSubnetGroupDescription,
  SubnetIds,
  Tags = NULL
)
}
\arguments{
\item{CacheSubnetGroupName}{[required] A name for the cache subnet group. This value is stored as a lowercase
string.

Constraints: Must contain no more than 255 alphanumeric characters or
hyphens.

Example: \code{mysubnetgroup}}

\item{CacheSubnetGroupDescription}{[required] A description for the cache subnet group.}

\item{SubnetIds}{[required] A list of VPC subnet IDs for the cache subnet group.}

\item{Tags}{A list of tags to be added to this resource. A tag is a key-value pair.
A tag key must be accompanied by a tag value, although null is accepted.}
}
\description{
Creates a new cache subnet group.

See \url{https://www.paws-r-sdk.com/docs/elasticache_create_cache_subnet_group/} for full documentation.
}
\keyword{internal}
