% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_decrease_replica_count}
\alias{elasticache_decrease_replica_count}
\title{Dynamically decreases the number of replicas in a Redis (cluster mode
disabled) replication group or the number of replica nodes in one or
more node groups (shards) of a Redis (cluster mode enabled) replication
group}
\usage{
elasticache_decrease_replica_count(ReplicationGroupId, NewReplicaCount,
  ReplicaConfiguration, ReplicasToRemove, ApplyImmediately)
}
\arguments{
\item{ReplicationGroupId}{[required] The id of the replication group from which you want to remove replica
nodes.}

\item{NewReplicaCount}{The number of read replica nodes you want at the completion of this
operation. For Redis (cluster mode disabled) replication groups, this is
the number of replica nodes in the replication group. For Redis (cluster
mode enabled) replication groups, this is the number of replica nodes in
each of the replication group's node groups.

The minimum number of replicas in a shard or replication group is:
\itemize{
\item Redis (cluster mode disabled)
\itemize{
\item If Multi-AZ is enabled: 1
\item If Multi-AZ is not enabled: 0
}
\item Redis (cluster mode enabled): 0 (though you will not be able to
failover to a replica if your primary node fails)
}}

\item{ReplicaConfiguration}{A list of \code{ConfigureShard} objects that can be used to configure each
shard in a Redis (cluster mode enabled) replication group. The
\code{ConfigureShard} has three members: \code{NewReplicaCount}, \code{NodeGroupId},
and \code{PreferredAvailabilityZones}.}

\item{ReplicasToRemove}{A list of the node ids to remove from the replication group or node
group (shard).}

\item{ApplyImmediately}{[required] If \code{True}, the number of replica nodes is decreased immediately.
\code{ApplyImmediately=False} is not currently supported.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ReplicationGroup = list(
    ReplicationGroupId = "string",
    Description = "string",
    GlobalReplicationGroupInfo = list(
      GlobalReplicationGroupId = "string",
      GlobalReplicationGroupMemberRole = "string"
    ),
    Status = "string",
    PendingModifiedValues = list(
      PrimaryClusterId = "string",
      AutomaticFailoverStatus = "enabled"|"disabled",
      Resharding = list(
        SlotMigration = list(
          ProgressPercentage = 123.0
        )
      ),
      AuthTokenStatus = "SETTING"|"ROTATING",
      UserGroups = list(
        UserGroupIdsToAdd = list(
          "string"
        ),
        UserGroupIdsToRemove = list(
          "string"
        )
      )
    ),
    MemberClusters = list(
      "string"
    ),
    NodeGroups = list(
      list(
        NodeGroupId = "string",
        Status = "string",
        PrimaryEndpoint = list(
          Address = "string",
          Port = 123
        ),
        ReaderEndpoint = list(
          Address = "string",
          Port = 123
        ),
        Slots = "string",
        NodeGroupMembers = list(
          list(
            CacheClusterId = "string",
            CacheNodeId = "string",
            ReadEndpoint = list(
              Address = "string",
              Port = 123
            ),
            PreferredAvailabilityZone = "string",
            PreferredOutpostArn = "string",
            CurrentRole = "string"
          )
        )
      )
    ),
    SnapshottingClusterId = "string",
    AutomaticFailover = "enabled"|"disabled"|"enabling"|"disabling",
    MultiAZ = "enabled"|"disabled",
    ConfigurationEndpoint = list(
      Address = "string",
      Port = 123
    ),
    SnapshotRetentionLimit = 123,
    SnapshotWindow = "string",
    ClusterEnabled = TRUE|FALSE,
    CacheNodeType = "string",
    AuthTokenEnabled = TRUE|FALSE,
    AuthTokenLastModifiedDate = as.POSIXct(
      "2015-01-01"
    ),
    TransitEncryptionEnabled = TRUE|FALSE,
    AtRestEncryptionEnabled = TRUE|FALSE,
    MemberClustersOutpostArns = list(
      "string"
    ),
    KmsKeyId = "string",
    ARN = "string",
    UserGroupIds = list(
      "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Dynamically decreases the number of replicas in a Redis (cluster mode
disabled) replication group or the number of replica nodes in one or
more node groups (shards) of a Redis (cluster mode enabled) replication
group. This operation is performed with no cluster down time.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$decrease_replica_count(
  ReplicationGroupId = "string",
  NewReplicaCount = 123,
  ReplicaConfiguration = list(
    list(
      NodeGroupId = "string",
      NewReplicaCount = 123,
      PreferredAvailabilityZones = list(
        "string"
      ),
      PreferredOutpostArns = list(
        "string"
      )
    )
  ),
  ReplicasToRemove = list(
    "string"
  ),
  ApplyImmediately = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
