% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_delete_cache_security_group}
\alias{elasticache_delete_cache_security_group}
\title{Deletes a cache security group}
\usage{
elasticache_delete_cache_security_group(CacheSecurityGroupName)
}
\arguments{
\item{CacheSecurityGroupName}{[required] The name of the cache security group to delete.

You cannot delete the default security group.}
}
\description{
Deletes a cache security group.

See \url{https://www.paws-r-sdk.com/docs/elasticache_delete_cache_security_group/} for full documentation.
}
\keyword{internal}
