% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_delete_cache_subnet_group}
\alias{elasticache_delete_cache_subnet_group}
\title{Deletes a cache subnet group}
\usage{
elasticache_delete_cache_subnet_group(CacheSubnetGroupName)
}
\arguments{
\item{CacheSubnetGroupName}{[required] The name of the cache subnet group to delete.

Constraints: Must contain no more than 255 alphanumeric characters or
hyphens.}
}
\description{
Deletes a cache subnet group.
}
\details{
You cannot delete a cache subnet group if it is associated with any
clusters.
}
\section{Request syntax}{
\preformatted{svc$delete_cache_subnet_group(
  CacheSubnetGroupName = "string"
)
}
}

\keyword{internal}
