% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_delete_global_replication_group}
\alias{elasticache_delete_global_replication_group}
\title{Deleting a Global datastore is a two-step process:}
\usage{
elasticache_delete_global_replication_group(
  GlobalReplicationGroupId,
  RetainPrimaryReplicationGroup
)
}
\arguments{
\item{GlobalReplicationGroupId}{[required] The name of the Global datastore}

\item{RetainPrimaryReplicationGroup}{[required] The primary replication group is retained as a standalone replication
group.}
}
\description{
Deleting a Global datastore is a two-step process:

See \url{https://paws-r.github.io/docs/elasticache/delete_global_replication_group.html} for full documentation.
}
\keyword{internal}
