% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_delete_serverless_cache}
\alias{elasticache_delete_serverless_cache}
\title{Deletes a specified existing serverless cache}
\usage{
elasticache_delete_serverless_cache(
  ServerlessCacheName,
  FinalSnapshotName = NULL
)
}
\arguments{
\item{ServerlessCacheName}{[required] The identifier of the serverless cache to be deleted.}

\item{FinalSnapshotName}{Name of the final snapshot to be taken before the serverless cache is
deleted. Available for Valkey, Redis OSS and Serverless Memcached only.
Default: NULL, i.e. a final snapshot is not taken.}
}
\description{
Deletes a specified existing serverless cache.

See \url{https://www.paws-r-sdk.com/docs/elasticache_delete_serverless_cache/} for full documentation.
}
\keyword{internal}
