% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_delete_serverless_cache_snapshot}
\alias{elasticache_delete_serverless_cache_snapshot}
\title{Deletes an existing serverless cache snapshot}
\usage{
elasticache_delete_serverless_cache_snapshot(ServerlessCacheSnapshotName)
}
\arguments{
\item{ServerlessCacheSnapshotName}{[required] Idenfitier of the snapshot to be deleted. Available for Redis OSS and
Serverless Memcached only.}
}
\description{
Deletes an existing serverless cache snapshot. Available for Redis OSS and Serverless Memcached only.

See \url{https://www.paws-r-sdk.com/docs/elasticache_delete_serverless_cache_snapshot/} for full documentation.
}
\keyword{internal}
