% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_describe_cache_security_groups}
\alias{elasticache_describe_cache_security_groups}
\title{Returns a list of cache security group descriptions}
\usage{
elasticache_describe_cache_security_groups(CacheSecurityGroupName,
  MaxRecords, Marker)
}
\arguments{
\item{CacheSecurityGroupName}{The name of the cache security group to return details for.}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a marker is
included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.}

\item{Marker}{An optional marker returned from a prior request. Use this marker for
pagination of results from this operation. If this parameter is
specified, the response includes only records beyond the marker, up to
the value specified by \code{MaxRecords}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Marker = "string",
  CacheSecurityGroups = list(
    list(
      OwnerId = "string",
      CacheSecurityGroupName = "string",
      Description = "string",
      EC2SecurityGroups = list(
        list(
          Status = "string",
          EC2SecurityGroupName = "string",
          EC2SecurityGroupOwnerId = "string"
        )
      ),
      ARN = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns a list of cache security group descriptions. If a cache security
group name is specified, the list contains only the description of that
group. This applicable only when you have ElastiCache in Classic setup
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_cache_security_groups(
  CacheSecurityGroupName = "string",
  MaxRecords = 123,
  Marker = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
