% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_describe_serverless_cache_snapshots}
\alias{elasticache_describe_serverless_cache_snapshots}
\title{Returns information about serverless cache snapshots}
\usage{
elasticache_describe_serverless_cache_snapshots(
  ServerlessCacheName = NULL,
  ServerlessCacheSnapshotName = NULL,
  SnapshotType = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ServerlessCacheName}{The identifier of serverless cache. If this parameter is specified, only
snapshots associated with that specific serverless cache are described.
Available for Redis only.}

\item{ServerlessCacheSnapshotName}{The identifier of the serverless cache’s snapshot. If this parameter is
specified, only this snapshot is described. Available for Redis only.}

\item{SnapshotType}{The type of snapshot that is being described. Available for Redis only.}

\item{NextToken}{An optional marker returned from a prior request to support pagination
of results from this operation. If this parameter is specified, the
response includes only records beyond the marker, up to the value
specified by max-results. Available for Redis only.}

\item{MaxResults}{The maximum number of records to include in the response. If more
records exist than the specified max-results value, a market is included
in the response so that remaining results can be retrieved. Available
for Redis only.The default is 50. The Validation Constraints are a
maximum of 50.}
}
\description{
Returns information about serverless cache snapshots. By default, this API lists all of the customer’s serverless cache snapshots. It can also describe a single serverless cache snapshot, or the snapshots associated with a particular serverless cache. Available for Redis only.

See \url{https://www.paws-r-sdk.com/docs/elasticache_describe_serverless_cache_snapshots/} for full documentation.
}
\keyword{internal}
