% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_describe_service_updates}
\alias{elasticache_describe_service_updates}
\title{Returns details of the service updates}
\usage{
elasticache_describe_service_updates(ServiceUpdateName,
  ServiceUpdateStatus, MaxRecords, Marker)
}
\arguments{
\item{ServiceUpdateName}{The unique ID of the service update}

\item{ServiceUpdateStatus}{The status of the service update}

\item{MaxRecords}{The maximum number of records to include in the response}

\item{Marker}{An optional marker returned from a prior request. Use this marker for
pagination of results from this operation. If this parameter is
specified, the response includes only records beyond the marker, up to
the value specified by \code{MaxRecords}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Marker = "string",
  ServiceUpdates = list(
    list(
      ServiceUpdateName = "string",
      ServiceUpdateReleaseDate = as.POSIXct(
        "2015-01-01"
      ),
      ServiceUpdateEndDate = as.POSIXct(
        "2015-01-01"
      ),
      ServiceUpdateSeverity = "critical"|"important"|"medium"|"low",
      ServiceUpdateRecommendedApplyByDate = as.POSIXct(
        "2015-01-01"
      ),
      ServiceUpdateStatus = "available"|"cancelled"|"expired",
      ServiceUpdateDescription = "string",
      ServiceUpdateType = "security-update",
      Engine = "string",
      EngineVersion = "string",
      AutoUpdateAfterRecommendedApplyByDate = TRUE|FALSE,
      EstimatedUpdateTime = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns details of the service updates
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_service_updates(
  ServiceUpdateName = "string",
  ServiceUpdateStatus = list(
    "available"|"cancelled"|"expired"
  ),
  MaxRecords = 123,
  Marker = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
