% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_disassociate_global_replication_group}
\alias{elasticache_disassociate_global_replication_group}
\title{Remove a secondary cluster from the Global datastore using the Global
datastore name}
\usage{
elasticache_disassociate_global_replication_group(
  GlobalReplicationGroupId,
  ReplicationGroupId,
  ReplicationGroupRegion
)
}
\arguments{
\item{GlobalReplicationGroupId}{[required] The name of the Global datastore}

\item{ReplicationGroupId}{[required] The name of the secondary cluster you wish to remove from the Global
datastore}

\item{ReplicationGroupRegion}{[required] The Amazon region of secondary cluster you wish to remove from the
Global datastore}
}
\description{
Remove a secondary cluster from the Global datastore using the Global datastore name. The secondary cluster will no longer receive updates from the primary cluster, but will remain as a standalone cluster in that Amazon region.

See \url{https://www.paws-r-sdk.com/docs/elasticache_disassociate_global_replication_group/} for full documentation.
}
\keyword{internal}
