% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_export_serverless_cache_snapshot}
\alias{elasticache_export_serverless_cache_snapshot}
\title{Provides the functionality to export the serverless cache snapshot data
to Amazon S3}
\usage{
elasticache_export_serverless_cache_snapshot(
  ServerlessCacheSnapshotName,
  S3BucketName
)
}
\arguments{
\item{ServerlessCacheSnapshotName}{[required] The identifier of the serverless cache snapshot to be exported to S3.
Available for Valkey and Redis OSS only.}

\item{S3BucketName}{[required] Name of the Amazon S3 bucket to export the snapshot to. The Amazon S3
bucket must also be in same region as the snapshot. Available for Valkey
and Redis OSS only.}
}
\description{
Provides the functionality to export the serverless cache snapshot data to Amazon S3. Available for Valkey and Redis OSS only.

See \url{https://www.paws-r-sdk.com/docs/elasticache_export_serverless_cache_snapshot/} for full documentation.
}
\keyword{internal}
