% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_increase_node_groups_in_global_replication_group}
\alias{elasticache_increase_node_groups_in_global_replication_group}
\title{Increase the number of node groups in the Global datastore}
\usage{
elasticache_increase_node_groups_in_global_replication_group(
  GlobalReplicationGroupId,
  NodeGroupCount,
  RegionalConfigurations = NULL,
  ApplyImmediately
)
}
\arguments{
\item{GlobalReplicationGroupId}{[required] The name of the Global datastore}

\item{NodeGroupCount}{[required] Total number of node groups you want}

\item{RegionalConfigurations}{Describes the replication group IDs, the Amazon regions where they are
stored and the shard configuration for each that comprise the Global
datastore}

\item{ApplyImmediately}{[required] Indicates that the process begins immediately. At present, the only
permitted value for this parameter is true.}
}
\description{
Increase the number of node groups in the Global datastore

See \url{https://www.paws-r-sdk.com/docs/elasticache_increase_node_groups_in_global_replication_group/} for full documentation.
}
\keyword{internal}
