% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_modify_cache_subnet_group}
\alias{elasticache_modify_cache_subnet_group}
\title{Modifies an existing cache subnet group}
\usage{
elasticache_modify_cache_subnet_group(CacheSubnetGroupName,
  CacheSubnetGroupDescription, SubnetIds)
}
\arguments{
\item{CacheSubnetGroupName}{[required] The name for the cache subnet group. This value is stored as a lowercase
string.

Constraints: Must contain no more than 255 alphanumeric characters or
hyphens.

Example: \code{mysubnetgroup}}

\item{CacheSubnetGroupDescription}{A description of the cache subnet group.}

\item{SubnetIds}{The EC2 subnet IDs for the cache subnet group.}
}
\value{
A list with the following syntax:\preformatted{list(
  CacheSubnetGroup = list(
    CacheSubnetGroupName = "string",
    CacheSubnetGroupDescription = "string",
    VpcId = "string",
    Subnets = list(
      list(
        SubnetIdentifier = "string",
        SubnetAvailabilityZone = list(
          Name = "string"
        ),
        SubnetOutpost = list(
          SubnetOutpostArn = "string"
        )
      )
    ),
    ARN = "string"
  )
)
}
}
\description{
Modifies an existing cache subnet group.
}
\section{Request syntax}{
\preformatted{svc$modify_cache_subnet_group(
  CacheSubnetGroupName = "string",
  CacheSubnetGroupDescription = "string",
  SubnetIds = list(
    "string"
  )
)
}
}

\keyword{internal}
