% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_reboot_cache_cluster}
\alias{elasticache_reboot_cache_cluster}
\title{Reboots some, or all, of the cache nodes within a provisioned cluster}
\usage{
elasticache_reboot_cache_cluster(CacheClusterId, CacheNodeIdsToReboot)
}
\arguments{
\item{CacheClusterId}{[required] The cluster identifier. This parameter is stored as a lowercase string.}

\item{CacheNodeIdsToReboot}{[required] A list of cache node IDs to reboot. A node ID is a numeric identifier
(0001, 0002, etc.). To reboot an entire cluster, specify all of the
cache node IDs.}
}
\description{
Reboots some, or all, of the cache nodes within a provisioned cluster. This operation applies any modified cache parameter groups to the cluster. The reboot operation takes place as soon as possible, and results in a momentary outage to the cluster. During the reboot, the cluster status is set to REBOOTING.

See \url{https://www.paws-r-sdk.com/docs/elasticache_reboot_cache_cluster/} for full documentation.
}
\keyword{internal}
