% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_test_migration}
\alias{elasticache_test_migration}
\title{Async API to test connection between source and target replication group}
\usage{
elasticache_test_migration(ReplicationGroupId, CustomerNodeEndpointList)
}
\arguments{
\item{ReplicationGroupId}{[required] The ID of the replication group to which data is to be migrated.}

\item{CustomerNodeEndpointList}{[required] List of endpoints from which data should be migrated. List should have
only one element.}
}
\description{
Async API to test connection between source and target replication group.

See \url{https://www.paws-r-sdk.com/docs/elasticache_test_migration/} for full documentation.
}
\keyword{internal}
