% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyspaces_operations.R
\name{keyspaces_create_keyspace}
\alias{keyspaces_create_keyspace}
\title{The CreateKeyspace operation adds a new keyspace to your account}
\usage{
keyspaces_create_keyspace(keyspaceName, tags = NULL)
}
\arguments{
\item{keyspaceName}{[required] The name of the keyspace to be created.}

\item{tags}{A list of key-value pair tags to be attached to the keyspace.

For more information, see \href{https://docs.aws.amazon.com/keyspaces/latest/devguide/tagging-keyspaces.html}{Adding tags and labels to Amazon Keyspaces resources}
in the \emph{Amazon Keyspaces Developer Guide}.}
}
\description{
The \code{\link[=keyspaces_create_keyspace]{create_keyspace}} operation adds a new keyspace to your account. In an Amazon Web Services account, keyspace names must be unique within each Region.

See \url{https://paws-r.github.io/docs/keyspaces/create_keyspace.html} for full documentation.
}
\keyword{internal}
