% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyspaces_operations.R
\name{keyspaces_delete_keyspace}
\alias{keyspaces_delete_keyspace}
\title{The DeleteKeyspace operation deletes a keyspace and all of its tables}
\usage{
keyspaces_delete_keyspace(keyspaceName)
}
\arguments{
\item{keyspaceName}{[required] The name of the keyspace to be deleted.}
}
\description{
The \code{\link[=keyspaces_delete_keyspace]{delete_keyspace}} operation deletes a keyspace and all of its tables.

See \url{https://paws-r.github.io/docs/keyspaces/delete_keyspace.html} for full documentation.
}
\keyword{internal}
