% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyspaces_operations.R
\name{keyspaces_delete_type}
\alias{keyspaces_delete_type}
\title{The DeleteType operation deletes a user-defined type (UDT)}
\usage{
keyspaces_delete_type(keyspaceName, typeName)
}
\arguments{
\item{keyspaceName}{[required] The name of the keyspace of the to be deleted type.}

\item{typeName}{[required] The name of the type to be deleted.}
}
\description{
The \code{\link[=keyspaces_delete_type]{delete_type}} operation deletes a user-defined type (UDT). You can only delete a type that is not used in a table or another UDT.

See \url{https://www.paws-r-sdk.com/docs/keyspaces_delete_type/} for full documentation.
}
\keyword{internal}
