% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyspaces_operations.R
\name{keyspaces_get_table}
\alias{keyspaces_get_table}
\title{Returns information about the table, including the table's name and
current status, the keyspace name, configuration settings, and metadata}
\usage{
keyspaces_get_table(keyspaceName, tableName)
}
\arguments{
\item{keyspaceName}{[required] The name of the keyspace that the table is stored in.}

\item{tableName}{[required] The name of the table.}
}
\description{
Returns information about the table, including the table's name and current status, the keyspace name, configuration settings, and metadata.

See \url{https://paws-r.github.io/docs/keyspaces/get_table.html} for full documentation.
}
\keyword{internal}
