% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyspaces_operations.R
\name{keyspaces_get_table_auto_scaling_settings}
\alias{keyspaces_get_table_auto_scaling_settings}
\title{Returns auto scaling related settings of the specified table in JSON
format}
\usage{
keyspaces_get_table_auto_scaling_settings(keyspaceName, tableName)
}
\arguments{
\item{keyspaceName}{[required] The name of the keyspace.}

\item{tableName}{[required] The name of the table.}
}
\description{
Returns auto scaling related settings of the specified table in JSON format. If the table is a multi-Region table, the Amazon Web Services Region specific auto scaling settings of the table are included.

See \url{https://www.paws-r-sdk.com/docs/keyspaces_get_table_auto_scaling_settings/} for full documentation.
}
\keyword{internal}
