% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyspaces_operations.R
\name{keyspaces_list_tags_for_resource}
\alias{keyspaces_list_tags_for_resource}
\title{Returns a list of all tags associated with the specified Amazon
Keyspaces resource}
\usage{
keyspaces_list_tags_for_resource(
  resourceArn,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the Amazon Keyspaces resource.}

\item{nextToken}{The pagination token. To resume pagination, provide the \code{NextToken}
value as argument of a subsequent API invocation.}

\item{maxResults}{The total number of tags to return in the output. If the total number of
tags available is more than the value specified, a \code{NextToken} is
provided in the output. To resume pagination, provide the \code{NextToken}
value as an argument of a subsequent API invocation.}
}
\description{
Returns a list of all tags associated with the specified Amazon Keyspaces resource.

See \url{https://www.paws-r-sdk.com/docs/keyspaces_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
