% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyspaces_operations.R
\name{keyspaces_tag_resource}
\alias{keyspaces_tag_resource}
\title{Associates a set of tags with a Amazon Keyspaces resource}
\usage{
keyspaces_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the Amazon Keyspaces resource to which
to add tags.}

\item{tags}{[required] The tags to be assigned to the Amazon Keyspaces resource.}
}
\description{
Associates a set of tags with a Amazon Keyspaces resource. You can then activate these user-defined tags so that they appear on the Cost Management Console for cost allocation tracking. For more information, see \href{https://docs.aws.amazon.com/keyspaces/latest/devguide/tagging-keyspaces.html}{Adding tags and labels to Amazon Keyspaces resources} in the \emph{Amazon Keyspaces Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/keyspaces_tag_resource/} for full documentation.
}
\keyword{internal}
