% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyspaces_operations.R
\name{keyspaces_untag_resource}
\alias{keyspaces_untag_resource}
\title{Removes the association of tags from a Amazon Keyspaces resource}
\usage{
keyspaces_untag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Keyspaces resource that the tags will be removed from. This
value is an Amazon Resource Name (ARN).}

\item{tags}{[required] A list of existing tags to be removed from the Amazon Keyspaces
resource.}
}
\description{
Removes the association of tags from a Amazon Keyspaces resource.

See \url{https://www.paws-r-sdk.com/docs/keyspaces_untag_resource/} for full documentation.
}
\keyword{internal}
