% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyspaces_operations.R
\name{keyspaces_update_table}
\alias{keyspaces_update_table}
\title{Adds new columns to the table or updates one of the table's settings,
for example capacity mode, encryption, point-in-time recovery, or ttl
settings}
\usage{
keyspaces_update_table(
  keyspaceName,
  tableName,
  addColumns = NULL,
  capacitySpecification = NULL,
  encryptionSpecification = NULL,
  pointInTimeRecovery = NULL,
  ttl = NULL,
  defaultTimeToLive = NULL
)
}
\arguments{
\item{keyspaceName}{[required] The name of the keyspace the specified table is stored in.}

\item{tableName}{[required] The name of the table.}

\item{addColumns}{For each column to be added to the specified table:

• \code{name} - The name of the column.

• \code{type} - An Amazon Keyspaces data type. For more information, see
\href{https://docs.aws.amazon.com/keyspaces/latest/devguide/cql.elements.html#cql.data-types}{Data types}
in the \emph{Amazon Keyspaces Developer Guide}.}

\item{capacitySpecification}{Modifies the read/write throughput capacity mode for the table. The
options are:

• \code{throughputMode:PAY_PER_REQUEST} and

• \code{throughputMode:PROVISIONED} - Provisioned capacity mode requires
\code{readCapacityUnits} and \code{writeCapacityUnits} as input.

The default is \code{throughput_mode:PAY_PER_REQUEST}.

For more information, see \href{https://docs.aws.amazon.com/keyspaces/latest/devguide/ReadWriteCapacityMode.html}{Read/write capacity modes}
in the \emph{Amazon Keyspaces Developer Guide}.}

\item{encryptionSpecification}{Modifies the encryption settings of the table. You can choose one of the
following KMS key (KMS key):

• \code{type:AWS_OWNED_KMS_KEY} - This key is owned by Amazon Keyspaces.

• \code{type:CUSTOMER_MANAGED_KMS_KEY} - This key is stored in your account
and is created, owned, and managed by you. This option requires the
\code{kms_key_identifier} of the KMS key in Amazon Resource Name (ARN) format
as input.

The default is \code{AWS_OWNED_KMS_KEY}.

For more information, see \href{https://docs.aws.amazon.com/keyspaces/latest/devguide/EncryptionAtRest.html}{Encryption at rest}
in the \emph{Amazon Keyspaces Developer Guide}.}

\item{pointInTimeRecovery}{Modifies the \code{pointInTimeRecovery} settings of the table. The options
are:

• \code{ENABLED}

• \code{DISABLED}

If it's not specified, the default is \code{DISABLED}.

For more information, see \href{https://docs.aws.amazon.com/keyspaces/latest/devguide/PointInTimeRecovery.html}{Point-in-time recovery}
in the \emph{Amazon Keyspaces Developer Guide}.}

\item{ttl}{Modifies Time to Live custom settings for the table. The options are:

• \code{status:enabled}

• \code{status:disabled}

The default is \code{status:disabled}. After \code{ttl} is enabled, you can't
disable it for the table.

For more information, see \href{https://docs.aws.amazon.com/keyspaces/latest/devguide/TTL.html}{Expiring data by using Amazon Keyspaces Time to Live (TTL)}
in the \emph{Amazon Keyspaces Developer Guide}.}

\item{defaultTimeToLive}{The default Time to Live setting in seconds for the table.

For more information, see \href{https://docs.aws.amazon.com/keyspaces/latest/devguide/TTL-how-it-works.html#ttl-howitworks_default_ttl}{Setting the default TTL value for a table}
in the \emph{Amazon Keyspaces Developer Guide}.}
}
\description{
Adds new columns to the table or updates one of the table's settings, for example capacity mode, encryption, point-in-time recovery, or ttl settings. Note that you can only update one specific table setting per update operation.

See \url{https://paws-r.github.io/docs/keyspaces/update_table.html} for full documentation.
}
\keyword{internal}
