% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_add_lf_tags_to_resource}
\alias{lakeformation_add_lf_tags_to_resource}
\title{Attaches one or more LF-tags to an existing resource}
\usage{
lakeformation_add_lf_tags_to_resource(CatalogId = NULL, Resource, LFTags)
}
\arguments{
\item{CatalogId}{The identifier for the Data Catalog. By default, the account ID. The
Data Catalog is the persistent metadata store. It contains database
definitions, table definitions, and other control information to manage
your Lake Formation environment.}

\item{Resource}{[required] The database, table, or column resource to which to attach an LF-tag.}

\item{LFTags}{[required] The LF-tags to attach to the resource.}
}
\description{
Attaches one or more LF-tags to an existing resource.

See \url{https://www.paws-r-sdk.com/docs/lakeformation_add_lf_tags_to_resource/} for full documentation.
}
\keyword{internal}
