% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_batch_revoke_permissions}
\alias{lakeformation_batch_revoke_permissions}
\title{Batch operation to revoke permissions from the principal}
\usage{
lakeformation_batch_revoke_permissions(CatalogId = NULL, Entries)
}
\arguments{
\item{CatalogId}{The identifier for the Data Catalog. By default, the account ID. The
Data Catalog is the persistent metadata store. It contains database
definitions, table definitions, and other control information to manage
your Lake Formation environment.}

\item{Entries}{[required] A list of up to 20 entries for resource permissions to be revoked by
batch operation to the principal.}
}
\description{
Batch operation to revoke permissions from the principal.

See \url{https://paws-r.github.io/docs/lakeformation/batch_revoke_permissions.html} for full documentation.
}
\keyword{internal}
