% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_cancel_transaction}
\alias{lakeformation_cancel_transaction}
\title{Attempts to cancel the specified transaction}
\usage{
lakeformation_cancel_transaction(TransactionId)
}
\arguments{
\item{TransactionId}{[required] The transaction to cancel.}
}
\description{
Attempts to cancel the specified transaction. Returns an exception if the transaction was previously committed.

See \url{https://paws-r.github.io/docs/lakeformation/cancel_transaction.html} for full documentation.
}
\keyword{internal}
