% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_commit_transaction}
\alias{lakeformation_commit_transaction}
\title{Attempts to commit the specified transaction}
\usage{
lakeformation_commit_transaction(TransactionId)
}
\arguments{
\item{TransactionId}{[required] The transaction to commit.}
}
\description{
Attempts to commit the specified transaction. Returns an exception if the transaction was previously aborted. This API action is idempotent if called multiple times for the same transaction.

See \url{https://paws-r.github.io/docs/lakeformation/commit_transaction.html} for full documentation.
}
\keyword{internal}
