% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_create_data_cells_filter}
\alias{lakeformation_create_data_cells_filter}
\title{Creates a data cell filter to allow one to grant access to certain
columns on certain rows}
\usage{
lakeformation_create_data_cells_filter(TableData)
}
\arguments{
\item{TableData}{[required] A \code{DataCellsFilter} structure containing information about the data
cells filter.}
}
\description{
Creates a data cell filter to allow one to grant access to certain columns on certain rows.

See \url{https://paws-r.github.io/docs/lakeformation/create_data_cells_filter.html} for full documentation.
}
\keyword{internal}
