% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_create_lf_tag}
\alias{lakeformation_create_lf_tag}
\title{Creates an LF-tag with the specified name and values}
\usage{
lakeformation_create_lf_tag(CatalogId = NULL, TagKey, TagValues)
}
\arguments{
\item{CatalogId}{The identifier for the Data Catalog. By default, the account ID. The
Data Catalog is the persistent metadata store. It contains database
definitions, table definitions, and other control information to manage
your Lake Formation environment.}

\item{TagKey}{[required] The key-name for the LF-tag.}

\item{TagValues}{[required] A list of possible values an attribute can take.}
}
\description{
Creates an LF-tag with the specified name and values.

See \url{https://www.paws-r-sdk.com/docs/lakeformation_create_lf_tag/} for full documentation.
}
\keyword{internal}
