% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_delete_data_cells_filter}
\alias{lakeformation_delete_data_cells_filter}
\title{Deletes a data cell filter}
\usage{
lakeformation_delete_data_cells_filter(
  TableCatalogId = NULL,
  DatabaseName = NULL,
  TableName = NULL,
  Name = NULL
)
}
\arguments{
\item{TableCatalogId}{The ID of the catalog to which the table belongs.}

\item{DatabaseName}{A database in the Glue Data Catalog.}

\item{TableName}{A table in the database.}

\item{Name}{The name given by the user to the data filter cell.}
}
\description{
Deletes a data cell filter.

See \url{https://paws-r.github.io/docs/lakeformation/delete_data_cells_filter.html} for full documentation.
}
\keyword{internal}
