% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_delete_lake_formation_identity_center_configuration}
\alias{lakeformation_delete_lake_formation_identity_center_configuration}
\title{Deletes an IAM Identity Center connection with Lake Formation}
\usage{
lakeformation_delete_lake_formation_identity_center_configuration(
  CatalogId = NULL
)
}
\arguments{
\item{CatalogId}{The identifier for the Data Catalog. By default, the account ID. The
Data Catalog is the persistent metadata store. It contains database
definitions, table definitions, view definition, and other control
information to manage your Lake Formation environment.}
}
\description{
Deletes an IAM Identity Center connection with Lake Formation.

See \url{https://www.paws-r-sdk.com/docs/lakeformation_delete_lake_formation_identity_center_configuration/} for full documentation.
}
\keyword{internal}
