% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_delete_lf_tag}
\alias{lakeformation_delete_lf_tag}
\title{Deletes the specified LF-tag given a key name}
\usage{
lakeformation_delete_lf_tag(CatalogId = NULL, TagKey)
}
\arguments{
\item{CatalogId}{The identifier for the Data Catalog. By default, the account ID. The
Data Catalog is the persistent metadata store. It contains database
definitions, table definitions, and other control information to manage
your Lake Formation environment.}

\item{TagKey}{[required] The key-name for the LF-tag to delete.}
}
\description{
Deletes the specified LF-tag given a key name. If the input parameter tag key was not found, then the operation will throw an exception. When you delete an LF-tag, the \code{LFTagPolicy} attached to the LF-tag becomes invalid. If the deleted LF-tag was still assigned to any resource, the tag policy attach to the deleted LF-tag will no longer be applied to the resource.

See \url{https://www.paws-r-sdk.com/docs/lakeformation_delete_lf_tag/} for full documentation.
}
\keyword{internal}
