% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_describe_lake_formation_identity_center_configuration}
\alias{lakeformation_describe_lake_formation_identity_center_configuration}
\title{Retrieves the instance ARN and application ARN for the connection}
\usage{
lakeformation_describe_lake_formation_identity_center_configuration(
  CatalogId = NULL
)
}
\arguments{
\item{CatalogId}{The identifier for the Data Catalog. By default, the account ID. The
Data Catalog is the persistent metadata store. It contains database
definitions, table definitions, and other control information to manage
your Lake Formation environment.}
}
\description{
Retrieves the instance ARN and application ARN for the connection.

See \url{https://www.paws-r-sdk.com/docs/lakeformation_describe_lake_formation_identity_center_configuration/} for full documentation.
}
\keyword{internal}
