% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_describe_transaction}
\alias{lakeformation_describe_transaction}
\title{Returns the details of a single transaction}
\usage{
lakeformation_describe_transaction(TransactionId)
}
\arguments{
\item{TransactionId}{[required] The transaction for which to return status.}
}
\description{
Returns the details of a single transaction.

See \url{https://www.paws-r-sdk.com/docs/lakeformation_describe_transaction/} for full documentation.
}
\keyword{internal}
