% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_extend_transaction}
\alias{lakeformation_extend_transaction}
\title{Indicates to the service that the specified transaction is still active
and should not be treated as idle and aborted}
\usage{
lakeformation_extend_transaction(TransactionId = NULL)
}
\arguments{
\item{TransactionId}{The transaction to extend.}
}
\description{
Indicates to the service that the specified transaction is still active and should not be treated as idle and aborted.

See \url{https://www.paws-r-sdk.com/docs/lakeformation_extend_transaction/} for full documentation.
}
\keyword{internal}
