% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_get_data_cells_filter}
\alias{lakeformation_get_data_cells_filter}
\title{Returns a data cells filter}
\usage{
lakeformation_get_data_cells_filter(
  TableCatalogId,
  DatabaseName,
  TableName,
  Name
)
}
\arguments{
\item{TableCatalogId}{[required] The ID of the catalog to which the table belongs.}

\item{DatabaseName}{[required] A database in the Glue Data Catalog.}

\item{TableName}{[required] A table in the database.}

\item{Name}{[required] The name given by the user to the data filter cell.}
}
\description{
Returns a data cells filter.

See \url{https://www.paws-r-sdk.com/docs/lakeformation_get_data_cells_filter/} for full documentation.
}
\keyword{internal}
