% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_get_data_lake_principal}
\alias{lakeformation_get_data_lake_principal}
\title{Returns the identity of the invoking principal}
\usage{
lakeformation_get_data_lake_principal()
}
\description{
Returns the identity of the invoking principal.

See \url{https://www.paws-r-sdk.com/docs/lakeformation_get_data_lake_principal/} for full documentation.
}
\keyword{internal}
