% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_get_data_lake_settings}
\alias{lakeformation_get_data_lake_settings}
\title{Retrieves the list of the data lake administrators of a Lake
Formation-managed data lake}
\usage{
lakeformation_get_data_lake_settings(CatalogId = NULL)
}
\arguments{
\item{CatalogId}{The identifier for the Data Catalog. By default, the account ID. The
Data Catalog is the persistent metadata store. It contains database
definitions, table definitions, and other control information to manage
your Lake Formation environment.}
}
\description{
Retrieves the list of the data lake administrators of a Lake Formation-managed data lake.

See \url{https://paws-r.github.io/docs/lakeformation/get_data_lake_settings.html} for full documentation.
}
\keyword{internal}
