% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_get_lf_tag_expression}
\alias{lakeformation_get_lf_tag_expression}
\title{Returns the details about the LF-Tag expression}
\usage{
lakeformation_get_lf_tag_expression(Name, CatalogId = NULL)
}
\arguments{
\item{Name}{[required] The name for the LF-Tag expression}

\item{CatalogId}{The identifier for the Data Catalog. By default, the account ID.}
}
\description{
Returns the details about the LF-Tag expression. The caller must be a data lake admin or must have \code{DESCRIBE} permission on the LF-Tag expression resource.

See \url{https://www.paws-r-sdk.com/docs/lakeformation_get_lf_tag_expression/} for full documentation.
}
\keyword{internal}
