% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_get_query_state}
\alias{lakeformation_get_query_state}
\title{Returns the state of a query previously submitted}
\usage{
lakeformation_get_query_state(QueryId)
}
\arguments{
\item{QueryId}{[required] The ID of the plan query operation.}
}
\description{
Returns the state of a query previously submitted. Clients are expected to poll \code{\link[=lakeformation_get_query_state]{get_query_state}} to monitor the current state of the planning before retrieving the work units. A query state is only visible to the principal that made the initial call to \code{\link[=lakeformation_start_query_planning]{start_query_planning}}.

See \url{https://paws-r.github.io/docs/lakeformation/get_query_state.html} for full documentation.
}
\keyword{internal}
