% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_get_query_statistics}
\alias{lakeformation_get_query_statistics}
\title{Retrieves statistics on the planning and execution of a query}
\usage{
lakeformation_get_query_statistics(QueryId)
}
\arguments{
\item{QueryId}{[required] The ID of the plan query operation.}
}
\description{
Retrieves statistics on the planning and execution of a query.

See \url{https://paws-r.github.io/docs/lakeformation/get_query_statistics.html} for full documentation.
}
\keyword{internal}
