% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_get_resource_lf_tags}
\alias{lakeformation_get_resource_lf_tags}
\title{Returns the LF-tags applied to a resource}
\usage{
lakeformation_get_resource_lf_tags(
  CatalogId = NULL,
  Resource,
  ShowAssignedLFTags = NULL
)
}
\arguments{
\item{CatalogId}{The identifier for the Data Catalog. By default, the account ID. The
Data Catalog is the persistent metadata store. It contains database
definitions, table definitions, and other control information to manage
your Lake Formation environment.}

\item{Resource}{[required] The database, table, or column resource for which you want to return
LF-tags.}

\item{ShowAssignedLFTags}{Indicates whether to show the assigned LF-tags.}
}
\description{
Returns the LF-tags applied to a resource.

See \url{https://paws-r.github.io/docs/lakeformation/get_resource_lf_tags.html} for full documentation.
}
\keyword{internal}
