% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_get_temporary_data_location_credentials}
\alias{lakeformation_get_temporary_data_location_credentials}
\title{Allows a user or application in a secure environment to access data in a
specific Amazon S3 location registered with Lake Formation by providing
temporary scoped credentials that are limited to the requested data
location and the caller's authorized access level}
\usage{
lakeformation_get_temporary_data_location_credentials(
  DurationSeconds = NULL,
  AuditContext = NULL,
  DataLocations = NULL,
  CredentialsScope = NULL
)
}
\arguments{
\item{DurationSeconds}{The time period, between 900 and 43,200 seconds, for the timeout of the
temporary credentials.}

\item{DataLocations}{The Amazon S3 data location that you want to access.}

\item{CredentialsScope}{The credential scope is determined by the caller's Lake Formation
permission on the associated table. Credential scope can be either:
\itemize{
\item READ - Provides read-only access to the data location.
\item READ_WRITE - Provides both read and write access to the data
location.
}}
}
\description{
Allows a user or application in a secure environment to access data in a specific Amazon S3 location registered with Lake Formation by providing temporary scoped credentials that are limited to the requested data location and the caller's authorized access level.

See \url{https://www.paws-r-sdk.com/docs/lakeformation_get_temporary_data_location_credentials/} for full documentation.
}
\keyword{internal}
