% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_get_work_unit_results}
\alias{lakeformation_get_work_unit_results}
\title{Returns the work units resulting from the query}
\usage{
lakeformation_get_work_unit_results(QueryId, WorkUnitId, WorkUnitToken)
}
\arguments{
\item{QueryId}{[required] The ID of the plan query operation for which to get results.}

\item{WorkUnitId}{[required] The work unit ID for which to get results. Value generated by
enumerating \code{WorkUnitIdMin} to \code{WorkUnitIdMax} (inclusive) from the
\code{WorkUnitRange} in the output of
\code{\link[=lakeformation_get_work_units]{get_work_units}}.}

\item{WorkUnitToken}{[required] A work token used to query the execution service. Token output from
\code{\link[=lakeformation_get_work_units]{get_work_units}}.}
}
\description{
Returns the work units resulting from the query. Work units can be executed in any order and in parallel.

See \url{https://www.paws-r-sdk.com/docs/lakeformation_get_work_unit_results/} for full documentation.
}
\keyword{internal}
