% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_list_data_cells_filter}
\alias{lakeformation_list_data_cells_filter}
\title{Lists all the data cell filters on a table}
\usage{
lakeformation_list_data_cells_filter(
  Table = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Table}{A table in the Glue Data Catalog.}

\item{NextToken}{A continuation token, if this is a continuation call.}

\item{MaxResults}{The maximum size of the response.}
}
\description{
Lists all the data cell filters on a table.

See \url{https://paws-r.github.io/docs/lakeformation/list_data_cells_filter.html} for full documentation.
}
\keyword{internal}
