% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_list_lake_formation_opt_ins}
\alias{lakeformation_list_lake_formation_opt_ins}
\title{Retrieve the current list of resources and principals that are opt in to
enforce Lake Formation permissions}
\usage{
lakeformation_list_lake_formation_opt_ins(
  Principal = NULL,
  Resource = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Principal}{}

\item{Resource}{A structure for the resource.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{A continuation token, if this is not the first call to retrieve this
list.}
}
\description{
Retrieve the current list of resources and principals that are opt in to enforce Lake Formation permissions.

See \url{https://www.paws-r-sdk.com/docs/lakeformation_list_lake_formation_opt_ins/} for full documentation.
}
\keyword{internal}
