% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_list_lf_tag_expressions}
\alias{lakeformation_list_lf_tag_expressions}
\title{Returns the LF-Tag expressions in caller’s account filtered based on
caller's permissions}
\usage{
lakeformation_list_lf_tag_expressions(
  CatalogId = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{CatalogId}{The identifier for the Data Catalog. By default, the account ID.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{A continuation token, if this is not the first call to retrieve this
list.}
}
\description{
Returns the LF-Tag expressions in caller’s account filtered based on caller's permissions. Data Lake and read only admins implicitly can see all tag expressions in their account, else caller needs DESCRIBE permissions on tag expression.

See \url{https://www.paws-r-sdk.com/docs/lakeformation_list_lf_tag_expressions/} for full documentation.
}
\keyword{internal}
