% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_register_resource}
\alias{lakeformation_register_resource}
\title{Registers the resource as managed by the Data Catalog}
\usage{
lakeformation_register_resource(
  ResourceArn,
  UseServiceLinkedRole = NULL,
  RoleArn = NULL
)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to
register.}

\item{UseServiceLinkedRole}{Designates an Identity and Access Management (IAM) service-linked role
by registering this role with the Data Catalog. A service-linked role is
a unique type of IAM role that is linked directly to Lake Formation.

For more information, see Using Service-Linked Roles for Lake Formation.}

\item{RoleArn}{The identifier for the role that registers the resource.}
}
\description{
Registers the resource as managed by the Data Catalog.

See \url{https://paws-r.github.io/docs/lakeformation/register_resource.html} for full documentation.
}
\keyword{internal}
