% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_remove_lf_tags_from_resource}
\alias{lakeformation_remove_lf_tags_from_resource}
\title{Removes an LF-tag from the resource}
\usage{
lakeformation_remove_lf_tags_from_resource(CatalogId = NULL, Resource, LFTags)
}
\arguments{
\item{CatalogId}{The identifier for the Data Catalog. By default, the account ID. The
Data Catalog is the persistent metadata store. It contains database
definitions, table definitions, and other control information to manage
your Lake Formation environment.}

\item{Resource}{[required] The database, table, or column resource where you want to remove an
LF-tag.}

\item{LFTags}{[required] The LF-tags to be removed from the resource.}
}
\description{
Removes an LF-tag from the resource. Only database, table, or tableWithColumns resource are allowed. To tag columns, use the column inclusion list in \code{tableWithColumns} to specify column input.

See \url{https://www.paws-r-sdk.com/docs/lakeformation_remove_lf_tags_from_resource/} for full documentation.
}
\keyword{internal}
