% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_start_query_planning}
\alias{lakeformation_start_query_planning}
\title{Submits a request to process a query statement}
\usage{
lakeformation_start_query_planning(QueryPlanningContext, QueryString)
}
\arguments{
\item{QueryPlanningContext}{[required] A structure containing information about the query plan.}

\item{QueryString}{[required] A PartiQL query statement used as an input to the planner service.}
}
\description{
Submits a request to process a query statement.

See \url{https://www.paws-r-sdk.com/docs/lakeformation_start_query_planning/} for full documentation.
}
\keyword{internal}
