% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_start_transaction}
\alias{lakeformation_start_transaction}
\title{Starts a new transaction and returns its transaction ID}
\usage{
lakeformation_start_transaction(TransactionType = NULL)
}
\arguments{
\item{TransactionType}{Indicates whether this transaction should be read only or read and
write. Writes made using a read-only transaction ID will be rejected.
Read-only transactions do not need to be committed.}
}
\description{
Starts a new transaction and returns its transaction ID. Transaction IDs are opaque objects that you can use to identify a transaction.

See \url{https://paws-r.github.io/docs/lakeformation/start_transaction.html} for full documentation.
}
\keyword{internal}
