% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_update_lake_formation_identity_center_configuration}
\alias{lakeformation_update_lake_formation_identity_center_configuration}
\title{Updates the IAM Identity Center connection parameters}
\usage{
lakeformation_update_lake_formation_identity_center_configuration(
  CatalogId = NULL,
  ApplicationStatus = NULL,
  ExternalFiltering = NULL
)
}
\arguments{
\item{CatalogId}{The identifier for the Data Catalog. By default, the account ID. The
Data Catalog is the persistent metadata store. It contains database
definitions, table definitions, view definitions, and other control
information to manage your Lake Formation environment.}

\item{ApplicationStatus}{Allows to enable or disable the IAM Identity Center connection.}

\item{ExternalFiltering}{A list of the account IDs of Amazon Web Services accounts of third-party
applications that are allowed to access data managed by Lake Formation.}
}
\description{
Updates the IAM Identity Center connection parameters.

See \url{https://www.paws-r-sdk.com/docs/lakeformation_update_lake_formation_identity_center_configuration/} for full documentation.
}
\keyword{internal}
