% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_update_resource}
\alias{lakeformation_update_resource}
\title{Updates the data access role used for vending access to the given
(registered) resource in Lake Formation}
\usage{
lakeformation_update_resource(
  RoleArn,
  ResourceArn,
  WithFederation = NULL,
  HybridAccessEnabled = NULL
)
}
\arguments{
\item{RoleArn}{[required] The new role to use for the given resource registered in Lake Formation.}

\item{ResourceArn}{[required] The resource ARN.}

\item{WithFederation}{Whether or not the resource is a federated resource.}

\item{HybridAccessEnabled}{Specifies whether the data access of tables pointing to the location can
be managed by both Lake Formation permissions as well as Amazon S3
bucket policies.}
}
\description{
Updates the data access role used for vending access to the given (registered) resource in Lake Formation.

See \url{https://www.paws-r-sdk.com/docs/lakeformation_update_resource/} for full documentation.
}
\keyword{internal}
