% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_update_table_objects}
\alias{lakeformation_update_table_objects}
\title{Updates the manifest of Amazon S3 objects that make up the specified
governed table}
\usage{
lakeformation_update_table_objects(
  CatalogId = NULL,
  DatabaseName,
  TableName,
  TransactionId = NULL,
  WriteOperations
)
}
\arguments{
\item{CatalogId}{The catalog containing the governed table to update. Defaults to the
caller’s account ID.}

\item{DatabaseName}{[required] The database containing the governed table to update.}

\item{TableName}{[required] The governed table to update.}

\item{TransactionId}{The transaction at which to do the write.}

\item{WriteOperations}{[required] A list of \code{WriteOperation} objects that define an object to add to or
delete from the manifest for a governed table.}
}
\description{
Updates the manifest of Amazon S3 objects that make up the specified governed table.

See \url{https://paws-r.github.io/docs/lakeformation/update_table_objects.html} for full documentation.
}
\keyword{internal}
