% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_update_table_storage_optimizer}
\alias{lakeformation_update_table_storage_optimizer}
\title{Updates the configuration of the storage optimizers for a table}
\usage{
lakeformation_update_table_storage_optimizer(
  CatalogId = NULL,
  DatabaseName,
  TableName,
  StorageOptimizerConfig
)
}
\arguments{
\item{CatalogId}{The Catalog ID of the table.}

\item{DatabaseName}{[required] Name of the database where the table is present.}

\item{TableName}{[required] Name of the table for which to enable the storage optimizer.}

\item{StorageOptimizerConfig}{[required] Name of the table for which to enable the storage optimizer.}
}
\description{
Updates the configuration of the storage optimizers for a table.

See \url{https://paws-r.github.io/docs/lakeformation/update_table_storage_optimizer.html} for full documentation.
}
\keyword{internal}
