% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_batch_update_cluster}
\alias{memorydb_batch_update_cluster}
\title{Apply the service update to a list of clusters supplied}
\usage{
memorydb_batch_update_cluster(ClusterNames, ServiceUpdate = NULL)
}
\arguments{
\item{ClusterNames}{[required] The cluster names to apply the updates.}

\item{ServiceUpdate}{The unique ID of the service update}
}
\description{
Apply the service update to a list of clusters supplied. For more information on service updates and applying them, see \href{https://docs.aws.amazon.com/memorydb/latest/devguide/managing-updates.html#applying-updates}{Applying the service updates}.

See \url{https://paws-r.github.io/docs/memorydb/batch_update_cluster.html} for full documentation.
}
\keyword{internal}
