% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_create_acl}
\alias{memorydb_create_acl}
\title{Creates an Access Control List}
\usage{
memorydb_create_acl(ACLName, UserNames = NULL, Tags = NULL)
}
\arguments{
\item{ACLName}{[required] The name of the Access Control List.}

\item{UserNames}{The list of users that belong to the Access Control List.}

\item{Tags}{A list of tags to be added to this resource. A tag is a key-value pair.
A tag key must be accompanied by a tag value, although null is accepted.}
}
\description{
Creates an Access Control List. For more information, see \href{https://docs.aws.amazon.com/memorydb/latest/devguide/clusters.acls.html}{Authenticating users with Access Contol Lists (ACLs)}.

See \url{https://www.paws-r-sdk.com/docs/memorydb_create_acl/} for full documentation.
}
\keyword{internal}
