% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_create_multi_region_cluster}
\alias{memorydb_create_multi_region_cluster}
\title{Creates a new multi-Region cluster}
\usage{
memorydb_create_multi_region_cluster(
  MultiRegionClusterNameSuffix,
  Description = NULL,
  Engine = NULL,
  EngineVersion = NULL,
  NodeType,
  MultiRegionParameterGroupName = NULL,
  NumShards = NULL,
  TLSEnabled = NULL,
  Tags = NULL
)
}
\arguments{
\item{MultiRegionClusterNameSuffix}{[required] A suffix to be added to the Multi-Region cluster name. Amazon MemoryDB
automatically applies a prefix to the Multi-Region cluster Name when it
is created. Each Amazon Region has its own prefix. For instance, a
Multi-Region cluster Name created in the US-West-1 region will begin
with "virxk", along with the suffix name you provide. The suffix
guarantees uniqueness of the Multi-Region cluster name across multiple
regions.}

\item{Description}{A description for the multi-Region cluster.}

\item{Engine}{The name of the engine to be used for the multi-Region cluster.}

\item{EngineVersion}{The version of the engine to be used for the multi-Region cluster.}

\item{NodeType}{[required] The node type to be used for the multi-Region cluster.}

\item{MultiRegionParameterGroupName}{The name of the multi-Region parameter group to be associated with the
cluster.}

\item{NumShards}{The number of shards for the multi-Region cluster.}

\item{TLSEnabled}{Whether to enable TLS encryption for the multi-Region cluster.}

\item{Tags}{A list of tags to be applied to the multi-Region cluster.}
}
\description{
Creates a new multi-Region cluster.

See \url{https://www.paws-r-sdk.com/docs/memorydb_create_multi_region_cluster/} for full documentation.
}
\keyword{internal}
