% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_create_snapshot}
\alias{memorydb_create_snapshot}
\title{Creates a copy of an entire cluster at a specific moment in time}
\usage{
memorydb_create_snapshot(
  ClusterName,
  SnapshotName,
  KmsKeyId = NULL,
  Tags = NULL
)
}
\arguments{
\item{ClusterName}{[required] The snapshot is created from this cluster.}

\item{SnapshotName}{[required] A name for the snapshot being created.}

\item{KmsKeyId}{The ID of the KMS key used to encrypt the snapshot.}

\item{Tags}{A list of tags to be added to this resource. A tag is a key-value pair.
A tag key must be accompanied by a tag value, although null is accepted.}
}
\description{
Creates a copy of an entire cluster at a specific moment in time.

See \url{https://www.paws-r-sdk.com/docs/memorydb_create_snapshot/} for full documentation.
}
\keyword{internal}
