% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_delete_acl}
\alias{memorydb_delete_acl}
\title{Deletes an Access Control List}
\usage{
memorydb_delete_acl(ACLName)
}
\arguments{
\item{ACLName}{[required] The name of the Access Control List to delete}
}
\description{
Deletes an Access Control List. The ACL must first be disassociated from the cluster before it can be deleted. For more information, see \href{https://docs.aws.amazon.com/memorydb/latest/devguide/clusters.acls.html}{Authenticating users with Access Contol Lists (ACLs)}.

See \url{https://paws-r.github.io/docs/memorydb/delete_acl.html} for full documentation.
}
\keyword{internal}
