% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_delete_multi_region_cluster}
\alias{memorydb_delete_multi_region_cluster}
\title{Deletes an existing multi-Region cluster}
\usage{
memorydb_delete_multi_region_cluster(MultiRegionClusterName)
}
\arguments{
\item{MultiRegionClusterName}{[required] The name of the multi-Region cluster to be deleted.}
}
\description{
Deletes an existing multi-Region cluster.

See \url{https://www.paws-r-sdk.com/docs/memorydb_delete_multi_region_cluster/} for full documentation.
}
\keyword{internal}
