% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_delete_parameter_group}
\alias{memorydb_delete_parameter_group}
\title{Deletes the specified parameter group}
\usage{
memorydb_delete_parameter_group(ParameterGroupName)
}
\arguments{
\item{ParameterGroupName}{[required] The name of the parameter group to delete.}
}
\description{
Deletes the specified parameter group. You cannot delete a parameter group if it is associated with any clusters. You cannot delete the default parameter groups in your account.

See \url{https://www.paws-r-sdk.com/docs/memorydb_delete_parameter_group/} for full documentation.
}
\keyword{internal}
