% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_delete_snapshot}
\alias{memorydb_delete_snapshot}
\title{Deletes an existing snapshot}
\usage{
memorydb_delete_snapshot(SnapshotName)
}
\arguments{
\item{SnapshotName}{[required] The name of the snapshot to delete}
}
\description{
Deletes an existing snapshot. When you receive a successful response from this operation, MemoryDB immediately begins deleting the snapshot; you cannot cancel or revert this operation.

See \url{https://www.paws-r-sdk.com/docs/memorydb_delete_snapshot/} for full documentation.
}
\keyword{internal}
