% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_delete_subnet_group}
\alias{memorydb_delete_subnet_group}
\title{Deletes a subnet group}
\usage{
memorydb_delete_subnet_group(SubnetGroupName)
}
\arguments{
\item{SubnetGroupName}{[required] The name of the subnet group to delete.}
}
\description{
Deletes a subnet group. You cannot delete a default subnet group or one that is associated with any clusters.

See \url{https://www.paws-r-sdk.com/docs/memorydb_delete_subnet_group/} for full documentation.
}
\keyword{internal}
