% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_delete_user}
\alias{memorydb_delete_user}
\title{Deletes a user}
\usage{
memorydb_delete_user(UserName)
}
\arguments{
\item{UserName}{[required] The name of the user to delete}
}
\description{
Deletes a user. The user will be removed from all ACLs and in turn removed from all clusters.

See \url{https://paws-r.github.io/docs/memorydb/delete_user.html} for full documentation.
}
\keyword{internal}
