% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_describe_multi_region_clusters}
\alias{memorydb_describe_multi_region_clusters}
\title{Returns details about one or more multi-Region clusters}
\usage{
memorydb_describe_multi_region_clusters(
  MultiRegionClusterName = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  ShowClusterDetails = NULL
)
}
\arguments{
\item{MultiRegionClusterName}{The name of a specific multi-Region cluster to describe.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{A token to specify where to start paginating.}

\item{ShowClusterDetails}{Details about the multi-Region cluster.}
}
\description{
Returns details about one or more multi-Region clusters.

See \url{https://www.paws-r-sdk.com/docs/memorydb_describe_multi_region_clusters/} for full documentation.
}
\keyword{internal}
