% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_describe_multi_region_parameter_groups}
\alias{memorydb_describe_multi_region_parameter_groups}
\title{Returns a list of multi-region parameter groups}
\usage{
memorydb_describe_multi_region_parameter_groups(
  MultiRegionParameterGroupName = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MultiRegionParameterGroupName}{The request for information on a specific multi-region parameter group.}

\item{MaxResults}{The maximum number of records to include in the response. If more
records exist than the specified MaxResults value, a token is included
in the response so that the remaining results can be retrieved.}

\item{NextToken}{An optional token returned from a prior request. Use this token for
pagination of results from this action. If this parameter is specified,
the response includes only results beyond the token, up to the value
specified by MaxResults.}
}
\description{
Returns a list of multi-region parameter groups.

See \url{https://www.paws-r-sdk.com/docs/memorydb_describe_multi_region_parameter_groups/} for full documentation.
}
\keyword{internal}
