% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_describe_service_updates}
\alias{memorydb_describe_service_updates}
\title{Returns details of the service updates}
\usage{
memorydb_describe_service_updates(
  ServiceUpdateName = NULL,
  ClusterNames = NULL,
  Status = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ServiceUpdateName}{The unique ID of the service update to describe.}

\item{ClusterNames}{The list of cluster names to identify service updates to apply}

\item{Status}{The status(es) of the service updates to filter on}

\item{MaxResults}{The maximum number of records to include in the response. If more
records exist than the specified MaxResults value, a token is included
in the response so that the remaining results can be retrieved.}

\item{NextToken}{An optional argument to pass in case the total number of records exceeds
the value of MaxResults. If nextToken is returned, there are more
results available. The value of nextToken is a unique pagination token
for each page. Make the call again using the returned token to retrieve
the next page. Keep all other arguments unchanged.}
}
\description{
Returns details of the service updates

See \url{https://paws-r.github.io/docs/memorydb/describe_service_updates.html} for full documentation.
}
\keyword{internal}
