% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_describe_subnet_groups}
\alias{memorydb_describe_subnet_groups}
\title{Returns a list of subnet group descriptions}
\usage{
memorydb_describe_subnet_groups(
  SubnetGroupName = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{SubnetGroupName}{The name of the subnet group to return details for.}

\item{MaxResults}{The maximum number of records to include in the response. If more
records exist than the specified MaxResults value, a token is included
in the response so that the remaining results can be retrieved.}

\item{NextToken}{An optional argument to pass in case the total number of records exceeds
the value of MaxResults. If nextToken is returned, there are more
results available. The value of nextToken is a unique pagination token
for each page. Make the call again using the returned token to retrieve
the next page. Keep all other arguments unchanged.}
}
\description{
Returns a list of subnet group descriptions. If a subnet group name is specified, the list contains only the description of that group.

See \url{https://www.paws-r-sdk.com/docs/memorydb_describe_subnet_groups/} for full documentation.
}
\keyword{internal}
