% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_failover_shard}
\alias{memorydb_failover_shard}
\title{Used to failover a shard}
\usage{
memorydb_failover_shard(ClusterName, ShardName)
}
\arguments{
\item{ClusterName}{[required] The cluster being failed over}

\item{ShardName}{[required] The name of the shard}
}
\description{
Used to failover a shard

See \url{https://paws-r.github.io/docs/memorydb/failover_shard.html} for full documentation.
}
\keyword{internal}
