% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_list_allowed_multi_region_cluster_updates}
\alias{memorydb_list_allowed_multi_region_cluster_updates}
\title{Lists the allowed updates for a multi-Region cluster}
\usage{
memorydb_list_allowed_multi_region_cluster_updates(MultiRegionClusterName)
}
\arguments{
\item{MultiRegionClusterName}{[required] The name of the multi-Region cluster.}
}
\description{
Lists the allowed updates for a multi-Region cluster.

See \url{https://www.paws-r-sdk.com/docs/memorydb_list_allowed_multi_region_cluster_updates/} for full documentation.
}
\keyword{internal}
