% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_purchase_reserved_nodes_offering}
\alias{memorydb_purchase_reserved_nodes_offering}
\title{Allows you to purchase a reserved node offering}
\usage{
memorydb_purchase_reserved_nodes_offering(
  ReservedNodesOfferingId,
  ReservationId = NULL,
  NodeCount = NULL,
  Tags = NULL
)
}
\arguments{
\item{ReservedNodesOfferingId}{[required] The ID of the reserved node offering to purchase.}

\item{ReservationId}{A customer-specified identifier to track this reservation.}

\item{NodeCount}{The number of node instances to reserve.}

\item{Tags}{A list of tags to be added to this resource. A tag is a key-value pair.
A tag key must be accompanied by a tag value, although null is accepted.}
}
\description{
Allows you to purchase a reserved node offering. Reserved nodes are not eligible for cancellation and are non-refundable.

See \url{https://paws-r.github.io/docs/memorydb/purchase_reserved_nodes_offering.html} for full documentation.
}
\keyword{internal}
