% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_untag_resource}
\alias{memorydb_untag_resource}
\title{Use this operation to remove tags on a resource}
\usage{
memorydb_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource to which the tags are to
be removed.}

\item{TagKeys}{[required] The list of keys of the tags that are to be removed.}
}
\description{
Use this operation to remove tags on a resource. A tag is a key-value pair where the key and value are case-sensitive. You can use tags to categorize and track all your MemoryDB resources. For more information, see \href{https://docs.aws.amazon.com/memorydb/latest/devguide/}{Tagging your MemoryDB resources}.

See \url{https://www.paws-r-sdk.com/docs/memorydb_untag_resource/} for full documentation.
}
\keyword{internal}
