% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_update_acl}
\alias{memorydb_update_acl}
\title{Changes the list of users that belong to the Access Control List}
\usage{
memorydb_update_acl(ACLName, UserNamesToAdd = NULL, UserNamesToRemove = NULL)
}
\arguments{
\item{ACLName}{[required] The name of the Access Control List}

\item{UserNamesToAdd}{The list of users to add to the Access Control List}

\item{UserNamesToRemove}{The list of users to remove from the Access Control List}
}
\description{
Changes the list of users that belong to the Access Control List.

See \url{https://paws-r.github.io/docs/memorydb/update_acl.html} for full documentation.
}
\keyword{internal}
