% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_update_multi_region_cluster}
\alias{memorydb_update_multi_region_cluster}
\title{Updates the configuration of an existing multi-Region cluster}
\usage{
memorydb_update_multi_region_cluster(
  MultiRegionClusterName,
  NodeType = NULL,
  Description = NULL,
  EngineVersion = NULL,
  ShardConfiguration = NULL,
  MultiRegionParameterGroupName = NULL,
  UpdateStrategy = NULL
)
}
\arguments{
\item{MultiRegionClusterName}{[required] The name of the multi-Region cluster to be updated.}

\item{NodeType}{The new node type to be used for the multi-Region cluster.}

\item{Description}{A new description for the multi-Region cluster.}

\item{EngineVersion}{The new engine version to be used for the multi-Region cluster.}

\item{ShardConfiguration}{}

\item{MultiRegionParameterGroupName}{The new multi-Region parameter group to be associated with the cluster.}

\item{UpdateStrategy}{The strategy to use for the update operation. Supported values are
"coordinated" or "uncoordinated".}
}
\description{
Updates the configuration of an existing multi-Region cluster.

See \url{https://www.paws-r-sdk.com/docs/memorydb_update_multi_region_cluster/} for full documentation.
}
\keyword{internal}
