% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_update_parameter_group}
\alias{memorydb_update_parameter_group}
\title{Updates the parameters of a parameter group}
\usage{
memorydb_update_parameter_group(ParameterGroupName, ParameterNameValues)
}
\arguments{
\item{ParameterGroupName}{[required] The name of the parameter group to update.}

\item{ParameterNameValues}{[required] An array of parameter names and values for the parameter update. You
must supply at least one parameter name and value; subsequent arguments
are optional. A maximum of 20 parameters may be updated per request.}
}
\description{
Updates the parameters of a parameter group. You can modify up to 20 parameters in a single request by submitting a list parameter name and value pairs.

See \url{https://paws-r.github.io/docs/memorydb/update_parameter_group.html} for full documentation.
}
\keyword{internal}
