% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_update_subnet_group}
\alias{memorydb_update_subnet_group}
\title{Updates a subnet group}
\usage{
memorydb_update_subnet_group(
  SubnetGroupName,
  Description = NULL,
  SubnetIds = NULL
)
}
\arguments{
\item{SubnetGroupName}{[required] The name of the subnet group}

\item{Description}{A description of the subnet group}

\item{SubnetIds}{The EC2 subnet IDs for the subnet group.}
}
\description{
Updates a subnet group. For more information, see \href{https://docs.aws.amazon.com/memorydb/latest/devguide/}{Updating a subnet group}

See \url{https://www.paws-r-sdk.com/docs/memorydb_update_subnet_group/} for full documentation.
}
\keyword{internal}
