% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_update_user}
\alias{memorydb_update_user}
\title{Changes user password(s) and/or access string}
\usage{
memorydb_update_user(UserName, AuthenticationMode = NULL, AccessString = NULL)
}
\arguments{
\item{UserName}{[required] The name of the user}

\item{AuthenticationMode}{Denotes the user's authentication properties, such as whether it
requires a password to authenticate.}

\item{AccessString}{Access permissions string used for this user.}
}
\description{
Changes user password(s) and/or access string.

See \url{https://paws-r.github.io/docs/memorydb/update_user.html} for full documentation.
}
\keyword{internal}
