% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_operations.R
\name{neptune_add_role_to_db_cluster}
\alias{neptune_add_role_to_db_cluster}
\title{Associates an Identity and Access Management (IAM) role with an Neptune
DB cluster}
\usage{
neptune_add_role_to_db_cluster(
  DBClusterIdentifier,
  RoleArn,
  FeatureName = NULL
)
}
\arguments{
\item{DBClusterIdentifier}{[required] The name of the DB cluster to associate the IAM role with.}

\item{RoleArn}{[required] The Amazon Resource Name (ARN) of the IAM role to associate with the
Neptune DB cluster, for example
\verb{arn:aws:iam::123456789012:role/NeptuneAccessRole}.}

\item{FeatureName}{The name of the feature for the Neptune DB cluster that the IAM role is
to be associated with. For the list of supported feature names, see
DBEngineVersion.}
}
\description{
Associates an Identity and Access Management (IAM) role with an Neptune DB cluster.

See \url{https://www.paws-r-sdk.com/docs/neptune_add_role_to_db_cluster/} for full documentation.
}
\keyword{internal}
