% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_operations.R
\name{neptune_copy_db_cluster_parameter_group}
\alias{neptune_copy_db_cluster_parameter_group}
\title{Copies the specified DB cluster parameter group}
\usage{
neptune_copy_db_cluster_parameter_group(
  SourceDBClusterParameterGroupIdentifier,
  TargetDBClusterParameterGroupIdentifier,
  TargetDBClusterParameterGroupDescription,
  Tags = NULL
)
}
\arguments{
\item{SourceDBClusterParameterGroupIdentifier}{[required] The identifier or Amazon Resource Name (ARN) for the source DB cluster parameter group. For information about creating an ARN, see \href{https://docs.aws.amazon.com/neptune/latest/userguide/#tagging.ARN.Constructing}{Constructing an Amazon Resource Name (ARN)}.

Constraints:
\itemize{
\item Must specify a valid DB cluster parameter group.
\item Must specify a valid DB cluster parameter group identifier, for example \code{my-db-cluster-param-group}, or a valid ARN.
\item The source DB cluster parameter group must be in the same Amazon Region as the copy. Neptune does not support cross-Region copying of parameter groups.
}}

\item{TargetDBClusterParameterGroupIdentifier}{[required] The identifier for the copied DB cluster parameter group.

Constraints:
\itemize{
\item Cannot be null, empty, or blank
\item Must contain from 1 to 255 letters, numbers, or hyphens
\item First character must be a letter
\item Cannot end with a hyphen or contain two consecutive hyphens
}

Example: \code{my-cluster-param-group1}}

\item{TargetDBClusterParameterGroupDescription}{[required] A description for the copied DB cluster parameter group.}

\item{Tags}{The tags to be assigned to the copied DB cluster parameter group.}
}
\description{
Copies the specified DB cluster parameter group.

See \url{https://www.paws-r-sdk.com/docs/neptune_copy_db_cluster_parameter_group/} for full documentation.
}
\keyword{internal}
