% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_operations.R
\name{neptune_delete_db_cluster_parameter_group}
\alias{neptune_delete_db_cluster_parameter_group}
\title{Deletes a specified DB cluster parameter group}
\usage{
neptune_delete_db_cluster_parameter_group(DBClusterParameterGroupName)
}
\arguments{
\item{DBClusterParameterGroupName}{[required] The name of the DB cluster parameter group.

Constraints:
\itemize{
\item Must be the name of an existing DB cluster parameter group.
\item You can't delete a default DB cluster parameter group.
\item Cannot be associated with any DB clusters.
}}
}
\description{
Deletes a specified DB cluster parameter group. The DB cluster parameter group to be deleted can't be associated with any DB clusters.

See \url{https://paws-r.github.io/docs/neptune/delete_db_cluster_parameter_group.html} for full documentation.
}
\keyword{internal}
