% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_operations.R
\name{neptune_delete_db_parameter_group}
\alias{neptune_delete_db_parameter_group}
\title{Deletes a specified DBParameterGroup}
\usage{
neptune_delete_db_parameter_group(DBParameterGroupName)
}
\arguments{
\item{DBParameterGroupName}{[required] The name of the DB parameter group.

Constraints:
\itemize{
\item Must be the name of an existing DB parameter group
\item You can't delete a default DB parameter group
\item Cannot be associated with any DB instances
}}
}
\value{
An empty list.
}
\description{
Deletes a specified DBParameterGroup. The DBParameterGroup to be deleted
can't be associated with any DB instances.
}
\section{Request syntax}{
\preformatted{svc$delete_db_parameter_group(
  DBParameterGroupName = "string"
)
}
}

\keyword{internal}
