% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_operations.R
\name{neptune_delete_event_subscription}
\alias{neptune_delete_event_subscription}
\title{Deletes an event notification subscription}
\usage{
neptune_delete_event_subscription(SubscriptionName)
}
\arguments{
\item{SubscriptionName}{[required] The name of the event notification subscription you want to delete.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  EventSubscription = list(
    CustomerAwsId = "string",
    CustSubscriptionId = "string",
    SnsTopicArn = "string",
    Status = "string",
    SubscriptionCreationTime = "string",
    SourceType = "string",
    SourceIdsList = list(
      "string"
    ),
    EventCategoriesList = list(
      "string"
    ),
    Enabled = TRUE|FALSE,
    EventSubscriptionArn = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deletes an event notification subscription.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_event_subscription(
  SubscriptionName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
