% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_operations.R
\name{neptune_delete_global_cluster}
\alias{neptune_delete_global_cluster}
\title{Deletes a global database}
\usage{
neptune_delete_global_cluster(GlobalClusterIdentifier)
}
\arguments{
\item{GlobalClusterIdentifier}{[required] The cluster identifier of the global database cluster being deleted.}
}
\description{
Deletes a global database. The primary and all secondary clusters must already be detached or deleted first.

See \url{https://www.paws-r-sdk.com/docs/neptune_delete_global_cluster/} for full documentation.
}
\keyword{internal}
