% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_operations.R
\name{neptune_describe_db_subnet_groups}
\alias{neptune_describe_db_subnet_groups}
\title{Returns a list of DBSubnetGroup descriptions}
\usage{
neptune_describe_db_subnet_groups(
  DBSubnetGroupName = NULL,
  Filters = NULL,
  MaxRecords = NULL,
  Marker = NULL
)
}
\arguments{
\item{DBSubnetGroupName}{The name of the DB subnet group to return details for.}

\item{Filters}{This parameter is not currently supported.}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
called a marker is included in the response so that the remaining
results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}

\item{Marker}{An optional pagination token provided by a previous
DescribeDBSubnetGroups request. If this parameter is specified, the
response includes only records beyond the marker, up to the value
specified by \code{MaxRecords}.}
}
\description{
Returns a list of DBSubnetGroup descriptions. If a DBSubnetGroupName is specified, the list will contain only the descriptions of the specified DBSubnetGroup.

See \url{https://www.paws-r-sdk.com/docs/neptune_describe_db_subnet_groups/} for full documentation.
}
\keyword{internal}
