% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_operations.R
\name{neptune_describe_orderable_db_instance_options}
\alias{neptune_describe_orderable_db_instance_options}
\title{Returns a list of orderable DB instance options for the specified engine}
\usage{
neptune_describe_orderable_db_instance_options(Engine, EngineVersion,
  DBInstanceClass, LicenseModel, Vpc, Filters, MaxRecords, Marker)
}
\arguments{
\item{Engine}{[required] The name of the engine to retrieve DB instance options for.}

\item{EngineVersion}{The engine version filter value. Specify this parameter to show only the
available offerings matching the specified engine version.}

\item{DBInstanceClass}{The DB instance class filter value. Specify this parameter to show only
the available offerings matching the specified DB instance class.}

\item{LicenseModel}{The license model filter value. Specify this parameter to show only the
available offerings matching the specified license model.}

\item{Vpc}{The VPC filter value. Specify this parameter to show only the available
VPC or non-VPC offerings.}

\item{Filters}{This parameter is not currently supported.}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
called a marker is included in the response so that the remaining
results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}

\item{Marker}{An optional pagination token provided by a previous
DescribeOrderableDBInstanceOptions request. If this parameter is
specified, the response includes only records beyond the marker, up to
the value specified by \code{MaxRecords} .}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  OrderableDBInstanceOptions = list(
    list(
      Engine = "string",
      EngineVersion = "string",
      DBInstanceClass = "string",
      LicenseModel = "string",
      AvailabilityZones = list(
        list(
          Name = "string"
        )
      ),
      MultiAZCapable = TRUE|FALSE,
      ReadReplicaCapable = TRUE|FALSE,
      Vpc = TRUE|FALSE,
      SupportsStorageEncryption = TRUE|FALSE,
      StorageType = "string",
      SupportsIops = TRUE|FALSE,
      SupportsEnhancedMonitoring = TRUE|FALSE,
      SupportsIAMDatabaseAuthentication = TRUE|FALSE,
      SupportsPerformanceInsights = TRUE|FALSE,
      MinStorageSize = 123,
      MaxStorageSize = 123,
      MinIopsPerDbInstance = 123,
      MaxIopsPerDbInstance = 123,
      MinIopsPerGib = 123.0,
      MaxIopsPerGib = 123.0
    )
  ),
  Marker = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns a list of orderable DB instance options for the specified
engine.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_orderable_db_instance_options(
  Engine = "string",
  EngineVersion = "string",
  DBInstanceClass = "string",
  LicenseModel = "string",
  Vpc = TRUE|FALSE,
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxRecords = 123,
  Marker = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
