% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_operations.R
\name{neptune_list_tags_for_resource}
\alias{neptune_list_tags_for_resource}
\title{Lists all tags on an Amazon Neptune resource}
\usage{
neptune_list_tags_for_resource(ResourceName, Filters)
}
\arguments{
\item{ResourceName}{[required] The Amazon Neptune resource with tags to be listed. This value is an
Amazon Resource Name (ARN). For information about creating an ARN, see
\href{https://docs.aws.amazon.com/neptune/latest/userguide/#tagging.ARN.Constructing}{Constructing an Amazon Resource Name (ARN)}.}

\item{Filters}{This parameter is not currently supported.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  TagList = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Lists all tags on an Amazon Neptune resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_tags_for_resource(
  ResourceName = "string",
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
