% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_operations.R
\name{neptune_promote_read_replica_db_cluster}
\alias{neptune_promote_read_replica_db_cluster}
\title{Not supported}
\usage{
neptune_promote_read_replica_db_cluster(DBClusterIdentifier)
}
\arguments{
\item{DBClusterIdentifier}{[required] Not supported.}
}
\description{
Not supported.

See \url{https://www.paws-r-sdk.com/docs/neptune_promote_read_replica_db_cluster/} for full documentation.
}
\keyword{internal}
