% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_operations.R
\name{neptune_remove_from_global_cluster}
\alias{neptune_remove_from_global_cluster}
\title{Detaches a Neptune DB cluster from a Neptune global database}
\usage{
neptune_remove_from_global_cluster(
  GlobalClusterIdentifier,
  DbClusterIdentifier
)
}
\arguments{
\item{GlobalClusterIdentifier}{[required] The identifier of the Neptune global database from which to detach the
specified Neptune DB cluster.}

\item{DbClusterIdentifier}{[required] The Amazon Resource Name (ARN) identifying the cluster to be detached
from the Neptune global database cluster.}
}
\description{
Detaches a Neptune DB cluster from a Neptune global database. A secondary cluster becomes a normal standalone cluster with read-write capability instead of being read-only, and no longer receives data from a the primary cluster.

See \url{https://www.paws-r-sdk.com/docs/neptune_remove_from_global_cluster/} for full documentation.
}
\keyword{internal}
