% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_operations.R
\name{neptune_remove_tags_from_resource}
\alias{neptune_remove_tags_from_resource}
\title{Removes metadata tags from an Amazon Neptune resource}
\usage{
neptune_remove_tags_from_resource(ResourceName, TagKeys)
}
\arguments{
\item{ResourceName}{[required] The Amazon Neptune resource that the tags are removed from. This value
is an Amazon Resource Name (ARN). For information about creating an ARN,
see \href{https://docs.aws.amazon.com/neptune/latest/userguide/#tagging.ARN.Constructing}{Constructing an Amazon Resource Name (ARN)}.}

\item{TagKeys}{[required] The tag key (name) of the tag to be removed.}
}
\description{
Removes metadata tags from an Amazon Neptune resource.

See \url{https://paws-r.github.io/docs/neptune/remove_tags_from_resource.html} for full documentation.
}
\keyword{internal}
