% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_operations.R
\name{neptune_stop_db_cluster}
\alias{neptune_stop_db_cluster}
\title{Stops an Amazon Neptune DB cluster}
\usage{
neptune_stop_db_cluster(DBClusterIdentifier)
}
\arguments{
\item{DBClusterIdentifier}{[required] The DB cluster identifier of the Neptune DB cluster to be stopped. This
parameter is stored as a lowercase string.}
}
\description{
Stops an Amazon Neptune DB cluster. When you stop a DB cluster, Neptune retains the DB cluster's metadata, including its endpoints and DB parameter groups.

See \url{https://paws-r.github.io/docs/neptune/stop_db_cluster.html} for full documentation.
}
\keyword{internal}
