% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_operations.R
\name{neptune_switchover_global_cluster}
\alias{neptune_switchover_global_cluster}
\title{Switches over the specified secondary DB cluster to be the new primary
DB cluster in the global database cluster}
\usage{
neptune_switchover_global_cluster(
  GlobalClusterIdentifier,
  TargetDbClusterIdentifier
)
}
\arguments{
\item{GlobalClusterIdentifier}{[required] The identifier of the global database cluster to switch over. This
parameter isn't case-sensitive.

Constraints: Must match the identifier of an existing global database
cluster.}

\item{TargetDbClusterIdentifier}{[required] The Amazon Resource Name (ARN) of the secondary Neptune DB cluster that
you want to promote to primary for the global database.}
}
\description{
Switches over the specified secondary DB cluster to be the new primary DB cluster in the global database cluster. Switchover operations were previously called "managed planned failovers."

See \url{https://www.paws-r-sdk.com/docs/neptune_switchover_global_cluster/} for full documentation.
}
\keyword{internal}
