% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptunedata_operations.R
\name{neptunedata_cancel_gremlin_query}
\alias{neptunedata_cancel_gremlin_query}
\title{Cancels a Gremlin query}
\usage{
neptunedata_cancel_gremlin_query(queryId)
}
\arguments{
\item{queryId}{[required] The unique identifier that identifies the query to be canceled.}
}
\description{
Cancels a Gremlin query. See \href{https://docs.aws.amazon.com/neptune/latest/userguide/gremlin-api-status-cancel.html}{Gremlin query cancellation} for more information.

See \url{https://www.paws-r-sdk.com/docs/neptunedata_cancel_gremlin_query/} for full documentation.
}
\keyword{internal}
