% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptunedata_operations.R
\name{neptunedata_cancel_loader_job}
\alias{neptunedata_cancel_loader_job}
\title{Cancels a specified load job}
\usage{
neptunedata_cancel_loader_job(loadId)
}
\arguments{
\item{loadId}{[required] The ID of the load job to be deleted.}
}
\description{
Cancels a specified load job. This is an HTTP \code{DELETE} request. See \href{https://docs.aws.amazon.com/neptune/latest/userguide/}{Neptune Loader Get-Status API} for more information.

See \url{https://www.paws-r-sdk.com/docs/neptunedata_cancel_loader_job/} for full documentation.
}
\keyword{internal}
