% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptunedata_operations.R
\name{neptunedata_cancel_open_cypher_query}
\alias{neptunedata_cancel_open_cypher_query}
\title{Cancels a specified openCypher query}
\usage{
neptunedata_cancel_open_cypher_query(queryId, silent = NULL)
}
\arguments{
\item{queryId}{[required] The unique ID of the openCypher query to cancel.}

\item{silent}{If set to \code{TRUE}, causes the cancelation of the openCypher query to
happen silently.}
}
\description{
Cancels a specified openCypher query. See \href{https://docs.aws.amazon.com/neptune/latest/userguide/access-graph-opencypher-status.html}{Neptune openCypher status endpoint} for more information.

See \url{https://www.paws-r-sdk.com/docs/neptunedata_cancel_open_cypher_query/} for full documentation.
}
\keyword{internal}
