% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptunedata_operations.R
\name{neptunedata_delete_propertygraph_statistics}
\alias{neptunedata_delete_propertygraph_statistics}
\title{Deletes statistics for Gremlin and openCypher (property graph) data}
\usage{
neptunedata_delete_propertygraph_statistics()
}
\description{
Deletes statistics for Gremlin and openCypher (property graph) data.

See \url{https://www.paws-r-sdk.com/docs/neptunedata_delete_propertygraph_statistics/} for full documentation.
}
\keyword{internal}
