% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptunedata_operations.R
\name{neptunedata_delete_sparql_statistics}
\alias{neptunedata_delete_sparql_statistics}
\title{Deletes SPARQL statistics}
\usage{
neptunedata_delete_sparql_statistics()
}
\description{
Deletes SPARQL statistics

See \url{https://www.paws-r-sdk.com/docs/neptunedata_delete_sparql_statistics/} for full documentation.
}
\keyword{internal}
