% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptunedata_operations.R
\name{neptunedata_execute_fast_reset}
\alias{neptunedata_execute_fast_reset}
\title{The fast reset REST API lets you reset a Neptune graph quicky and
easily, removing all of its data}
\usage{
neptunedata_execute_fast_reset(action, token = NULL)
}
\arguments{
\item{action}{[required] The fast reset action. One of the following values:
\itemize{
\item \strong{\code{initiateDatabaseReset}}   –   This action generates a unique
token needed to actually perform the fast reset.
\item \strong{\code{performDatabaseReset}}   –   This action uses the token
generated by the \code{initiateDatabaseReset} action to actually perform
the fast reset.
}}

\item{token}{The fast-reset token to initiate the reset.}
}
\description{
The fast reset REST API lets you reset a Neptune graph quicky and easily, removing all of its data.

See \url{https://www.paws-r-sdk.com/docs/neptunedata_execute_fast_reset/} for full documentation.
}
\keyword{internal}
