% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptunedata_operations.R
\name{neptunedata_execute_gremlin_explain_query}
\alias{neptunedata_execute_gremlin_explain_query}
\title{Executes a Gremlin Explain query}
\usage{
neptunedata_execute_gremlin_explain_query(gremlinQuery)
}
\arguments{
\item{gremlinQuery}{[required] The Gremlin explain query string.}
}
\description{
Executes a Gremlin Explain query.

See \url{https://www.paws-r-sdk.com/docs/neptunedata_execute_gremlin_explain_query/} for full documentation.
}
\keyword{internal}
