% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptunedata_operations.R
\name{neptunedata_execute_open_cypher_explain_query}
\alias{neptunedata_execute_open_cypher_explain_query}
\title{Executes an openCypher explain request}
\usage{
neptunedata_execute_open_cypher_explain_query(
  openCypherQuery,
  parameters = NULL,
  explainMode
)
}
\arguments{
\item{openCypherQuery}{[required] The openCypher query string.}

\item{parameters}{The openCypher query parameters.}

\item{explainMode}{[required] The openCypher \code{explain} mode. Can be one of: \code{static}, \code{dynamic}, or
\code{details}.}
}
\description{
Executes an openCypher \code{explain} request. See \href{https://docs.aws.amazon.com/neptune/latest/userguide/access-graph-opencypher-explain.html}{The openCypher explain feature} for more information.

See \url{https://www.paws-r-sdk.com/docs/neptunedata_execute_open_cypher_explain_query/} for full documentation.
}
\keyword{internal}
