% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptunedata_operations.R
\name{neptunedata_get_engine_status}
\alias{neptunedata_get_engine_status}
\title{Retrieves the status of the graph database on the host}
\usage{
neptunedata_get_engine_status()
}
\description{
Retrieves the status of the graph database on the host.

See \url{https://www.paws-r-sdk.com/docs/neptunedata_get_engine_status/} for full documentation.
}
\keyword{internal}
