% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptunedata_operations.R
\name{neptunedata_get_gremlin_query_status}
\alias{neptunedata_get_gremlin_query_status}
\title{Gets the status of a specified Gremlin query}
\usage{
neptunedata_get_gremlin_query_status(queryId)
}
\arguments{
\item{queryId}{[required] The unique identifier that identifies the Gremlin query.}
}
\description{
Gets the status of a specified Gremlin query.

See \url{https://www.paws-r-sdk.com/docs/neptunedata_get_gremlin_query_status/} for full documentation.
}
\keyword{internal}
