% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptunedata_operations.R
\name{neptunedata_get_ml_endpoint}
\alias{neptunedata_get_ml_endpoint}
\title{Retrieves details about an inference endpoint}
\usage{
neptunedata_get_ml_endpoint(id, neptuneIamRoleArn = NULL)
}
\arguments{
\item{id}{[required] The unique identifier of the inference endpoint.}

\item{neptuneIamRoleArn}{The ARN of an IAM role that provides Neptune access to SageMaker and
Amazon S3 resources. This must be listed in your DB cluster parameter
group or an error will occur.}
}
\description{
Retrieves details about an inference endpoint. See \href{https://docs.aws.amazon.com/neptune/latest/userguide/machine-learning-api-endpoints.html}{Managing inference endpoints using the endpoints command}.

See \url{https://www.paws-r-sdk.com/docs/neptunedata_get_ml_endpoint/} for full documentation.
}
\keyword{internal}
