% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptunedata_operations.R
\name{neptunedata_get_open_cypher_query_status}
\alias{neptunedata_get_open_cypher_query_status}
\title{Retrieves the status of a specified openCypher query}
\usage{
neptunedata_get_open_cypher_query_status(queryId)
}
\arguments{
\item{queryId}{[required] The unique ID of the openCypher query for which to retrieve the query
status.}
}
\description{
Retrieves the status of a specified openCypher query.

See \url{https://www.paws-r-sdk.com/docs/neptunedata_get_open_cypher_query_status/} for full documentation.
}
\keyword{internal}
