% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptunedata_operations.R
\name{neptunedata_get_propertygraph_statistics}
\alias{neptunedata_get_propertygraph_statistics}
\title{Gets property graph statistics (Gremlin and openCypher)}
\usage{
neptunedata_get_propertygraph_statistics()
}
\description{
Gets property graph statistics (Gremlin and openCypher).

See \url{https://www.paws-r-sdk.com/docs/neptunedata_get_propertygraph_statistics/} for full documentation.
}
\keyword{internal}
