% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptunedata_operations.R
\name{neptunedata_get_propertygraph_summary}
\alias{neptunedata_get_propertygraph_summary}
\title{Gets a graph summary for a property graph}
\usage{
neptunedata_get_propertygraph_summary(mode = NULL)
}
\arguments{
\item{mode}{Mode can take one of two values: \code{BASIC} (the default), and \code{DETAILED}.}
}
\description{
Gets a graph summary for a property graph.

See \url{https://www.paws-r-sdk.com/docs/neptunedata_get_propertygraph_summary/} for full documentation.
}
\keyword{internal}
