% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptunedata_operations.R
\name{neptunedata_get_sparql_statistics}
\alias{neptunedata_get_sparql_statistics}
\title{Gets RDF statistics (SPARQL)}
\usage{
neptunedata_get_sparql_statistics()
}
\description{
Gets RDF statistics (SPARQL).

See \url{https://www.paws-r-sdk.com/docs/neptunedata_get_sparql_statistics/} for full documentation.
}
\keyword{internal}
