% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptunedata_operations.R
\name{neptunedata_list_gremlin_queries}
\alias{neptunedata_list_gremlin_queries}
\title{Lists active Gremlin queries}
\usage{
neptunedata_list_gremlin_queries(includeWaiting = NULL)
}
\arguments{
\item{includeWaiting}{If set to \code{TRUE}, the list returned includes waiting queries. The
default is \code{FALSE};}
}
\description{
Lists active Gremlin queries. See \href{https://docs.aws.amazon.com/neptune/latest/userguide/gremlin-api-status.html}{Gremlin query status API} for details about the output.

See \url{https://www.paws-r-sdk.com/docs/neptunedata_list_gremlin_queries/} for full documentation.
}
\keyword{internal}
