% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptunedata_operations.R
\name{neptunedata_list_ml_data_processing_jobs}
\alias{neptunedata_list_ml_data_processing_jobs}
\title{Returns a list of Neptune ML data processing jobs}
\usage{
neptunedata_list_ml_data_processing_jobs(
  maxItems = NULL,
  neptuneIamRoleArn = NULL
)
}
\arguments{
\item{maxItems}{The maximum number of items to return (from 1 to 1024; the default is
10).}

\item{neptuneIamRoleArn}{The ARN of an IAM role that provides Neptune access to SageMaker and
Amazon S3 resources. This must be listed in your DB cluster parameter
group or an error will occur.}
}
\description{
Returns a list of Neptune ML data processing jobs. See \href{https://docs.aws.amazon.com/neptune/latest/userguide/machine-learning-api-dataprocessing.html#machine-learning-api-dataprocessing-list-jobs}{Listing active data-processing jobs using the Neptune ML dataprocessing command}.

See \url{https://www.paws-r-sdk.com/docs/neptunedata_list_ml_data_processing_jobs/} for full documentation.
}
\keyword{internal}
