% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptunedata_operations.R
\name{neptunedata_list_ml_endpoints}
\alias{neptunedata_list_ml_endpoints}
\title{Lists existing inference endpoints}
\usage{
neptunedata_list_ml_endpoints(maxItems = NULL, neptuneIamRoleArn = NULL)
}
\arguments{
\item{maxItems}{The maximum number of items to return (from 1 to 1024; the default is
10.}

\item{neptuneIamRoleArn}{The ARN of an IAM role that provides Neptune access to SageMaker and
Amazon S3 resources. This must be listed in your DB cluster parameter
group or an error will occur.}
}
\description{
Lists existing inference endpoints. See \href{https://docs.aws.amazon.com/neptune/latest/userguide/machine-learning-api-endpoints.html}{Managing inference endpoints using the endpoints command}.

See \url{https://www.paws-r-sdk.com/docs/neptunedata_list_ml_endpoints/} for full documentation.
}
\keyword{internal}
