% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptunedata_operations.R
\name{neptunedata_list_ml_model_training_jobs}
\alias{neptunedata_list_ml_model_training_jobs}
\title{Lists Neptune ML model-training jobs}
\usage{
neptunedata_list_ml_model_training_jobs(
  maxItems = NULL,
  neptuneIamRoleArn = NULL
)
}
\arguments{
\item{maxItems}{The maximum number of items to return (from 1 to 1024; the default is
10).}

\item{neptuneIamRoleArn}{The ARN of an IAM role that provides Neptune access to SageMaker and
Amazon S3 resources. This must be listed in your DB cluster parameter
group or an error will occur.}
}
\description{
Lists Neptune ML model-training jobs. See \href{https://docs.aws.amazon.com/neptune/latest/userguide/machine-learning-api-modeltraining.html}{Model training using the \code{modeltraining} command}.

See \url{https://www.paws-r-sdk.com/docs/neptunedata_list_ml_model_training_jobs/} for full documentation.
}
\keyword{internal}
