% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptunedata_operations.R
\name{neptunedata_list_open_cypher_queries}
\alias{neptunedata_list_open_cypher_queries}
\title{Lists active openCypher queries}
\usage{
neptunedata_list_open_cypher_queries(includeWaiting = NULL)
}
\arguments{
\item{includeWaiting}{When set to \code{TRUE} and other parameters are not present, causes status
information to be returned for waiting queries as well as for running
queries.}
}
\description{
Lists active openCypher queries. See \href{https://docs.aws.amazon.com/neptune/latest/userguide/access-graph-opencypher-status.html}{Neptune openCypher status endpoint} for more information.

See \url{https://www.paws-r-sdk.com/docs/neptunedata_list_open_cypher_queries/} for full documentation.
}
\keyword{internal}
