% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptunedata_operations.R
\name{neptunedata_manage_sparql_statistics}
\alias{neptunedata_manage_sparql_statistics}
\title{Manages the generation and use of RDF graph statistics}
\usage{
neptunedata_manage_sparql_statistics(mode = NULL)
}
\arguments{
\item{mode}{The statistics generation mode. One of: \code{DISABLE_AUTOCOMPUTE},
\code{ENABLE_AUTOCOMPUTE}, or \code{REFRESH}, the last of which manually triggers
DFE statistics generation.}
}
\description{
Manages the generation and use of RDF graph statistics.

See \url{https://www.paws-r-sdk.com/docs/neptunedata_manage_sparql_statistics/} for full documentation.
}
\keyword{internal}
