% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qldb_operations.R
\name{qldb_cancel_journal_kinesis_stream}
\alias{qldb_cancel_journal_kinesis_stream}
\title{Ends a given Amazon QLDB journal stream}
\usage{
qldb_cancel_journal_kinesis_stream(LedgerName, StreamId)
}
\arguments{
\item{LedgerName}{[required] The name of the ledger.}

\item{StreamId}{[required] The UUID (represented in Base62-encoded text) of the QLDB journal stream
to be canceled.}
}
\description{
Ends a given Amazon QLDB journal stream. Before a stream can be canceled, its current status must be \code{ACTIVE}.

See \url{https://paws-r.github.io/docs/qldb/cancel_journal_kinesis_stream.html} for full documentation.
}
\keyword{internal}
