% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qldb_operations.R
\name{qldb_delete_ledger}
\alias{qldb_delete_ledger}
\title{Deletes a ledger and all of its contents}
\usage{
qldb_delete_ledger(Name)
}
\arguments{
\item{Name}{[required] The name of the ledger that you want to delete.}
}
\description{
Deletes a ledger and all of its contents. This action is irreversible.

See \url{https://www.paws-r-sdk.com/docs/qldb_delete_ledger/} for full documentation.
}
\keyword{internal}
