% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qldb_operations.R
\name{qldb_describe_journal_s3_export}
\alias{qldb_describe_journal_s3_export}
\title{Returns information about a journal export job, including the ledger
name, export ID, creation time, current status, and the parameters of
the original export creation request}
\usage{
qldb_describe_journal_s3_export(Name, ExportId)
}
\arguments{
\item{Name}{[required] The name of the ledger.}

\item{ExportId}{[required] The UUID (represented in Base62-encoded text) of the journal export job
to describe.}
}
\description{
Returns information about a journal export job, including the ledger name, export ID, creation time, current status, and the parameters of the original export creation request.

See \url{https://paws-r.github.io/docs/qldb/describe_journal_s3_export.html} for full documentation.
}
\keyword{internal}
