% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qldb_operations.R
\name{qldb_get_digest}
\alias{qldb_get_digest}
\title{Returns the digest of a ledger at the latest committed block in the
journal}
\usage{
qldb_get_digest(Name)
}
\arguments{
\item{Name}{[required] The name of the ledger.}
}
\description{
Returns the digest of a ledger at the latest committed block in the journal. The response includes a 256-bit hash value and a block address.

See \url{https://www.paws-r-sdk.com/docs/qldb_get_digest/} for full documentation.
}
\keyword{internal}
