% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qldb_operations.R
\name{qldb_list_journal_s3_exports_for_ledger}
\alias{qldb_list_journal_s3_exports_for_ledger}
\title{Returns all journal export jobs for a specified ledger}
\usage{
qldb_list_journal_s3_exports_for_ledger(
  Name,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Name}{[required] The name of the ledger.}

\item{MaxResults}{The maximum number of results to return in a single
\code{\link[=qldb_list_journal_s3_exports_for_ledger]{list_journal_s3_exports_for_ledger}}
request. (The actual number of results returned might be fewer.)}

\item{NextToken}{A pagination token, indicating that you want to retrieve the next page
of results. If you received a value for \code{NextToken} in the response from
a previous
\code{\link[=qldb_list_journal_s3_exports_for_ledger]{list_journal_s3_exports_for_ledger}}
call, then you should use that value as input here.}
}
\description{
Returns all journal export jobs for a specified ledger.

See \url{https://paws-r.github.io/docs/qldb/list_journal_s3_exports_for_ledger.html} for full documentation.
}
\keyword{internal}
