% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qldb_operations.R
\name{qldb_list_ledgers}
\alias{qldb_list_ledgers}
\title{Returns all ledgers that are associated with the current Amazon Web
Services account and Region}
\usage{
qldb_list_ledgers(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of results to return in a single
\code{\link[=qldb_list_ledgers]{list_ledgers}} request. (The actual number of
results returned might be fewer.)}

\item{NextToken}{A pagination token, indicating that you want to retrieve the next page
of results. If you received a value for \code{NextToken} in the response from
a previous \code{\link[=qldb_list_ledgers]{list_ledgers}} call, then you should use
that value as input here.}
}
\description{
Returns all ledgers that are associated with the current Amazon Web Services account and Region.

See \url{https://paws-r.github.io/docs/qldb/list_ledgers.html} for full documentation.
}
\keyword{internal}
