% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qldb_operations.R
\name{qldb_untag_resource}
\alias{qldb_untag_resource}
\title{Removes one or more tags from a specified Amazon QLDB resource}
\usage{
qldb_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) from which to remove the tags. For
example:

\code{arn:aws:qldb:us-east-1:123456789012:ledger/exampleLedger}}

\item{TagKeys}{[required] The list of tag keys to remove.}
}
\description{
Removes one or more tags from a specified Amazon QLDB resource. You can specify up to 50 tag keys to remove.

See \url{https://paws-r.github.io/docs/qldb/untag_resource.html} for full documentation.
}
\keyword{internal}
