% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_delete_blue_green_deployment}
\alias{rds_delete_blue_green_deployment}
\title{Deletes a blue/green deployment}
\usage{
rds_delete_blue_green_deployment(
  BlueGreenDeploymentIdentifier,
  DeleteTarget = NULL
)
}
\arguments{
\item{BlueGreenDeploymentIdentifier}{[required] The unique identifier of the blue/green deployment to delete. This
parameter isn't case-sensitive.

Constraints:
\itemize{
\item Must match an existing blue/green deployment identifier.
}}

\item{DeleteTarget}{Specifies whether to delete the resources in the green environment. You
can't specify this option if the blue/green deployment
\href{https://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_BlueGreenDeployment.html}{status}
is \code{SWITCHOVER_COMPLETED}.}
}
\description{
Deletes a blue/green deployment.

See \url{https://www.paws-r-sdk.com/docs/rds_delete_blue_green_deployment/} for full documentation.
}
\keyword{internal}
