% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_delete_custom_availability_zone}
\alias{rds_delete_custom_availability_zone}
\title{Deletes a custom Availability Zone (AZ)}
\usage{
rds_delete_custom_availability_zone(CustomAvailabilityZoneId)
}
\arguments{
\item{CustomAvailabilityZoneId}{[required] The custom AZ identifier.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  CustomAvailabilityZone = list(
    CustomAvailabilityZoneId = "string",
    CustomAvailabilityZoneName = "string",
    CustomAvailabilityZoneStatus = "string",
    VpnDetails = list(
      VpnId = "string",
      VpnTunnelOriginatorIP = "string",
      VpnGatewayIp = "string",
      VpnPSK = "string",
      VpnName = "string",
      VpnState = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deletes a custom Availability Zone (AZ).

A custom AZ is an on-premises AZ that is integrated with a VMware
vSphere cluster.

For more information about RDS on VMware, see the \href{https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/Welcome.html}{RDS on VMware User Guide.}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_custom_availability_zone(
  CustomAvailabilityZoneId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
