% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_delete_custom_db_engine_version}
\alias{rds_delete_custom_db_engine_version}
\title{Deletes a custom engine version}
\usage{
rds_delete_custom_db_engine_version(Engine, EngineVersion)
}
\arguments{
\item{Engine}{[required] The database engine.

RDS Custom for Oracle supports the following values:
\itemize{
\item \code{custom-oracle-ee}
\item \code{custom-oracle-ee-cdb}
\item \code{custom-oracle-se2}
\item \code{custom-oracle-se2-cdb}
}

RDS Custom for SQL Server supports the following values:
\itemize{
\item \code{custom-sqlserver-ee}
\item \code{custom-sqlserver-se}
\item \code{custom-sqlserver-web}
\item \code{custom-sqlserver-dev}
}

RDS for SQL Server supports the following values:
\itemize{
\item \code{sqlserver-ee} (Bring Your Own Media)
\item \code{sqlserver-se} (Bring Your Own Media)
\item \code{sqlserver-dev-ee}
}}

\item{EngineVersion}{[required] The custom engine version (CEV) for your DB instance. This option is required for RDS Custom, but optional for Amazon RDS. The combination of \code{Engine} and \code{EngineVersion} is unique per customer per Amazon Web Services Region.}
}
\description{
Deletes a custom engine version. To run this command, make sure you meet the following prerequisites:

See \url{https://www.paws-r-sdk.com/docs/rds_delete_custom_db_engine_version/} for full documentation.
}
\keyword{internal}
