% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_delete_db_cluster_automated_backup}
\alias{rds_delete_db_cluster_automated_backup}
\title{Deletes automated backups using the DbClusterResourceId value of the
source DB cluster or the Amazon Resource Name (ARN) of the automated
backups}
\usage{
rds_delete_db_cluster_automated_backup(DbClusterResourceId)
}
\arguments{
\item{DbClusterResourceId}{[required] The identifier for the source DB cluster, which can't be changed and
which is unique to an Amazon Web Services Region.}
}
\description{
Deletes automated backups using the \code{DbClusterResourceId} value of the source DB cluster or the Amazon Resource Name (ARN) of the automated backups.

See \url{https://www.paws-r-sdk.com/docs/rds_delete_db_cluster_automated_backup/} for full documentation.
}
\keyword{internal}
