% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_delete_db_cluster_endpoint}
\alias{rds_delete_db_cluster_endpoint}
\title{Deletes a custom endpoint and removes it from an Amazon Aurora DB
cluster}
\usage{
rds_delete_db_cluster_endpoint(DBClusterEndpointIdentifier)
}
\arguments{
\item{DBClusterEndpointIdentifier}{[required] The identifier associated with the custom endpoint. This parameter is
stored as a lowercase string.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DBClusterEndpointIdentifier = "string",
  DBClusterIdentifier = "string",
  DBClusterEndpointResourceIdentifier = "string",
  Endpoint = "string",
  Status = "string",
  EndpointType = "string",
  CustomEndpointType = "string",
  StaticMembers = list(
    "string"
  ),
  ExcludedMembers = list(
    "string"
  ),
  DBClusterEndpointArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Deletes a custom endpoint and removes it from an Amazon Aurora DB
cluster.

This action only applies to Aurora DB clusters.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_db_cluster_endpoint(
  DBClusterEndpointIdentifier = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
