% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_delete_db_cluster_snapshot}
\alias{rds_delete_db_cluster_snapshot}
\title{Deletes a DB cluster snapshot}
\usage{
rds_delete_db_cluster_snapshot(DBClusterSnapshotIdentifier)
}
\arguments{
\item{DBClusterSnapshotIdentifier}{[required] The identifier of the DB cluster snapshot to delete.

Constraints: Must be the name of an existing DB cluster snapshot in the
\code{available} state.}
}
\description{
Deletes a DB cluster snapshot. If the snapshot is being copied, the copy operation is terminated.

See \url{https://paws-r.github.io/docs/rds/delete_db_cluster_snapshot.html} for full documentation.
}
\keyword{internal}
