% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_delete_db_instance_automated_backup}
\alias{rds_delete_db_instance_automated_backup}
\title{Deletes automated backups based on the source instance's DbiResourceId
value or the restorable instance's resource ID}
\usage{
rds_delete_db_instance_automated_backup(DbiResourceId)
}
\arguments{
\item{DbiResourceId}{[required] The identifier for the source DB instance, which can't be changed and
which is unique to an AWS Region.}
}
\description{
Deletes automated backups based on the source instance's \code{DbiResourceId}
value or the restorable instance's resource ID.
}
\section{Request syntax}{
\preformatted{svc$delete_db_instance_automated_backup(
  DbiResourceId = "string"
)
}
}

\keyword{internal}
