% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_delete_db_proxy_endpoint}
\alias{rds_delete_db_proxy_endpoint}
\title{Deletes a DBProxyEndpoint}
\usage{
rds_delete_db_proxy_endpoint(DBProxyEndpointName)
}
\arguments{
\item{DBProxyEndpointName}{[required] The name of the DB proxy endpoint to delete.}
}
\value{
A list with the following syntax:\preformatted{list(
  DBProxyEndpoint = list(
    DBProxyEndpointName = "string",
    DBProxyEndpointArn = "string",
    DBProxyName = "string",
    Status = "available"|"modifying"|"incompatible-network"|"insufficient-resource-limits"|"creating"|"deleting",
    VpcId = "string",
    VpcSecurityGroupIds = list(
      "string"
    ),
    VpcSubnetIds = list(
      "string"
    ),
    Endpoint = "string",
    CreatedDate = as.POSIXct(
      "2015-01-01"
    ),
    TargetRole = "READ_WRITE"|"READ_ONLY",
    IsDefault = TRUE|FALSE
  )
)
}
}
\description{
Deletes a \code{DBProxyEndpoint}. Doing so removes the ability to access the
DB proxy using the endpoint that you defined. The endpoint that you
delete might have provided capabilities such as read/write or read-only
operations, or using a different VPC than the DB proxy's default VPC.
}
\section{Request syntax}{
\preformatted{svc$delete_db_proxy_endpoint(
  DBProxyEndpointName = "string"
)
}
}

\keyword{internal}
