% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_delete_db_shard_group}
\alias{rds_delete_db_shard_group}
\title{Deletes an Aurora Limitless Database DB shard group}
\usage{
rds_delete_db_shard_group(DBShardGroupIdentifier)
}
\arguments{
\item{DBShardGroupIdentifier}{[required] The name of the DB shard group to delete.}
}
\description{
Deletes an Aurora Limitless Database DB shard group.

See \url{https://www.paws-r-sdk.com/docs/rds_delete_db_shard_group/} for full documentation.
}
\keyword{internal}
