% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_delete_db_subnet_group}
\alias{rds_delete_db_subnet_group}
\title{Deletes a DB subnet group}
\usage{
rds_delete_db_subnet_group(DBSubnetGroupName)
}
\arguments{
\item{DBSubnetGroupName}{[required] The name of the database subnet group to delete.

You can't delete the default subnet group.

Constraints: Must match the name of an existing DBSubnetGroup. Must not
be default.

Example: \code{mydbsubnetgroup}}
}
\description{
Deletes a DB subnet group.

See \url{https://www.paws-r-sdk.com/docs/rds_delete_db_subnet_group/} for full documentation.
}
\keyword{internal}
