% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_delete_integration}
\alias{rds_delete_integration}
\title{Deletes a zero-ETL integration with Amazon Redshift}
\usage{
rds_delete_integration(IntegrationIdentifier)
}
\arguments{
\item{IntegrationIdentifier}{[required] The unique identifier of the integration.}
}
\description{
Deletes a zero-ETL integration with Amazon Redshift.

See \url{https://www.paws-r-sdk.com/docs/rds_delete_integration/} for full documentation.
}
\keyword{internal}
