% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_describe_account_attributes}
\alias{rds_describe_account_attributes}
\title{Lists all of the attributes for a customer account}
\usage{
rds_describe_account_attributes()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AccountQuotas = list(
    list(
      AccountQuotaName = "string",
      Used = 123,
      Max = 123
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Lists all of the attributes for a customer account. The attributes
include Amazon RDS quotas for the account, such as the number of DB
instances allowed. The description for a quota includes the quota name,
current usage toward that quota, and the quota's maximum value.

This command doesn't take any parameters.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_account_attributes()
}\if{html}{\out{</div>}}
}

\keyword{internal}
