% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_describe_db_cluster_snapshots}
\alias{rds_describe_db_cluster_snapshots}
\title{Returns information about DB cluster snapshots}
\usage{
rds_describe_db_cluster_snapshots(
  DBClusterIdentifier = NULL,
  DBClusterSnapshotIdentifier = NULL,
  SnapshotType = NULL,
  Filters = NULL,
  MaxRecords = NULL,
  Marker = NULL,
  IncludeShared = NULL,
  IncludePublic = NULL,
  DbClusterResourceId = NULL
)
}
\arguments{
\item{DBClusterIdentifier}{The ID of the DB cluster to retrieve the list of DB cluster snapshots
for. This parameter can't be used in conjunction with the
\code{DBClusterSnapshotIdentifier} parameter. This parameter isn't
case-sensitive.

Constraints:
\itemize{
\item If supplied, must match the identifier of an existing DBCluster.
}}

\item{DBClusterSnapshotIdentifier}{A specific DB cluster snapshot identifier to describe. This parameter
can't be used in conjunction with the \code{DBClusterIdentifier} parameter.
This value is stored as a lowercase string.

Constraints:
\itemize{
\item If supplied, must match the identifier of an existing
DBClusterSnapshot.
\item If this identifier is for an automated snapshot, the \code{SnapshotType}
parameter must also be specified.
}}

\item{SnapshotType}{The type of DB cluster snapshots to be returned. You can specify one of
the following values:
\itemize{
\item \code{automated} - Return all DB cluster snapshots that have been
automatically taken by Amazon RDS for my Amazon Web Services account.
\item \code{manual} - Return all DB cluster snapshots that have been taken by my
Amazon Web Services account.
\item \code{shared} - Return all manual DB cluster snapshots that have been
shared to my Amazon Web Services account.
\item \code{public} - Return all DB cluster snapshots that have been marked as
public.
}

If you don't specify a \code{SnapshotType} value, then both automated and
manual DB cluster snapshots are returned. You can include shared DB
cluster snapshots with these results by enabling the \code{IncludeShared}
parameter. You can include public DB cluster snapshots with these
results by enabling the \code{IncludePublic} parameter.

The \code{IncludeShared} and \code{IncludePublic} parameters don't apply for
\code{SnapshotType} values of \code{manual} or \code{automated}. The \code{IncludePublic}
parameter doesn't apply when \code{SnapshotType} is set to \code{shared}. The
\code{IncludeShared} parameter doesn't apply when \code{SnapshotType} is set to
\code{public}.}

\item{Filters}{A filter that specifies one or more DB cluster snapshots to describe.

Supported filters:
\itemize{
\item \code{db-cluster-id} - Accepts DB cluster identifiers and DB cluster Amazon
Resource Names (ARNs).
\item \code{db-cluster-snapshot-id} - Accepts DB cluster snapshot identifiers.
\item \code{snapshot-type} - Accepts types of DB cluster snapshots.
\item \code{engine} - Accepts names of database engines.
}}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
called a marker is included in the response so you can retrieve the
remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.}

\item{Marker}{An optional pagination token provided by a previous
\code{\link[=rds_describe_db_cluster_snapshots]{describe_db_cluster_snapshots}}
request. If this parameter is specified, the response includes only
records beyond the marker, up to the value specified by \code{MaxRecords}.}

\item{IncludeShared}{Specifies whether to include shared manual DB cluster snapshots from
other Amazon Web Services accounts that this Amazon Web Services account
has been given permission to copy or restore. By default, these
snapshots are not included.

You can give an Amazon Web Services account permission to restore a
manual DB cluster snapshot from another Amazon Web Services account by
the
\code{\link[=rds_modify_db_cluster_snapshot_attribute]{modify_db_cluster_snapshot_attribute}}
API action.}

\item{IncludePublic}{Specifies whether to include manual DB cluster snapshots that are public
and can be copied or restored by any Amazon Web Services account. By
default, the public snapshots are not included.

You can share a manual DB cluster snapshot as public by using the
\code{\link[=rds_modify_db_cluster_snapshot_attribute]{modify_db_cluster_snapshot_attribute}}
API action.}

\item{DbClusterResourceId}{A specific DB cluster resource ID to describe.}
}
\description{
Returns information about DB cluster snapshots. This API action supports pagination.

See \url{https://www.paws-r-sdk.com/docs/rds_describe_db_cluster_snapshots/} for full documentation.
}
\keyword{internal}
