% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_describe_db_proxy_targets}
\alias{rds_describe_db_proxy_targets}
\title{Returns information about DBProxyTarget objects}
\usage{
rds_describe_db_proxy_targets(DBProxyName, TargetGroupName, Filters,
  Marker, MaxRecords)
}
\arguments{
\item{DBProxyName}{[required] The identifier of the \code{DBProxyTarget} to describe.}

\item{TargetGroupName}{The identifier of the \code{DBProxyTargetGroup} to describe.}

\item{Filters}{This parameter is not currently supported.}

\item{Marker}{An optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
called a marker is included in the response so that the remaining
results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Targets = list(
    list(
      TargetArn = "string",
      Endpoint = "string",
      TrackedClusterId = "string",
      RdsResourceId = "string",
      Port = 123,
      Type = "RDS_INSTANCE"|"RDS_SERVERLESS_ENDPOINT"|"TRACKED_CLUSTER",
      TargetHealth = list(
        State = "REGISTERING"|"AVAILABLE"|"UNAVAILABLE",
        Reason = "UNREACHABLE"|"CONNECTION_FAILED"|"AUTH_FAILURE"|"PENDING_PROXY_CAPACITY",
        Description = "string"
      )
    )
  ),
  Marker = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns information about \code{DBProxyTarget} objects. This API supports
pagination.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_db_proxy_targets(
  DBProxyName = "string",
  TargetGroupName = "string",
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  Marker = "string",
  MaxRecords = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
