% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_describe_db_snapshot_attributes}
\alias{rds_describe_db_snapshot_attributes}
\title{Returns a list of DB snapshot attribute names and values for a manual DB
snapshot}
\usage{
rds_describe_db_snapshot_attributes(DBSnapshotIdentifier)
}
\arguments{
\item{DBSnapshotIdentifier}{[required] The identifier for the DB snapshot to describe the attributes for.}
}
\description{
Returns a list of DB snapshot attribute names and values for a manual DB snapshot.

See \url{https://paws-r.github.io/docs/rds/describe_db_snapshot_attributes.html} for full documentation.
}
\keyword{internal}
