% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_describe_event_categories}
\alias{rds_describe_event_categories}
\title{Displays a list of categories for all event source types, or, if
specified, for a specified source type}
\usage{
rds_describe_event_categories(SourceType = NULL, Filters = NULL)
}
\arguments{
\item{SourceType}{The type of source that is generating the events. For RDS Proxy events,
specify \code{db-proxy}.

Valid values: \code{db-instance} | \code{db-cluster} | \code{db-parameter-group} |
\code{db-security-group} | \code{db-snapshot} | \code{db-cluster-snapshot} | \code{db-proxy}}

\item{Filters}{This parameter isn't currently supported.}
}
\description{
Displays a list of categories for all event source types, or, if specified, for a specified source type. You can also see this list in the "Amazon RDS event categories and event messages" section of the \href{https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/USER_Events.Messages.html}{\emph{Amazon RDS User Guide}} or the \href{https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/USER_Events.Messages.html}{\emph{Amazon Aurora User Guide}} .

See \url{https://paws-r.github.io/docs/rds/describe_event_categories.html} for full documentation.
}
\keyword{internal}
