% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_describe_integrations}
\alias{rds_describe_integrations}
\title{Describe one or more zero-ETL integrations with Amazon Redshift}
\usage{
rds_describe_integrations(
  IntegrationIdentifier = NULL,
  Filters = NULL,
  MaxRecords = NULL,
  Marker = NULL
)
}
\arguments{
\item{IntegrationIdentifier}{The unique identifier of the integration.}

\item{Filters}{A filter that specifies one or more resources to return.}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
called a marker is included in the response so that you can retrieve the
remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.}

\item{Marker}{An optional pagination token provided by a previous
\code{\link[=rds_describe_integrations]{describe_integrations}} request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}
}
\description{
Describe one or more zero-ETL integrations with Amazon Redshift.

See \url{https://www.paws-r-sdk.com/docs/rds_describe_integrations/} for full documentation.
}
\keyword{internal}
