% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_describe_option_group_options}
\alias{rds_describe_option_group_options}
\title{Describes all available options for the specified engine}
\usage{
rds_describe_option_group_options(
  EngineName,
  MajorEngineVersion = NULL,
  Filters = NULL,
  MaxRecords = NULL,
  Marker = NULL
)
}
\arguments{
\item{EngineName}{[required] The name of the engine to describe options for.

Valid Values:
\itemize{
\item \code{db2-ae}
\item \code{db2-ce}
\item \code{db2-se}
\item \code{mariadb}
\item \code{mysql}
\item \code{oracle-ee}
\item \code{oracle-ee-cdb}
\item \code{oracle-se2}
\item \code{oracle-se2-cdb}
\item \code{postgres}
\item \code{sqlserver-ee}
\item \code{sqlserver-se}
\item \code{sqlserver-ex}
\item \code{sqlserver-web}
}}

\item{MajorEngineVersion}{If specified, filters the results to include only options for the specified major engine version.}

\item{Filters}{This parameter isn't currently supported.}

\item{MaxRecords}{The maximum number of records to include in the response. If more records exist than the specified \code{MaxRecords} value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.}

\item{Marker}{An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by \code{MaxRecords}.}
}
\description{
Describes all available options for the specified engine.

See \url{https://www.paws-r-sdk.com/docs/rds_describe_option_group_options/} for full documentation.
}
\keyword{internal}
