% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_describe_serverless_v2_platform_versions}
\alias{rds_describe_serverless_v2_platform_versions}
\title{Describes the properties of specific platform versions for Aurora
Serverless v2}
\usage{
rds_describe_serverless_v2_platform_versions(
  ServerlessV2PlatformVersion = NULL,
  Engine = NULL,
  Filters = NULL,
  DefaultOnly = NULL,
  IncludeAll = NULL,
  MaxRecords = NULL,
  Marker = NULL
)
}
\arguments{
\item{ServerlessV2PlatformVersion}{A specific platform version to return details for.

Example: \code{3}}

\item{Engine}{The database engine to return platform version details for.

Valid Values:
\itemize{
\item \code{aurora-mysql}
\item \code{aurora-postgresql}
}}

\item{Filters}{This parameter isn't currently supported.}

\item{DefaultOnly}{Specifies whether to return only the default platform versions for each
engine. The default platform version is the version used for new DB
clusters.}

\item{IncludeAll}{Specifies whether to also include platform versions which are no longer
in use.}

\item{MaxRecords}{The maximum number of records to include in the response. If more than
the \code{MaxRecords} value is available, a pagination token called a marker
is included in the response so you can retrieve the remaining results.

Default: 20

Constraints: Minimum 1, maximum 200.}

\item{Marker}{An optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}
}
\description{
Describes the properties of specific platform versions for Aurora Serverless v2.

See \url{https://www.paws-r-sdk.com/docs/rds_describe_serverless_v2_platform_versions/} for full documentation.
}
\keyword{internal}
