% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_disable_http_endpoint}
\alias{rds_disable_http_endpoint}
\title{Disables the HTTP endpoint for the specified DB cluster}
\usage{
rds_disable_http_endpoint(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the DB cluster.}
}
\description{
Disables the HTTP endpoint for the specified DB cluster. Disabling this endpoint disables RDS Data API.

See \url{https://www.paws-r-sdk.com/docs/rds_disable_http_endpoint/} for full documentation.
}
\keyword{internal}
