% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_enable_http_endpoint}
\alias{rds_enable_http_endpoint}
\title{Enables the HTTP endpoint for the DB cluster}
\usage{
rds_enable_http_endpoint(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the DB cluster.}
}
\description{
Enables the HTTP endpoint for the DB cluster. By default, the HTTP endpoint isn't enabled.

See \url{https://www.paws-r-sdk.com/docs/rds_enable_http_endpoint/} for full documentation.
}
\keyword{internal}
