% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_failover_db_cluster}
\alias{rds_failover_db_cluster}
\title{Forces a failover for a DB cluster}
\usage{
rds_failover_db_cluster(DBClusterIdentifier, TargetDBInstanceIdentifier = NULL)
}
\arguments{
\item{DBClusterIdentifier}{[required] A DB cluster identifier to force a failover for. This parameter isn't
case-sensitive.

Constraints:
\itemize{
\item Must match the identifier of an existing DBCluster.
}}

\item{TargetDBInstanceIdentifier}{The name of the DB instance to promote to the primary DB instance.

Specify the DB instance identifier for an Aurora Replica or a Multi-AZ
readable standby in the DB cluster, for example \code{mydbcluster-replica1}.

This setting isn't supported for RDS for MySQL Multi-AZ DB clusters.}
}
\description{
Forces a failover for a DB cluster.

See \url{https://www.paws-r-sdk.com/docs/rds_failover_db_cluster/} for full documentation.
}
\keyword{internal}
