% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_modify_db_cluster_snapshot_attribute}
\alias{rds_modify_db_cluster_snapshot_attribute}
\title{Adds an attribute and values to, or removes an attribute and values
from, a manual DB cluster snapshot}
\usage{
rds_modify_db_cluster_snapshot_attribute(DBClusterSnapshotIdentifier,
  AttributeName, ValuesToAdd, ValuesToRemove)
}
\arguments{
\item{DBClusterSnapshotIdentifier}{[required] The identifier for the DB cluster snapshot to modify the attributes for.}

\item{AttributeName}{[required] The name of the DB cluster snapshot attribute to modify.

To manage authorization for other AWS accounts to copy or restore a
manual DB cluster snapshot, set this value to \code{restore}.

To view the list of attributes available to modify, use the
\code{\link[=rds_describe_db_cluster_snapshot_attributes]{describe_db_cluster_snapshot_attributes}}
API action.}

\item{ValuesToAdd}{A list of DB cluster snapshot attributes to add to the attribute
specified by \code{AttributeName}.

To authorize other AWS accounts to copy or restore a manual DB cluster
snapshot, set this list to include one or more AWS account IDs, or \code{all}
to make the manual DB cluster snapshot restorable by any AWS account. Do
not add the \code{all} value for any manual DB cluster snapshots that contain
private information that you don't want available to all AWS accounts.}

\item{ValuesToRemove}{A list of DB cluster snapshot attributes to remove from the attribute
specified by \code{AttributeName}.

To remove authorization for other AWS accounts to copy or restore a
manual DB cluster snapshot, set this list to include one or more AWS
account identifiers, or \code{all} to remove authorization for any AWS
account to copy or restore the DB cluster snapshot. If you specify
\code{all}, an AWS account whose account ID is explicitly added to the
\code{restore} attribute can still copy or restore a manual DB cluster
snapshot.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DBClusterSnapshotAttributesResult = list(
    DBClusterSnapshotIdentifier = "string",
    DBClusterSnapshotAttributes = list(
      list(
        AttributeName = "string",
        AttributeValues = list(
          "string"
        )
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Adds an attribute and values to, or removes an attribute and values
from, a manual DB cluster snapshot.

To share a manual DB cluster snapshot with other AWS accounts, specify
\code{restore} as the \code{AttributeName} and use the \code{ValuesToAdd} parameter to
add a list of IDs of the AWS accounts that are authorized to restore the
manual DB cluster snapshot. Use the value \code{all} to make the manual DB
cluster snapshot public, which means that it can be copied or restored
by all AWS accounts.

Don't add the \code{all} value for any manual DB cluster snapshots that
contain private information that you don't want available to all AWS
accounts.

If a manual DB cluster snapshot is encrypted, it can be shared, but only
by specifying a list of authorized AWS account IDs for the \code{ValuesToAdd}
parameter. You can't use \code{all} as a value for that parameter in this
case.

To view which AWS accounts have access to copy or restore a manual DB
cluster snapshot, or whether a manual DB cluster snapshot is public or
private, use the
\code{\link[=rds_describe_db_cluster_snapshot_attributes]{describe_db_cluster_snapshot_attributes}}
API action. The accounts are returned as values for the \code{restore}
attribute.

This action only applies to Aurora DB clusters.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$modify_db_cluster_snapshot_attribute(
  DBClusterSnapshotIdentifier = "string",
  AttributeName = "string",
  ValuesToAdd = list(
    "string"
  ),
  ValuesToRemove = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
