% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_modify_db_proxy_endpoint}
\alias{rds_modify_db_proxy_endpoint}
\title{Changes the settings for an existing DB proxy endpoint}
\usage{
rds_modify_db_proxy_endpoint(
  DBProxyEndpointName,
  NewDBProxyEndpointName = NULL,
  VpcSecurityGroupIds = NULL
)
}
\arguments{
\item{DBProxyEndpointName}{[required] The name of the DB proxy sociated with the DB proxy endpoint that you
want to modify.}

\item{NewDBProxyEndpointName}{The new identifier for the \code{DBProxyEndpoint}. An identifier must begin
with a letter and must contain only ASCII letters, digits, and hyphens;
it can't end with a hyphen or contain two consecutive hyphens.}

\item{VpcSecurityGroupIds}{The VPC security group IDs for the DB proxy endpoint. When the DB proxy
endpoint uses a different VPC than the original proxy, you also specify
a different set of security group IDs than for the original proxy.}
}
\description{
Changes the settings for an existing DB proxy endpoint.

See \url{https://paws-r.github.io/docs/rds/modify_db_proxy_endpoint.html} for full documentation.
}
\keyword{internal}
