% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_modify_db_proxy_target_group}
\alias{rds_modify_db_proxy_target_group}
\title{Modifies the properties of a DBProxyTargetGroup}
\usage{
rds_modify_db_proxy_target_group(
  TargetGroupName,
  DBProxyName,
  ConnectionPoolConfig = NULL,
  NewName = NULL
)
}
\arguments{
\item{TargetGroupName}{[required] The name of the target group to modify.}

\item{DBProxyName}{[required] The name of the proxy.}

\item{ConnectionPoolConfig}{The settings that determine the size and behavior of the connection pool
for the target group.}

\item{NewName}{The new name for the modified \code{DBProxyTarget}. An identifier must begin
with a letter and must contain only ASCII letters, digits, and hyphens;
it can't end with a hyphen or contain two consecutive hyphens.

You can't rename the \code{default} target group.}
}
\description{
Modifies the properties of a \code{DBProxyTargetGroup}.

See \url{https://www.paws-r-sdk.com/docs/rds_modify_db_proxy_target_group/} for full documentation.
}
\keyword{internal}
