% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_modify_db_recommendation}
\alias{rds_modify_db_recommendation}
\title{Updates the recommendation status and recommended action status for the
specified recommendation}
\usage{
rds_modify_db_recommendation(
  RecommendationId,
  Locale = NULL,
  Status = NULL,
  RecommendedActionUpdates = NULL
)
}
\arguments{
\item{RecommendationId}{[required] The identifier of the recommendation to update.}

\item{Locale}{The language of the modified recommendation.}

\item{Status}{The recommendation status to update.

Valid values:
\itemize{
\item active
\item dismissed
}}

\item{RecommendedActionUpdates}{The list of recommended action status to update. You can update multiple
recommended actions at one time.}
}
\description{
Updates the recommendation status and recommended action status for the specified recommendation.

See \url{https://www.paws-r-sdk.com/docs/rds_modify_db_recommendation/} for full documentation.
}
\keyword{internal}
