% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_modify_integration}
\alias{rds_modify_integration}
\title{Modifies a zero-ETL integration with Amazon Redshift}
\usage{
rds_modify_integration(
  IntegrationIdentifier,
  IntegrationName = NULL,
  DataFilter = NULL,
  Description = NULL
)
}
\arguments{
\item{IntegrationIdentifier}{[required] The unique identifier of the integration to modify.}

\item{IntegrationName}{A new name for the integration.}

\item{DataFilter}{A new data filter for the integration. For more information, see \href{https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/}{Data filtering for Aurora zero-ETL integrations with Amazon Redshift}
or \href{https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/zero-etl.filtering.html}{Data filtering for Amazon RDS zero-ETL integrations with Amazon Redshift}.}

\item{Description}{A new description for the integration.}
}
\description{
Modifies a zero-ETL integration with Amazon Redshift.

See \url{https://www.paws-r-sdk.com/docs/rds_modify_integration/} for full documentation.
}
\keyword{internal}
