% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_purchase_reserved_db_instances_offering}
\alias{rds_purchase_reserved_db_instances_offering}
\title{Purchases a reserved DB instance offering}
\usage{
rds_purchase_reserved_db_instances_offering(
  ReservedDBInstancesOfferingId,
  ReservedDBInstanceId = NULL,
  DBInstanceCount = NULL,
  Tags = NULL
)
}
\arguments{
\item{ReservedDBInstancesOfferingId}{[required] The ID of the Reserved DB instance offering to purchase.

Example: 438012d3-4052-4cc7-b2e3-8d3372e0e706}

\item{ReservedDBInstanceId}{Customer-specified identifier to track this reservation.

Example: myreservationID}

\item{DBInstanceCount}{The number of instances to reserve.

Default: \code{1}}

\item{Tags}{A list of tags.

For more information, see \href{https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/USER_Tagging.html}{Tagging Amazon RDS resources} in the \emph{Amazon RDS User Guide} or \href{https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/USER_Tagging.html}{Tagging Amazon Aurora and Amazon RDS resources} in the \emph{Amazon Aurora User Guide}.}
}
\description{
Purchases a reserved DB instance offering.

See \url{https://www.paws-r-sdk.com/docs/rds_purchase_reserved_db_instances_offering/} for full documentation.
}
\keyword{internal}
