% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_reboot_db_cluster}
\alias{rds_reboot_db_cluster}
\title{You might need to reboot your DB cluster, usually for maintenance
reasons}
\usage{
rds_reboot_db_cluster(DBClusterIdentifier)
}
\arguments{
\item{DBClusterIdentifier}{[required] The DB cluster identifier. This parameter is stored as a lowercase
string.

Constraints:
\itemize{
\item Must match the identifier of an existing DBCluster.
}}
}
\description{
You might need to reboot your DB cluster, usually for maintenance reasons. For example, if you make certain modifications, or if you change the DB cluster parameter group associated with the DB cluster, reboot the DB cluster for the changes to take effect.

See \url{https://paws-r.github.io/docs/rds/reboot_db_cluster.html} for full documentation.
}
\keyword{internal}
