% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_reboot_db_shard_group}
\alias{rds_reboot_db_shard_group}
\title{You might need to reboot your DB shard group, usually for maintenance
reasons}
\usage{
rds_reboot_db_shard_group(DBShardGroupIdentifier)
}
\arguments{
\item{DBShardGroupIdentifier}{[required] The name of the DB shard group to reboot.}
}
\description{
You might need to reboot your DB shard group, usually for maintenance reasons. For example, if you make certain modifications, reboot the DB shard group for the changes to take effect.

See \url{https://www.paws-r-sdk.com/docs/rds_reboot_db_shard_group/} for full documentation.
}
\keyword{internal}
