% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_register_db_proxy_targets}
\alias{rds_register_db_proxy_targets}
\title{Associate one or more DBProxyTarget data structures with a
DBProxyTargetGroup}
\usage{
rds_register_db_proxy_targets(
  DBProxyName,
  TargetGroupName = NULL,
  DBInstanceIdentifiers = NULL,
  DBClusterIdentifiers = NULL
)
}
\arguments{
\item{DBProxyName}{[required] The identifier of the \code{DBProxy} that is associated with the
\code{DBProxyTargetGroup}.}

\item{TargetGroupName}{The identifier of the \code{DBProxyTargetGroup}.}

\item{DBInstanceIdentifiers}{One or more DB instance identifiers.}

\item{DBClusterIdentifiers}{One or more DB cluster identifiers.}
}
\description{
Associate one or more \code{DBProxyTarget} data structures with a \code{DBProxyTargetGroup}.

See \url{https://www.paws-r-sdk.com/docs/rds_register_db_proxy_targets/} for full documentation.
}
\keyword{internal}
