% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_start_db_cluster}
\alias{rds_start_db_cluster}
\title{Starts an Amazon Aurora DB cluster that was stopped using the Amazon Web
Services console, the stop-db-cluster CLI command, or the StopDBCluster
action}
\usage{
rds_start_db_cluster(DBClusterIdentifier)
}
\arguments{
\item{DBClusterIdentifier}{[required] The DB cluster identifier of the Amazon Aurora DB cluster to be started.
This parameter is stored as a lowercase string.}
}
\description{
Starts an Amazon Aurora DB cluster that was stopped using the Amazon Web Services console, the stop-db-cluster CLI command, or the StopDBCluster action.

See \url{https://www.paws-r-sdk.com/docs/rds_start_db_cluster/} for full documentation.
}
\keyword{internal}
