% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_start_db_instance}
\alias{rds_start_db_instance}
\title{Starts an Amazon RDS DB instance that was stopped using the Amazon Web
Services console, the stop-db-instance CLI command, or the
StopDBInstance operation}
\usage{
rds_start_db_instance(DBInstanceIdentifier)
}
\arguments{
\item{DBInstanceIdentifier}{[required] The user-supplied instance identifier.}
}
\description{
Starts an Amazon RDS DB instance that was stopped using the Amazon Web Services console, the stop-db-instance CLI command, or the \code{\link[=rds_stop_db_instance]{stop_db_instance}} operation.

See \url{https://www.paws-r-sdk.com/docs/rds_start_db_instance/} for full documentation.
}
\keyword{internal}
