% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_start_export_task}
\alias{rds_start_export_task}
\title{Starts an export of a snapshot to Amazon S3}
\usage{
rds_start_export_task(
  ExportTaskIdentifier,
  SourceArn,
  S3BucketName,
  IamRoleArn,
  KmsKeyId,
  S3Prefix = NULL,
  ExportOnly = NULL
)
}
\arguments{
\item{ExportTaskIdentifier}{[required] A unique identifier for the snapshot export task. This ID isn't an
identifier for the Amazon S3 bucket where the snapshot is to be exported
to.}

\item{SourceArn}{[required] The Amazon Resource Name (ARN) of the snapshot to export to Amazon S3.}

\item{S3BucketName}{[required] The name of the Amazon S3 bucket to export the snapshot to.}

\item{IamRoleArn}{[required] The name of the IAM role to use for writing to the Amazon S3 bucket when
exporting a snapshot.}

\item{KmsKeyId}{[required] The ID of the Amazon Web Services KMS key to use to encrypt the snapshot
exported to Amazon S3. The Amazon Web Services KMS key identifier is the
key ARN, key ID, alias ARN, or alias name for the KMS key. The caller of
this operation must be authorized to execute the following operations.
These can be set in the Amazon Web Services KMS key policy:
\itemize{
\item GrantOperation.Encrypt
\item GrantOperation.Decrypt
\item GrantOperation.GenerateDataKey
\item GrantOperation.GenerateDataKeyWithoutPlaintext
\item GrantOperation.ReEncryptFrom
\item GrantOperation.ReEncryptTo
\item GrantOperation.CreateGrant
\item GrantOperation.DescribeKey
\item GrantOperation.RetireGrant
}}

\item{S3Prefix}{The Amazon S3 bucket prefix to use as the file name and path of the
exported snapshot.}

\item{ExportOnly}{The data to be exported from the snapshot. If this parameter is not
provided, all the snapshot data is exported. Valid values are the
following:
\itemize{
\item \code{database} - Export all the data from a specified database.
\item \code{database.table} \emph{table-name} - Export a table of the snapshot. This
format is valid only for RDS for MySQL, RDS for MariaDB, and Aurora
MySQL.
\item \code{database.schema} \emph{schema-name} - Export a database schema of the
snapshot. This format is valid only for RDS for PostgreSQL and
Aurora PostgreSQL.
\item \code{database.schema.table} \emph{table-name} - Export a table of the
database schema. This format is valid only for RDS for PostgreSQL
and Aurora PostgreSQL.
}}
}
\description{
Starts an export of a snapshot to Amazon S3. The provided IAM role must have access to the S3 bucket.

See \url{https://paws-r.github.io/docs/rds/start_export_task.html} for full documentation.
}
\keyword{internal}
