% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_switchover_read_replica}
\alias{rds_switchover_read_replica}
\title{Switches over an Oracle standby database in an Oracle Data Guard
environment, making it the new primary database}
\usage{
rds_switchover_read_replica(DBInstanceIdentifier)
}
\arguments{
\item{DBInstanceIdentifier}{[required] The DB instance identifier of the current standby database. This value
is stored as a lowercase string.

Constraints:
\itemize{
\item Must match the identiﬁer of an existing Oracle read replica DB
instance.
}}
}
\description{
Switches over an Oracle standby database in an Oracle Data Guard environment, making it the new primary database. Issue this command in the Region that hosts the current standby database.

See \url{https://www.paws-r-sdk.com/docs/rds_switchover_read_replica/} for full documentation.
}
\keyword{internal}
