% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_authorize_endpoint_access}
\alias{redshift_authorize_endpoint_access}
\title{Grants access to a cluster}
\usage{
redshift_authorize_endpoint_access(
  ClusterIdentifier = NULL,
  Account,
  VpcIds = NULL
)
}
\arguments{
\item{ClusterIdentifier}{The cluster identifier of the cluster to grant access to.}

\item{Account}{[required] The Amazon Web Services account ID to grant access to.}

\item{VpcIds}{The virtual private cloud (VPC) identifiers to grant access to.}
}
\description{
Grants access to a cluster.

See \url{https://paws-r.github.io/docs/redshift/authorize_endpoint_access.html} for full documentation.
}
\keyword{internal}
