% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_cancel_resize}
\alias{redshift_cancel_resize}
\title{Cancels a resize operation for a cluster}
\usage{
redshift_cancel_resize(ClusterIdentifier)
}
\arguments{
\item{ClusterIdentifier}{[required] The unique identifier for the cluster that you want to cancel a resize
operation for.}
}
\description{
Cancels a resize operation for a cluster.

See \url{https://paws-r.github.io/docs/redshift/cancel_resize.html} for full documentation.
}
\keyword{internal}
