% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_create_authentication_profile}
\alias{redshift_create_authentication_profile}
\title{Creates an authentication profile with the specified parameters}
\usage{
redshift_create_authentication_profile(
  AuthenticationProfileName,
  AuthenticationProfileContent
)
}
\arguments{
\item{AuthenticationProfileName}{[required] The name of the authentication profile to be created.}

\item{AuthenticationProfileContent}{[required] The content of the authentication profile in JSON format. The maximum
length of the JSON string is determined by a quota for your account.}
}
\description{
Creates an authentication profile with the specified parameters.

See \url{https://paws-r.github.io/docs/redshift/create_authentication_profile.html} for full documentation.
}
\keyword{internal}
