% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_create_custom_domain_association}
\alias{redshift_create_custom_domain_association}
\title{Used to create a custom domain name for a cluster}
\usage{
redshift_create_custom_domain_association(
  CustomDomainName,
  CustomDomainCertificateArn,
  ClusterIdentifier
)
}
\arguments{
\item{CustomDomainName}{[required] The custom domain name for a custom domain association.}

\item{CustomDomainCertificateArn}{[required] The certificate Amazon Resource Name (ARN) for the custom domain name
association.}

\item{ClusterIdentifier}{[required] The cluster identifier that the custom domain is associated with.}
}
\description{
Used to create a custom domain name for a cluster. Properties include the custom domain name, the cluster the custom domain is associated with, and the certificate Amazon Resource Name (ARN).

See \url{https://www.paws-r-sdk.com/docs/redshift_create_custom_domain_association/} for full documentation.
}
\keyword{internal}
