% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_create_hsm_client_certificate}
\alias{redshift_create_hsm_client_certificate}
\title{Creates an HSM client certificate that an Amazon Redshift cluster will
use to connect to the client's HSM in order to store and retrieve the
keys used to encrypt the cluster databases}
\usage{
redshift_create_hsm_client_certificate(
  HsmClientCertificateIdentifier,
  Tags = NULL
)
}
\arguments{
\item{HsmClientCertificateIdentifier}{[required] The identifier to be assigned to the new HSM client certificate that the
cluster will use to connect to the HSM to use the database encryption
keys.}

\item{Tags}{A list of tag instances.}
}
\description{
Creates an HSM client certificate that an Amazon Redshift cluster will use to connect to the client's HSM in order to store and retrieve the keys used to encrypt the cluster databases.

See \url{https://paws-r.github.io/docs/redshift/create_hsm_client_certificate.html} for full documentation.
}
\keyword{internal}
