% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_create_qev_2_idc_application}
\alias{redshift_create_qev_2_idc_application}
\title{Creates an Amazon Redshift Query Editor (QEV2) IAM Identity Center
application}
\usage{
redshift_create_qev_2_idc_application(
  IdcInstanceArn,
  Qev2IdcApplicationName,
  IdcDisplayName,
  Tags = NULL
)
}
\arguments{
\item{IdcInstanceArn}{[required] The Amazon Resource Name (ARN) of the IAM Identity Center instance used to create the Amazon Redshift Query Editor (QEV2) managed application.}

\item{Qev2IdcApplicationName}{[required] The name of the Amazon Redshift Query Editor (QEV2) application in IAM Identity Center.}

\item{IdcDisplayName}{[required] The display name for the Amazon Redshift Query Editor (QEV2) IAM Identity Center application. It appears in the console.}

\item{Tags}{A list of tags to associate with the application. Tags are key-value pairs that you can use to organize and identify your resources.}
}
\description{
Creates an Amazon Redshift Query Editor (QEV2) IAM Identity Center application.

See \url{https://www.paws-r-sdk.com/docs/redshift_create_qev_2_idc_application/} for full documentation.
}
\keyword{internal}
