% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_deauthorize_data_share}
\alias{redshift_deauthorize_data_share}
\title{From a datashare producer account, removes authorization from the
specified datashare}
\usage{
redshift_deauthorize_data_share(DataShareArn, ConsumerIdentifier)
}
\arguments{
\item{DataShareArn}{[required] The Amazon Resource Name (ARN) of the datashare to remove authorization
from.}

\item{ConsumerIdentifier}{[required] The identifier of the data consumer that is to have authorization
removed from the datashare. This identifier is an Amazon Web Services
account ID or a keyword, such as ADX.}
}
\description{
From a datashare producer account, removes authorization from the specified datashare.

See \url{https://www.paws-r-sdk.com/docs/redshift_deauthorize_data_share/} for full documentation.
}
\keyword{internal}
