% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_delete_cluster_subnet_group}
\alias{redshift_delete_cluster_subnet_group}
\title{Deletes the specified cluster subnet group}
\usage{
redshift_delete_cluster_subnet_group(ClusterSubnetGroupName)
}
\arguments{
\item{ClusterSubnetGroupName}{[required] The name of the cluster subnet group name to be deleted.}
}
\description{
Deletes the specified cluster subnet group.

See \url{https://www.paws-r-sdk.com/docs/redshift_delete_cluster_subnet_group/} for full documentation.
}
\keyword{internal}
