% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_delete_endpoint_access}
\alias{redshift_delete_endpoint_access}
\title{Deletes a Redshift-managed VPC endpoint}
\usage{
redshift_delete_endpoint_access(EndpointName)
}
\arguments{
\item{EndpointName}{[required] The Redshift-managed VPC endpoint to delete.}
}
\value{
A list with the following syntax:\preformatted{list(
  ClusterIdentifier = "string",
  ResourceOwner = "string",
  SubnetGroupName = "string",
  EndpointStatus = "string",
  EndpointName = "string",
  EndpointCreateTime = as.POSIXct(
    "2015-01-01"
  ),
  Port = 123,
  Address = "string",
  VpcSecurityGroups = list(
    list(
      VpcSecurityGroupId = "string",
      Status = "string"
    )
  ),
  VpcEndpoint = list(
    VpcEndpointId = "string",
    VpcId = "string",
    NetworkInterfaces = list(
      list(
        NetworkInterfaceId = "string",
        SubnetId = "string",
        PrivateIpAddress = "string",
        AvailabilityZone = "string"
      )
    )
  )
)
}
}
\description{
Deletes a Redshift-managed VPC endpoint.
}
\section{Request syntax}{
\preformatted{svc$delete_endpoint_access(
  EndpointName = "string"
)
}
}

\keyword{internal}
