% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_delete_event_subscription}
\alias{redshift_delete_event_subscription}
\title{Deletes an Amazon Redshift event notification subscription}
\usage{
redshift_delete_event_subscription(SubscriptionName)
}
\arguments{
\item{SubscriptionName}{[required] The name of the Amazon Redshift event notification subscription to be
deleted.}
}
\value{
An empty list.
}
\description{
Deletes an Amazon Redshift event notification subscription.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_event_subscription(
  SubscriptionName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
