% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_delete_hsm_configuration}
\alias{redshift_delete_hsm_configuration}
\title{Deletes the specified Amazon Redshift HSM configuration}
\usage{
redshift_delete_hsm_configuration(HsmConfigurationIdentifier)
}
\arguments{
\item{HsmConfigurationIdentifier}{[required] The identifier of the Amazon Redshift HSM configuration to be deleted.}
}
\value{
An empty list.
}
\description{
Deletes the specified Amazon Redshift HSM configuration.
}
\section{Request syntax}{
\preformatted{svc$delete_hsm_configuration(
  HsmConfigurationIdentifier = "string"
)
}
}

\keyword{internal}
