% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_delete_integration}
\alias{redshift_delete_integration}
\title{Deletes a zero-ETL integration or S3 event integration with Amazon
Redshift}
\usage{
redshift_delete_integration(IntegrationArn)
}
\arguments{
\item{IntegrationArn}{[required] The unique identifier of the integration to delete.}
}
\description{
Deletes a zero-ETL integration or S3 event integration with Amazon Redshift.

See \url{https://www.paws-r-sdk.com/docs/redshift_delete_integration/} for full documentation.
}
\keyword{internal}
