% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_delete_qev_2_idc_application}
\alias{redshift_delete_qev_2_idc_application}
\title{Deletes an Amazon Redshift Query Editor (QEV2) IAM Identity Center
application}
\usage{
redshift_delete_qev_2_idc_application(Qev2IdcApplicationArn)
}
\arguments{
\item{Qev2IdcApplicationArn}{[required] The Amazon Resource Name (ARN) for the Amazon Redshift Query Editor (QEV2) IAM Identity Center application to delete.}
}
\description{
Deletes an Amazon Redshift Query Editor (QEV2) IAM Identity Center application.

See \url{https://www.paws-r-sdk.com/docs/redshift_delete_qev_2_idc_application/} for full documentation.
}
\keyword{internal}
