% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_delete_redshift_idc_application}
\alias{redshift_delete_redshift_idc_application}
\title{Deletes an Amazon Redshift IAM Identity Center application}
\usage{
redshift_delete_redshift_idc_application(RedshiftIdcApplicationArn)
}
\arguments{
\item{RedshiftIdcApplicationArn}{[required] The ARN for a deleted Amazon Redshift IAM Identity Center application.}
}
\description{
Deletes an Amazon Redshift IAM Identity Center application.

See \url{https://www.paws-r-sdk.com/docs/redshift_delete_redshift_idc_application/} for full documentation.
}
\keyword{internal}
