% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_deregister_namespace}
\alias{redshift_deregister_namespace}
\title{Deregisters a cluster or serverless namespace from the Amazon Web
Services Glue Data Catalog}
\usage{
redshift_deregister_namespace(NamespaceIdentifier, ConsumerIdentifiers)
}
\arguments{
\item{NamespaceIdentifier}{[required] The unique identifier of the cluster or serverless namespace that you
want to deregister.}

\item{ConsumerIdentifiers}{[required] An array containing the ID of the consumer account that you want to
deregister the cluster or serverless namespace from.}
}
\description{
Deregisters a cluster or serverless namespace from the Amazon Web Services Glue Data Catalog.

See \url{https://www.paws-r-sdk.com/docs/redshift_deregister_namespace/} for full documentation.
}
\keyword{internal}
