% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_describe_account_attributes}
\alias{redshift_describe_account_attributes}
\title{Returns a list of attributes attached to an account}
\usage{
redshift_describe_account_attributes(AttributeNames = NULL)
}
\arguments{
\item{AttributeNames}{A list of attribute names.}
}
\description{
Returns a list of attributes attached to an account

See \url{https://paws-r.github.io/docs/redshift/describe_account_attributes.html} for full documentation.
}
\keyword{internal}
