% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_describe_authentication_profiles}
\alias{redshift_describe_authentication_profiles}
\title{Describes an authentication profile}
\usage{
redshift_describe_authentication_profiles(AuthenticationProfileName = NULL)
}
\arguments{
\item{AuthenticationProfileName}{The name of the authentication profile to describe. If not specified
then all authentication profiles owned by the account are listed.}
}
\description{
Describes an authentication profile.

See \url{https://www.paws-r-sdk.com/docs/redshift_describe_authentication_profiles/} for full documentation.
}
\keyword{internal}
