% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_describe_custom_domain_associations}
\alias{redshift_describe_custom_domain_associations}
\title{Contains information about custom domain associations for a cluster}
\usage{
redshift_describe_custom_domain_associations(
  CustomDomainName = NULL,
  CustomDomainCertificateArn = NULL,
  MaxRecords = NULL,
  Marker = NULL
)
}
\arguments{
\item{CustomDomainName}{The custom domain name for the custom domain association.}

\item{CustomDomainCertificateArn}{The certificate Amazon Resource Name (ARN) for the custom domain
association.}

\item{MaxRecords}{The maximum records setting for the associated custom domain.}

\item{Marker}{The marker for the custom domain association.}
}
\description{
Contains information about custom domain associations for a cluster.

See \url{https://www.paws-r-sdk.com/docs/redshift_describe_custom_domain_associations/} for full documentation.
}
\keyword{internal}
