% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_describe_event_categories}
\alias{redshift_describe_event_categories}
\title{Displays a list of event categories for all event source types, or for a
specified source type}
\usage{
redshift_describe_event_categories(SourceType)
}
\arguments{
\item{SourceType}{The source type, such as cluster or parameter group, to which the
described event categories apply.

Valid values: cluster, cluster-snapshot, cluster-parameter-group,
cluster-security-group, and scheduled-action.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  EventCategoriesMapList = list(
    list(
      SourceType = "string",
      Events = list(
        list(
          EventId = "string",
          EventCategories = list(
            "string"
          ),
          EventDescription = "string",
          Severity = "string"
        )
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Displays a list of event categories for all event source types, or for a
specified source type. For a list of the event categories and source
types, go to \href{https://docs.aws.amazon.com/redshift/latest/mgmt/working-with-event-notifications.html}{Amazon Redshift Event Notifications}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_event_categories(
  SourceType = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
