% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_describe_redshift_idc_applications}
\alias{redshift_describe_redshift_idc_applications}
\title{Lists the Amazon Redshift IAM Identity Center applications}
\usage{
redshift_describe_redshift_idc_applications(
  RedshiftIdcApplicationArn = NULL,
  MaxRecords = NULL,
  Marker = NULL
)
}
\arguments{
\item{RedshiftIdcApplicationArn}{The ARN for the Redshift application that integrates with IAM Identity
Center.}

\item{MaxRecords}{The maximum number of response records to return in each call. If the
number of remaining response records exceeds the specified MaxRecords
value, a value is returned in a marker field of the response. You can
retrieve the next set of records by retrying the command with the
returned marker value.}

\item{Marker}{A value that indicates the starting point for the next set of response
records in a subsequent request. If a value is returned in a response,
you can retrieve the next set of records by providing this returned
marker value in the Marker parameter and retrying the command. If the
Marker field is empty, all response records have been retrieved for the
request.}
}
\description{
Lists the Amazon Redshift IAM Identity Center applications.

See \url{https://www.paws-r-sdk.com/docs/redshift_describe_redshift_idc_applications/} for full documentation.
}
\keyword{internal}
