% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_describe_resize}
\alias{redshift_describe_resize}
\title{Returns information about the last resize operation for the specified
cluster}
\usage{
redshift_describe_resize(ClusterIdentifier)
}
\arguments{
\item{ClusterIdentifier}{[required] The unique identifier of a cluster whose resize progress you are
requesting. This parameter is case-sensitive.

By default, resize operations for all clusters defined for an AWS
account are returned.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  TargetNodeType = "string",
  TargetNumberOfNodes = 123,
  TargetClusterType = "string",
  Status = "string",
  ImportTablesCompleted = list(
    "string"
  ),
  ImportTablesInProgress = list(
    "string"
  ),
  ImportTablesNotStarted = list(
    "string"
  ),
  AvgResizeRateInMegaBytesPerSecond = 123.0,
  TotalResizeDataInMegaBytes = 123,
  ProgressInMegaBytes = 123,
  ElapsedTimeInSeconds = 123,
  EstimatedTimeToCompletionInSeconds = 123,
  ResizeType = "string",
  Message = "string",
  TargetEncryptionType = "string",
  DataTransferProgressPercent = 123.0
)
}\if{html}{\out{</div>}}
}
\description{
Returns information about the last resize operation for the specified
cluster. If no resize operation has ever been initiated for the
specified cluster, a \verb{HTTP 404} error is returned. If a resize operation
was initiated and completed, the status of the resize remains as
\code{SUCCEEDED} until the next resize.

A resize operation can be requested using
\code{\link[=redshift_modify_cluster]{modify_cluster}} and specifying a different
number or type of nodes for the cluster.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_resize(
  ClusterIdentifier = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
