% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_describe_storage}
\alias{redshift_describe_storage}
\title{Returns account level backups storage size and provisional storage}
\usage{
redshift_describe_storage()
}
\value{
A list with the following syntax:\preformatted{list(
  TotalBackupSizeInMegaBytes = 123.0,
  TotalProvisionedStorageInMegaBytes = 123.0
)
}
}
\description{
Returns account level backups storage size and provisional storage.
}
\keyword{internal}
