% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_disassociate_data_share_consumer}
\alias{redshift_disassociate_data_share_consumer}
\title{From a datashare consumer account, remove association for the specified
datashare}
\usage{
redshift_disassociate_data_share_consumer(
  DataShareArn,
  DisassociateEntireAccount = NULL,
  ConsumerArn = NULL,
  ConsumerRegion = NULL
)
}
\arguments{
\item{DataShareArn}{[required] The Amazon Resource Name (ARN) of the datashare to remove association
for.}

\item{DisassociateEntireAccount}{A value that specifies whether association for the datashare is removed
from the entire account.}

\item{ConsumerArn}{The Amazon Resource Name (ARN) of the consumer namespace that
association for the datashare is removed from.}

\item{ConsumerRegion}{From a datashare consumer account, removes association of a datashare
from all the existing and future namespaces in the specified Amazon Web
Services Region.}
}
\description{
From a datashare consumer account, remove association for the specified datashare.

See \url{https://www.paws-r-sdk.com/docs/redshift_disassociate_data_share_consumer/} for full documentation.
}
\keyword{internal}
