% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_enable_logging}
\alias{redshift_enable_logging}
\title{Starts logging information, such as queries and connection attempts, for
the specified Amazon Redshift cluster}
\usage{
redshift_enable_logging(
  ClusterIdentifier,
  BucketName = NULL,
  S3KeyPrefix = NULL,
  LogDestinationType = NULL,
  LogExports = NULL,
  S3TableKmsKeyId = NULL,
  S3TableGranularity = NULL
)
}
\arguments{
\item{ClusterIdentifier}{[required] The identifier of the cluster on which logging is to be started.

Example: \code{examplecluster}}

\item{BucketName}{The name of an existing S3 bucket where the log files are to be stored.

Constraints:
\itemize{
\item Must be in the same region as the cluster
\item The cluster must have read bucket and put object permissions
}}

\item{S3KeyPrefix}{The prefix applied to the log file names.

Valid characters are any letter from any language, any whitespace character, any numeric character, and the following characters: underscore (\verb{_}), period (\code{.}), colon (\code{:}), slash (\code{/}), equal (\code{=}), plus (\code{+}), backslash (\verb{\\}), hyphen (\code{-}), at symbol (\code{@}).}

\item{LogDestinationType}{The log destination type. An enum with possible values of \code{s3}, \code{cloudwatch}, and \code{s3table}.}

\item{LogExports}{The collection of exported log types. When \code{LogDestinationType} is \code{s3} or \code{cloudwatch}, possible values are \code{connectionlog}, \code{useractivitylog}, and \code{userlog}. When \code{LogDestinationType} is \code{s3table}, the values are the names of the system tables to publish. Omitting this parameter, passing an empty list, or including the value \code{all} publishes all current and future system tables.}

\item{S3TableKmsKeyId}{The identifier of a customer managed KMS key used to encrypt the S3 tables. This parameter is valid only when \code{LogDestinationType} is \code{s3table}.}

\item{S3TableGranularity}{The scope of system table publishing. Valid values are \code{cluster} and \code{account}. A value of \code{cluster} scopes publishing to the individual cluster. A value of \code{account} scopes publishing to the Amazon Web Services account. This parameter is valid only when \code{LogDestinationType} is \code{s3table}.}
}
\description{
Starts logging information, such as queries and connection attempts, for the specified Amazon Redshift cluster.

See \url{https://www.paws-r-sdk.com/docs/redshift_enable_logging/} for full documentation.
}
\keyword{internal}
