% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_failover_primary_compute}
\alias{redshift_failover_primary_compute}
\title{Fails over the primary compute unit of the specified Multi-AZ cluster to
another Availability Zone}
\usage{
redshift_failover_primary_compute(ClusterIdentifier)
}
\arguments{
\item{ClusterIdentifier}{[required] The unique identifier of the cluster for which the primary compute unit
will be failed over to another Availability Zone.}
}
\description{
Fails over the primary compute unit of the specified Multi-AZ cluster to another Availability Zone.

See \url{https://www.paws-r-sdk.com/docs/redshift_failover_primary_compute/} for full documentation.
}
\keyword{internal}
