% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_get_identity_center_auth_token}
\alias{redshift_get_identity_center_auth_token}
\title{Generates an encrypted authentication token that propagates the caller's
Amazon Web Services IAM Identity Center identity to Amazon Redshift
clusters}
\usage{
redshift_get_identity_center_auth_token(ClusterIds)
}
\arguments{
\item{ClusterIds}{[required] A list of cluster identifiers that the generated token can be used with.
The token will be scoped to only allow authentication to the specified
clusters.

Constraints:
\itemize{
\item \code{ClusterIds} must contain at least 1 cluster identifier.
\item \code{ClusterIds} can hold a maximum of 20 cluster identifiers.
\item Cluster identifiers must be 1 to 63 characters in length.
\item The characters accepted for cluster identifiers are the following:
\itemize{
\item Alphanumeric characters
\item Hyphens
}
\item Cluster identifiers must start with a letter.
\item Cluster identifiers can't end with a hyphen or contain two
consecutive hyphens.
}}
}
\description{
Generates an encrypted authentication token that propagates the caller's Amazon Web Services IAM Identity Center identity to Amazon Redshift clusters. This API extracts the Amazon Web Services IAM Identity Center identity from enhanced credentials and creates a secure token that Amazon Redshift drivers can use for authentication.

See \url{https://www.paws-r-sdk.com/docs/redshift_get_identity_center_auth_token/} for full documentation.
}
\keyword{internal}
