% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_get_resource_policy}
\alias{redshift_get_resource_policy}
\title{Get the resource policy for a specified resource}
\usage{
redshift_get_resource_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource of which its resource
policy is fetched.}
}
\description{
Get the resource policy for a specified resource.

See \url{https://www.paws-r-sdk.com/docs/redshift_get_resource_policy/} for full documentation.
}
\keyword{internal}
