% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_modify_cluster_parameter_group}
\alias{redshift_modify_cluster_parameter_group}
\title{Modifies the parameters of a parameter group}
\usage{
redshift_modify_cluster_parameter_group(ParameterGroupName, Parameters)
}
\arguments{
\item{ParameterGroupName}{[required] The name of the parameter group to be modified.}

\item{Parameters}{[required] An array of parameters to be modified. A maximum of 20 parameters can be
modified in a single request.

For each parameter to be modified, you must supply at least the
parameter name and parameter value; other name-value pairs of the
parameter are optional.

For the workload management (WLM) configuration, you must supply all the
name-value pairs in the wlm_json_configuration parameter.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ParameterGroupName = "string",
  ParameterGroupStatus = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Modifies the parameters of a parameter group.

For more information about parameters and parameter groups, go to
\href{https://docs.aws.amazon.com/redshift/latest/mgmt/working-with-parameter-groups.html}{Amazon Redshift Parameter Groups}
in the \emph{Amazon Redshift Cluster Management Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$modify_cluster_parameter_group(
  ParameterGroupName = "string",
  Parameters = list(
    list(
      ParameterName = "string",
      ParameterValue = "string",
      Description = "string",
      Source = "string",
      DataType = "string",
      AllowedValues = "string",
      ApplyType = "static"|"dynamic",
      IsModifiable = TRUE|FALSE,
      MinimumEngineVersion = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
