% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_modify_custom_domain_association}
\alias{redshift_modify_custom_domain_association}
\title{Contains information for changing a custom domain association}
\usage{
redshift_modify_custom_domain_association(
  CustomDomainName,
  CustomDomainCertificateArn,
  ClusterIdentifier
)
}
\arguments{
\item{CustomDomainName}{[required] The custom domain name for a changed custom domain association.}

\item{CustomDomainCertificateArn}{[required] The certificate Amazon Resource Name (ARN) for the changed custom domain
association.}

\item{ClusterIdentifier}{[required] The identifier of the cluster to change a custom domain association for.}
}
\description{
Contains information for changing a custom domain association.

See \url{https://www.paws-r-sdk.com/docs/redshift_modify_custom_domain_association/} for full documentation.
}
\keyword{internal}
