% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_modify_integration}
\alias{redshift_modify_integration}
\title{Modifies a zero-ETL integration or S3 event integration with Amazon
Redshift}
\usage{
redshift_modify_integration(
  IntegrationArn,
  Description = NULL,
  IntegrationName = NULL
)
}
\arguments{
\item{IntegrationArn}{[required] The unique identifier of the integration to modify.}

\item{Description}{A new description for the integration.}

\item{IntegrationName}{A new name for the integration.}
}
\description{
Modifies a zero-ETL integration or S3 event integration with Amazon Redshift.

See \url{https://www.paws-r-sdk.com/docs/redshift_modify_integration/} for full documentation.
}
\keyword{internal}
