% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_modify_qev_2_idc_application}
\alias{redshift_modify_qev_2_idc_application}
\title{Modifies an Amazon Redshift Query Editor (QEV2) IAM Identity Center
application}
\usage{
redshift_modify_qev_2_idc_application(
  Qev2IdcApplicationArn,
  IdcDisplayName = NULL
)
}
\arguments{
\item{Qev2IdcApplicationArn}{[required] The Amazon Resource Name (ARN) for the Amazon Redshift Query Editor (QEV2) application that integrates with IAM Identity Center.}

\item{IdcDisplayName}{The display name for the Amazon Redshift Query Editor (QEV2) IAM Identity Center application. It appears in the console.}
}
\description{
Modifies an Amazon Redshift Query Editor (QEV2) IAM Identity Center application.

See \url{https://www.paws-r-sdk.com/docs/redshift_modify_qev_2_idc_application/} for full documentation.
}
\keyword{internal}
