% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_put_resource_policy}
\alias{redshift_put_resource_policy}
\title{Updates the resource policy for a specified resource}
\usage{
redshift_put_resource_policy(ResourceArn, Policy)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource of which its resource
policy is updated.}

\item{Policy}{[required] The content of the resource policy being updated.}
}
\description{
Updates the resource policy for a specified resource.

See \url{https://www.paws-r-sdk.com/docs/redshift_put_resource_policy/} for full documentation.
}
\keyword{internal}
