% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_register_namespace}
\alias{redshift_register_namespace}
\title{Registers a cluster or serverless namespace to the Amazon Web Services
Glue Data Catalog}
\usage{
redshift_register_namespace(NamespaceIdentifier, ConsumerIdentifiers)
}
\arguments{
\item{NamespaceIdentifier}{[required] The unique identifier of the cluster or serverless namespace that you
want to register.}

\item{ConsumerIdentifiers}{[required] An array containing the ID of the consumer account that you want to
register the namespace to.}
}
\description{
Registers a cluster or serverless namespace to the Amazon Web Services Glue Data Catalog.

See \url{https://www.paws-r-sdk.com/docs/redshift_register_namespace/} for full documentation.
}
\keyword{internal}
