% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_reject_data_share}
\alias{redshift_reject_data_share}
\title{From a datashare consumer account, rejects the specified datashare}
\usage{
redshift_reject_data_share(DataShareArn)
}
\arguments{
\item{DataShareArn}{[required] The Amazon Resource Name (ARN) of the datashare to reject.}
}
\description{
From a datashare consumer account, rejects the specified datashare.

See \url{https://paws-r.github.io/docs/redshift/reject_data_share.html} for full documentation.
}
\keyword{internal}
