% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_resume_cluster}
\alias{redshift_resume_cluster}
\title{Resumes a paused cluster}
\usage{
redshift_resume_cluster(ClusterIdentifier)
}
\arguments{
\item{ClusterIdentifier}{[required] The identifier of the cluster to be resumed.}
}
\description{
Resumes a paused cluster.

See \url{https://paws-r.github.io/docs/redshift/resume_cluster.html} for full documentation.
}
\keyword{internal}
