% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_revoke_endpoint_access}
\alias{redshift_revoke_endpoint_access}
\title{Revokes access to a cluster}
\usage{
redshift_revoke_endpoint_access(
  ClusterIdentifier = NULL,
  Account = NULL,
  VpcIds = NULL,
  Force = NULL
)
}
\arguments{
\item{ClusterIdentifier}{The cluster to revoke access from.}

\item{Account}{The Amazon Web Services account ID whose access is to be revoked.}

\item{VpcIds}{The virtual private cloud (VPC) identifiers for which access is to be
revoked.}

\item{Force}{Indicates whether to force the revoke action. If true, the
Redshift-managed VPC endpoints associated with the endpoint
authorization are also deleted.}
}
\description{
Revokes access to a cluster.

See \url{https://www.paws-r-sdk.com/docs/redshift_revoke_endpoint_access/} for full documentation.
}
\keyword{internal}
