% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftdataapiservice_operations.R
\name{redshiftdataapiservice_cancel_statement}
\alias{redshiftdataapiservice_cancel_statement}
\title{Cancels a running query}
\usage{
redshiftdataapiservice_cancel_statement(Id)
}
\arguments{
\item{Id}{[required] The identifier of the SQL statement to cancel. This value is a
universally unique identifier (UUID) generated by Amazon Redshift Data
API. This identifier is returned by \code{BatchExecuteStatment},
\code{ExecuteStatment}, and
\code{\link[=redshiftdataapiservice_list_statements]{list_statements}}.}
}
\description{
Cancels a running query. To be canceled, a query must be running.

See \url{https://www.paws-r-sdk.com/docs/redshiftdataapiservice_cancel_statement/} for full documentation.
}
\keyword{internal}
