% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftdataapiservice_operations.R
\name{redshiftdataapiservice_list_sessions}
\alias{redshiftdataapiservice_list_sessions}
\title{Lists the sessions that the caller created in the last 24 hours}
\usage{
redshiftdataapiservice_list_sessions(
  NextToken = NULL,
  MaxResults = NULL,
  SessionId = NULL,
  Status = NULL,
  RoleLevel = NULL,
  ClusterIdentifier = NULL,
  WorkgroupName = NULL,
  Database = NULL
)
}
\arguments{
\item{NextToken}{A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned NextToken value in the next NextToken parameter and retrying the command. If the NextToken field is empty, all response records have been retrieved for the request.}

\item{MaxResults}{The maximum number of sessions to return in the response. If more sessions exist than fit in one response, the operation returns \code{NextToken} to paginate the results.}

\item{SessionId}{The identifier of a specific session to return metadata for. This value is a universally unique identifier (UUID) generated by Amazon Redshift Data API. When you provide \code{SessionId}, you can't specify \code{Status}, \code{ClusterIdentifier}, \code{WorkgroupName}, or \code{Database}.}

\item{Status}{The status of the sessions to list. If no status is specified, sessions with a status of \code{AVAILABLE} or \code{BUSY} are returned. Status values are defined as follows:
\itemize{
\item AVAILABLE – The session is open and ready to run a SQL statement.
\item BUSY – The session is currently running a SQL statement.
\item CLOSED – The session is closed and can no longer run SQL statements.
}}

\item{RoleLevel}{Specifies whether to return all sessions created by the caller's IAM role, including sessions from previous IAM sessions. If false, only sessions created in the current IAM session are returned. The default is true.}

\item{ClusterIdentifier}{The cluster identifier. Only sessions on this cluster are returned. When providing \code{ClusterIdentifier}, then \code{WorkgroupName} can't be specified.}

\item{WorkgroupName}{The serverless workgroup name or Amazon Resource Name (ARN). Only sessions on this workgroup are returned. When providing \code{WorkgroupName}, then \code{ClusterIdentifier} can't be specified.}

\item{Database}{The name of the database. Only sessions connected to this database are returned.}
}
\description{
Lists the sessions that the caller created in the last 24 hours. By default, only sessions with a status of \code{AVAILABLE} or \code{BUSY} are returned. You can filter the results by session status, compute target (cluster or serverless workgroup), or database. To retrieve the metadata for a single session, provide the \code{SessionId} parameter. Use \code{NextToken} to page through the session list.

See \url{https://www.paws-r-sdk.com/docs/redshiftdataapiservice_list_sessions/} for full documentation.
}
\keyword{internal}
