% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_create_custom_domain_association}
\alias{redshiftserverless_create_custom_domain_association}
\title{Creates a custom domain association for Amazon Redshift Serverless}
\usage{
redshiftserverless_create_custom_domain_association(
  customDomainCertificateArn,
  customDomainName,
  workgroupName
)
}
\arguments{
\item{customDomainCertificateArn}{[required] The custom domain name’s certificate Amazon resource name (ARN).}

\item{customDomainName}{[required] The custom domain name to associate with the workgroup.}

\item{workgroupName}{[required] The name of the workgroup associated with the database.}
}
\description{
Creates a custom domain association for Amazon Redshift Serverless.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_create_custom_domain_association/} for full documentation.
}
\keyword{internal}
