% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_create_snapshot}
\alias{redshiftserverless_create_snapshot}
\title{Creates a snapshot of all databases in a namespace}
\usage{
redshiftserverless_create_snapshot(
  namespaceName,
  retentionPeriod = NULL,
  snapshotName
)
}
\arguments{
\item{namespaceName}{[required] The namespace to create a snapshot for.}

\item{retentionPeriod}{How long to retain the created snapshot.}

\item{snapshotName}{[required] The name of the snapshot.}
}
\description{
Creates a snapshot of all databases in a namespace. For more information about snapshots, see \href{https://docs.aws.amazon.com/redshift/latest/mgmt/serverless-snapshots-recovery.html}{Working with snapshots and recovery points}.

See \url{https://paws-r.github.io/docs/redshiftserverless/create_snapshot.html} for full documentation.
}
\keyword{internal}
