% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_create_snapshot_copy_configuration}
\alias{redshiftserverless_create_snapshot_copy_configuration}
\title{Creates a snapshot copy configuration that lets you copy snapshots to
another Amazon Web Services Region}
\usage{
redshiftserverless_create_snapshot_copy_configuration(
  destinationKmsKeyId = NULL,
  destinationRegion,
  namespaceName,
  snapshotRetentionPeriod = NULL
)
}
\arguments{
\item{destinationKmsKeyId}{The KMS key to use to encrypt your snapshots in the destination Amazon
Web Services Region.}

\item{destinationRegion}{[required] The destination Amazon Web Services Region that you want to copy
snapshots to.}

\item{namespaceName}{[required] The name of the namespace to copy snapshots from.}

\item{snapshotRetentionPeriod}{The retention period of the snapshots that you copy to the destination
Amazon Web Services Region.}
}
\description{
Creates a snapshot copy configuration that lets you copy snapshots to another Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_create_snapshot_copy_configuration/} for full documentation.
}
\keyword{internal}
