% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_delete_custom_domain_association}
\alias{redshiftserverless_delete_custom_domain_association}
\title{Deletes a custom domain association for Amazon Redshift Serverless}
\usage{
redshiftserverless_delete_custom_domain_association(
  customDomainName,
  workgroupName
)
}
\arguments{
\item{customDomainName}{[required] The custom domain name associated with the workgroup.}

\item{workgroupName}{[required] The name of the workgroup associated with the database.}
}
\description{
Deletes a custom domain association for Amazon Redshift Serverless.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_delete_custom_domain_association/} for full documentation.
}
\keyword{internal}
