% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_delete_namespace}
\alias{redshiftserverless_delete_namespace}
\title{Deletes a namespace from Amazon Redshift Serverless}
\usage{
redshiftserverless_delete_namespace(
  finalSnapshotName = NULL,
  finalSnapshotRetentionPeriod = NULL,
  namespaceName
)
}
\arguments{
\item{finalSnapshotName}{The name of the snapshot to be created before the namespace is deleted.}

\item{finalSnapshotRetentionPeriod}{How long to retain the final snapshot.}

\item{namespaceName}{[required] The name of the namespace to delete.}
}
\description{
Deletes a namespace from Amazon Redshift Serverless. Before you delete the namespace, you can create a final snapshot that has all of the data within the namespace.

See \url{https://paws-r.github.io/docs/redshiftserverless/delete_namespace.html} for full documentation.
}
\keyword{internal}
