% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_delete_resource_policy}
\alias{redshiftserverless_delete_resource_policy}
\title{Deletes the specified resource policy}
\usage{
redshiftserverless_delete_resource_policy(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the policy to delete.}
}
\description{
Deletes the specified resource policy.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_delete_resource_policy/} for full documentation.
}
\keyword{internal}
