% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_delete_snapshot}
\alias{redshiftserverless_delete_snapshot}
\title{Deletes a snapshot from Amazon Redshift Serverless}
\usage{
redshiftserverless_delete_snapshot(snapshotName)
}
\arguments{
\item{snapshotName}{[required] The name of the snapshot to be deleted.}
}
\description{
Deletes a snapshot from Amazon Redshift Serverless.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_delete_snapshot/} for full documentation.
}
\keyword{internal}
