% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_delete_snapshot_copy_configuration}
\alias{redshiftserverless_delete_snapshot_copy_configuration}
\title{Deletes a snapshot copy configuration}
\usage{
redshiftserverless_delete_snapshot_copy_configuration(
  snapshotCopyConfigurationId
)
}
\arguments{
\item{snapshotCopyConfigurationId}{[required] The ID of the snapshot copy configuration to delete.}
}
\description{
Deletes a snapshot copy configuration

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_delete_snapshot_copy_configuration/} for full documentation.
}
\keyword{internal}
