% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_delete_usage_limit}
\alias{redshiftserverless_delete_usage_limit}
\title{Deletes a usage limit from Amazon Redshift Serverless}
\usage{
redshiftserverless_delete_usage_limit(usageLimitId)
}
\arguments{
\item{usageLimitId}{[required] The unique identifier of the usage limit to delete.}
}
\description{
Deletes a usage limit from Amazon Redshift Serverless.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_delete_usage_limit/} for full documentation.
}
\keyword{internal}
