% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_delete_workgroup}
\alias{redshiftserverless_delete_workgroup}
\title{Deletes a workgroup}
\usage{
redshiftserverless_delete_workgroup(workgroupName)
}
\arguments{
\item{workgroupName}{[required] The name of the workgroup to be deleted.}
}
\description{
Deletes a workgroup.

See \url{https://paws-r.github.io/docs/redshiftserverless/delete_workgroup.html} for full documentation.
}
\keyword{internal}
