% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_get_custom_domain_association}
\alias{redshiftserverless_get_custom_domain_association}
\title{Gets information about a specific custom domain association}
\usage{
redshiftserverless_get_custom_domain_association(
  customDomainName,
  workgroupName
)
}
\arguments{
\item{customDomainName}{[required] The custom domain name associated with the workgroup.}

\item{workgroupName}{[required] The name of the workgroup associated with the database.}
}
\description{
Gets information about a specific custom domain association.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_get_custom_domain_association/} for full documentation.
}
\keyword{internal}
