% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_get_endpoint_access}
\alias{redshiftserverless_get_endpoint_access}
\title{Returns information, such as the name, about a VPC endpoint}
\usage{
redshiftserverless_get_endpoint_access(endpointName)
}
\arguments{
\item{endpointName}{[required] The name of the VPC endpoint to return information for.}
}
\description{
Returns information, such as the name, about a VPC endpoint.

See \url{https://paws-r.github.io/docs/redshiftserverless/get_endpoint_access.html} for full documentation.
}
\keyword{internal}
