% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_get_identity_center_auth_token}
\alias{redshiftserverless_get_identity_center_auth_token}
\title{Returns an Identity Center authentication token for accessing Amazon
Redshift Serverless workgroups}
\usage{
redshiftserverless_get_identity_center_auth_token(workgroupNames)
}
\arguments{
\item{workgroupNames}{[required] A list of workgroup names for which to generate the Identity Center
authentication token.

Constraints:
\itemize{
\item Must contain between 1 and 20 workgroup names.
\item Each workgroup name must be a valid Amazon Redshift Serverless
workgroup identifier.
\item All specified workgroups must have Identity Center integration
enabled.
}}
}
\description{
Returns an Identity Center authentication token for accessing Amazon Redshift Serverless workgroups.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_get_identity_center_auth_token/} for full documentation.
}
\keyword{internal}
