% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_get_namespace}
\alias{redshiftserverless_get_namespace}
\title{Returns information about a namespace in Amazon Redshift Serverless}
\usage{
redshiftserverless_get_namespace(namespaceName)
}
\arguments{
\item{namespaceName}{[required] The name of the namespace to retrieve information for.}
}
\description{
Returns information about a namespace in Amazon Redshift Serverless.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_get_namespace/} for full documentation.
}
\keyword{internal}
