% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_get_reservation}
\alias{redshiftserverless_get_reservation}
\title{Gets an Amazon Redshift Serverless reservation}
\usage{
redshiftserverless_get_reservation(reservationId)
}
\arguments{
\item{reservationId}{[required] The ID of the reservation to retrieve.}
}
\description{
Gets an Amazon Redshift Serverless reservation. A reservation gives you the option to commit to a specified number of Redshift Processing Units (RPUs) for a year at a discount from Serverless on-demand (OD) rates.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_get_reservation/} for full documentation.
}
\keyword{internal}
