% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_get_reservation_offering}
\alias{redshiftserverless_get_reservation_offering}
\title{Returns the reservation offering}
\usage{
redshiftserverless_get_reservation_offering(offeringId)
}
\arguments{
\item{offeringId}{[required] The identifier for the offering..}
}
\description{
Returns the reservation offering. The offering determines the payment schedule for the reservation.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_get_reservation_offering/} for full documentation.
}
\keyword{internal}
