% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_get_resource_policy}
\alias{redshiftserverless_get_resource_policy}
\title{Returns a resource policy}
\usage{
redshiftserverless_get_resource_policy(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource to return.}
}
\description{
Returns a resource policy.

See \url{https://paws-r.github.io/docs/redshiftserverless/get_resource_policy.html} for full documentation.
}
\keyword{internal}
