% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_get_scheduled_action}
\alias{redshiftserverless_get_scheduled_action}
\title{Returns information about a scheduled action}
\usage{
redshiftserverless_get_scheduled_action(scheduledActionName)
}
\arguments{
\item{scheduledActionName}{[required] The name of the scheduled action.}
}
\description{
Returns information about a scheduled action.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_get_scheduled_action/} for full documentation.
}
\keyword{internal}
