% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_get_snapshot}
\alias{redshiftserverless_get_snapshot}
\title{Returns information about a specific snapshot}
\usage{
redshiftserverless_get_snapshot(
  ownerAccount = NULL,
  snapshotArn = NULL,
  snapshotName = NULL
)
}
\arguments{
\item{ownerAccount}{The owner Amazon Web Services account of a snapshot shared with another
user.}

\item{snapshotArn}{The Amazon Resource Name (ARN) of the snapshot to return.}

\item{snapshotName}{The name of the snapshot to return.}
}
\description{
Returns information about a specific snapshot.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_get_snapshot/} for full documentation.
}
\keyword{internal}
