% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_get_table_restore_status}
\alias{redshiftserverless_get_table_restore_status}
\title{Returns information about a TableRestoreStatus object}
\usage{
redshiftserverless_get_table_restore_status(tableRestoreRequestId)
}
\arguments{
\item{tableRestoreRequestId}{[required] The ID of the
\code{\link[=redshiftserverless_restore_table_from_snapshot]{restore_table_from_snapshot}}
request to return status for.}
}
\description{
Returns information about a \code{TableRestoreStatus} object.

See \url{https://paws-r.github.io/docs/redshiftserverless/get_table_restore_status.html} for full documentation.
}
\keyword{internal}
