% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_get_track}
\alias{redshiftserverless_get_track}
\title{Get the Redshift Serverless version for a specified track}
\usage{
redshiftserverless_get_track(trackName)
}
\arguments{
\item{trackName}{[required] The name of the track of which its version is fetched.}
}
\description{
Get the Redshift Serverless version for a specified track.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_get_track/} for full documentation.
}
\keyword{internal}
