% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_get_usage_limit}
\alias{redshiftserverless_get_usage_limit}
\title{Returns information about a usage limit}
\usage{
redshiftserverless_get_usage_limit(usageLimitId)
}
\arguments{
\item{usageLimitId}{[required] The unique identifier of the usage limit to return information for.}
}
\description{
Returns information about a usage limit.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_get_usage_limit/} for full documentation.
}
\keyword{internal}
