% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_get_workgroup}
\alias{redshiftserverless_get_workgroup}
\title{Returns information about a specific workgroup}
\usage{
redshiftserverless_get_workgroup(workgroupName)
}
\arguments{
\item{workgroupName}{[required] The name of the workgroup to return information for.}
}
\description{
Returns information about a specific workgroup.

See \url{https://paws-r.github.io/docs/redshiftserverless/get_workgroup.html} for full documentation.
}
\keyword{internal}
