% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_list_custom_domain_associations}
\alias{redshiftserverless_list_custom_domain_associations}
\title{Lists custom domain associations for Amazon Redshift Serverless}
\usage{
redshiftserverless_list_custom_domain_associations(
  customDomainCertificateArn = NULL,
  customDomainName = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{customDomainCertificateArn}{The custom domain name’s certificate Amazon resource name (ARN).}

\item{customDomainName}{The custom domain name associated with the workgroup.}

\item{maxResults}{An optional parameter that specifies the maximum number of results to
return. You can use \code{nextToken} to display the next page of results.}

\item{nextToken}{When \code{nextToken} is returned, there are more results available. The
value of \code{nextToken} is a unique pagination token for each page. Make
the call again using the returned token to retrieve the next page.}
}
\description{
Lists custom domain associations for Amazon Redshift Serverless.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_list_custom_domain_associations/} for full documentation.
}
\keyword{internal}
