% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_list_endpoint_access}
\alias{redshiftserverless_list_endpoint_access}
\title{Returns an array of EndpointAccess objects and relevant information}
\usage{
redshiftserverless_list_endpoint_access(
  maxResults = NULL,
  nextToken = NULL,
  vpcId = NULL,
  workgroupName = NULL
)
}
\arguments{
\item{maxResults}{An optional parameter that specifies the maximum number of results to
return. You can use \code{nextToken} to display the next page of results.}

\item{nextToken}{If your initial
\code{\link[=redshiftserverless_list_endpoint_access]{list_endpoint_access}}
operation returns a \code{nextToken}, you can include the returned
\code{nextToken} in following
\code{\link[=redshiftserverless_list_endpoint_access]{list_endpoint_access}}
operations, which returns results in the next page.}

\item{vpcId}{The unique identifier of the virtual private cloud with access to Amazon
Redshift Serverless.}

\item{workgroupName}{The name of the workgroup associated with the VPC endpoint to return.}
}
\description{
Returns an array of \code{EndpointAccess} objects and relevant information.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_list_endpoint_access/} for full documentation.
}
\keyword{internal}
