% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_list_managed_workgroups}
\alias{redshiftserverless_list_managed_workgroups}
\title{Returns information about a list of specified managed workgroups in your
account}
\usage{
redshiftserverless_list_managed_workgroups(
  maxResults = NULL,
  nextToken = NULL,
  sourceArn = NULL
)
}
\arguments{
\item{maxResults}{An optional parameter that specifies the maximum number of results to
return. You can use nextToken to display the next page of results.}

\item{nextToken}{If your initial ListManagedWorkgroups operation returns a nextToken, you
can include the returned nextToken in following ListManagedWorkgroups
operations, which returns results in the next page.}

\item{sourceArn}{The Amazon Resource Name (ARN) for the managed workgroup in the Glue
Data Catalog.}
}
\description{
Returns information about a list of specified managed workgroups in your account.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_list_managed_workgroups/} for full documentation.
}
\keyword{internal}
