% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_list_reservation_offerings}
\alias{redshiftserverless_list_reservation_offerings}
\title{Returns the current reservation offerings in your account}
\usage{
redshiftserverless_list_reservation_offerings(
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}

\item{nextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\description{
Returns the current reservation offerings in your account.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_list_reservation_offerings/} for full documentation.
}
\keyword{internal}
