% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_list_scheduled_actions}
\alias{redshiftserverless_list_scheduled_actions}
\title{Returns a list of scheduled actions}
\usage{
redshiftserverless_list_scheduled_actions(
  maxResults = NULL,
  namespaceName = NULL,
  nextToken = NULL
)
}
\arguments{
\item{maxResults}{An optional parameter that specifies the maximum number of results to
return. Use \code{nextToken} to display the next page of results.}

\item{namespaceName}{The name of namespace associated with the scheduled action to retrieve.}

\item{nextToken}{If \code{nextToken} is returned, there are more results available. The value
of \code{nextToken} is a unique pagination token for each page. Make the call
again using the returned token to retrieve the next page.}
}
\description{
Returns a list of scheduled actions. You can use the flags to filter the list of returned scheduled actions.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_list_scheduled_actions/} for full documentation.
}
\keyword{internal}
