% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_list_snapshot_copy_configurations}
\alias{redshiftserverless_list_snapshot_copy_configurations}
\title{Returns a list of snapshot copy configurations}
\usage{
redshiftserverless_list_snapshot_copy_configurations(
  maxResults = NULL,
  namespaceName = NULL,
  nextToken = NULL
)
}
\arguments{
\item{maxResults}{An optional parameter that specifies the maximum number of results to
return. You can use \code{nextToken} to display the next page of results.}

\item{namespaceName}{The namespace from which to list all snapshot copy configurations.}

\item{nextToken}{If \code{nextToken} is returned, there are more results available. The value
of \code{nextToken} is a unique pagination token for each page. Make the call
again using the returned token to retrieve the next page.}
}
\description{
Returns a list of snapshot copy configurations.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_list_snapshot_copy_configurations/} for full documentation.
}
\keyword{internal}
