% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_list_table_restore_status}
\alias{redshiftserverless_list_table_restore_status}
\title{Returns information about an array of TableRestoreStatus objects}
\usage{
redshiftserverless_list_table_restore_status(
  maxResults = NULL,
  namespaceName = NULL,
  nextToken = NULL,
  workgroupName = NULL
)
}
\arguments{
\item{maxResults}{An optional parameter that specifies the maximum number of results to
return. You can use nextToken to display the next page of results.}

\item{namespaceName}{The namespace from which to list all of the statuses of
\code{\link[=redshiftserverless_restore_table_from_snapshot]{restore_table_from_snapshot}}
operations .}

\item{nextToken}{If your initial
\code{\link[=redshiftserverless_list_table_restore_status]{list_table_restore_status}}
operation returns a nextToken, you can include the returned \code{nextToken}
in following
\code{\link[=redshiftserverless_list_table_restore_status]{list_table_restore_status}}
operations. This will return results on the next page.}

\item{workgroupName}{The workgroup from which to list all of the statuses of
\code{\link[=redshiftserverless_restore_table_from_snapshot]{restore_table_from_snapshot}}
operations.}
}
\description{
Returns information about an array of \code{TableRestoreStatus} objects.

See \url{https://paws-r.github.io/docs/redshiftserverless/list_table_restore_status.html} for full documentation.
}
\keyword{internal}
