% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_list_tracks}
\alias{redshiftserverless_list_tracks}
\title{List the Amazon Redshift Serverless versions}
\usage{
redshiftserverless_list_tracks(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of response records to return in each call. If the
number of remaining response records exceeds the specified MaxRecords
value, a value is returned in a marker field of the response. You can
retrieve the next set of records by retrying the command with the
returned marker value.}

\item{nextToken}{If your initial \code{ListTracksRequest} operation returns a \code{nextToken}, you
can include the returned \code{nextToken} in following \code{ListTracksRequest}
operations, which returns results in the next page.}
}
\description{
List the Amazon Redshift Serverless versions.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_list_tracks/} for full documentation.
}
\keyword{internal}
