% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_list_usage_limits}
\alias{redshiftserverless_list_usage_limits}
\title{Lists all usage limits within Amazon Redshift Serverless}
\usage{
redshiftserverless_list_usage_limits(
  maxResults = NULL,
  nextToken = NULL,
  resourceArn = NULL,
  usageType = NULL
)
}
\arguments{
\item{maxResults}{An optional parameter that specifies the maximum number of results to
return. You can use \code{nextToken} to get the next page of results. The
default is 100.}

\item{nextToken}{If your initial
\code{\link[=redshiftserverless_list_usage_limits]{list_usage_limits}} operation
returns a \code{nextToken}, you can include the returned \code{nextToken} in
following \code{\link[=redshiftserverless_list_usage_limits]{list_usage_limits}}
operations, which returns results in the next page.}

\item{resourceArn}{The Amazon Resource Name (ARN) associated with the resource whose usage
limits you want to list.}

\item{usageType}{The Amazon Redshift Serverless feature whose limits you want to see.}
}
\description{
Lists all usage limits within Amazon Redshift Serverless.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_list_usage_limits/} for full documentation.
}
\keyword{internal}
