% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_list_workgroups}
\alias{redshiftserverless_list_workgroups}
\title{Returns information about a list of specified workgroups}
\usage{
redshiftserverless_list_workgroups(
  maxResults = NULL,
  nextToken = NULL,
  ownerAccount = NULL
)
}
\arguments{
\item{maxResults}{An optional parameter that specifies the maximum number of results to
return. You can use \code{nextToken} to display the next page of results.}

\item{nextToken}{If your initial ListWorkgroups operation returns a \code{nextToken}, you can
include the returned \code{nextToken} in following ListNamespaces operations,
which returns results in the next page.}

\item{ownerAccount}{The owner Amazon Web Services account for the Amazon Redshift Serverless
workgroup.}
}
\description{
Returns information about a list of specified workgroups.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_list_workgroups/} for full documentation.
}
\keyword{internal}
