% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_restore_from_recovery_point}
\alias{redshiftserverless_restore_from_recovery_point}
\title{Restore the data from a recovery point}
\usage{
redshiftserverless_restore_from_recovery_point(
  maintainIntegration = NULL,
  namespaceName,
  recoveryPointId,
  workgroupName
)
}
\arguments{
\item{maintainIntegration}{If \code{true}, maintain existing data sharing, zero-ETL and S3 event integrations when restoring. Otherwise, integrations will not be maintained after the restore operation. Integrations are only maintained when restored to the same serverless namespace.

Default: true}

\item{namespaceName}{[required] The name of the namespace to restore data into.}

\item{recoveryPointId}{[required] The unique identifier of the recovery point to restore from.}

\item{workgroupName}{[required] The name of the workgroup used to restore data.}
}
\description{
Restore the data from a recovery point.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_restore_from_recovery_point/} for full documentation.
}
\keyword{internal}
