% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_tag_resource}
\alias{redshiftserverless_tag_resource}
\title{Assigns one or more tags to a resource}
\usage{
redshiftserverless_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource to tag.}

\item{tags}{[required] The map of the key-value pairs used to tag the resource.}
}
\description{
Assigns one or more tags to a resource.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_tag_resource/} for full documentation.
}
\keyword{internal}
