% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_untag_resource}
\alias{redshiftserverless_untag_resource}
\title{Removes a tag or set of tags from a resource}
\usage{
redshiftserverless_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource to remove tags from.}

\item{tagKeys}{[required] The tag or set of tags to remove from the resource.}
}
\description{
Removes a tag or set of tags from a resource.

See \url{https://paws-r.github.io/docs/redshiftserverless/untag_resource.html} for full documentation.
}
\keyword{internal}
