% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_update_custom_domain_association}
\alias{redshiftserverless_update_custom_domain_association}
\title{Updates an Amazon Redshift Serverless certificate associated with a
custom domain}
\usage{
redshiftserverless_update_custom_domain_association(
  customDomainCertificateArn,
  customDomainName,
  workgroupName
)
}
\arguments{
\item{customDomainCertificateArn}{[required] The custom domain name’s certificate Amazon resource name (ARN). This is
optional.}

\item{customDomainName}{[required] The custom domain name associated with the workgroup.}

\item{workgroupName}{[required] The name of the workgroup associated with the database.}
}
\description{
Updates an Amazon Redshift Serverless certificate associated with a custom domain.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_update_custom_domain_association/} for full documentation.
}
\keyword{internal}
