% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_update_endpoint_access}
\alias{redshiftserverless_update_endpoint_access}
\title{Updates an Amazon Redshift Serverless managed endpoint}
\usage{
redshiftserverless_update_endpoint_access(
  endpointName,
  vpcSecurityGroupIds = NULL
)
}
\arguments{
\item{endpointName}{[required] The name of the VPC endpoint to update.}

\item{vpcSecurityGroupIds}{The list of VPC security groups associated with the endpoint after the
endpoint is modified.}
}
\description{
Updates an Amazon Redshift Serverless managed endpoint.

See \url{https://paws-r.github.io/docs/redshiftserverless/update_endpoint_access.html} for full documentation.
}
\keyword{internal}
