% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_update_snapshot_copy_configuration}
\alias{redshiftserverless_update_snapshot_copy_configuration}
\title{Updates a snapshot copy configuration}
\usage{
redshiftserverless_update_snapshot_copy_configuration(
  snapshotCopyConfigurationId,
  snapshotRetentionPeriod = NULL
)
}
\arguments{
\item{snapshotCopyConfigurationId}{[required] The ID of the snapshot copy configuration to update.}

\item{snapshotRetentionPeriod}{The new retention period of how long to keep a snapshot in the
destination Amazon Web Services Region.}
}
\description{
Updates a snapshot copy configuration.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_update_snapshot_copy_configuration/} for full documentation.
}
\keyword{internal}
