% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpledb_operations.R
\name{simpledb_batch_delete_attributes}
\alias{simpledb_batch_delete_attributes}
\title{Performs multiple DeleteAttributes operations in a single call, which
reduces round trips and latencies}
\usage{
simpledb_batch_delete_attributes(DomainName, Items)
}
\arguments{
\item{DomainName}{[required] The name of the domain in which the attributes are being deleted.}

\item{Items}{[required] A list of items on which to perform the operation.}
}
\description{
Performs multiple DeleteAttributes operations in a single call, which reduces round trips and latencies. This enables Amazon SimpleDB to optimize requests, which generally yields better throughput.

See \url{https://paws-r.github.io/docs/simpledb/batch_delete_attributes.html} for full documentation.
}
\keyword{internal}
