% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpledb_operations.R
\name{simpledb_create_domain}
\alias{simpledb_create_domain}
\title{The CreateDomain operation creates a new domain}
\usage{
simpledb_create_domain(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the domain to create. The name can range between 3 and 255
characters and can contain the following characters: a-z, A-Z, 0-9,
'_', '-', and '.'.}
}
\value{
An empty list.
}
\description{
The \code{\link[=simpledb_create_domain]{create_domain}} operation creates a new
domain. The domain name should be unique among the domains associated
with the Access Key ID provided in the request. The
\code{\link[=simpledb_create_domain]{create_domain}} operation may take 10 or more
seconds to complete.

CreateDomain is an idempotent operation; running it multiple times using
the same domain name will not result in an error response.

The client can create up to 100 domains per account.

If the client requires additional domains, go to
http://aws.amazon.com/contact-us/simpledb-limit-request/.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_domain(
  DomainName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
