% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpledb_operations.R
\name{simpledb_list_domains}
\alias{simpledb_list_domains}
\title{The ListDomains operation lists all domains associated with the Access
Key ID}
\usage{
simpledb_list_domains(MaxNumberOfDomains, NextToken)
}
\arguments{
\item{MaxNumberOfDomains}{The maximum number of domain names you want returned. The range is 1 to
100. The default setting is 100.}

\item{NextToken}{A string informing Amazon SimpleDB where to start the next list of
domain names.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DomainNames = list(
    "string"
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
The \code{\link[=simpledb_list_domains]{list_domains}} operation lists all domains
associated with the Access Key ID. It returns domain names up to the
limit set by MaxNumberOfDomains. A NextToken is returned if there are
more than \code{MaxNumberOfDomains} domains. Calling
\code{\link[=simpledb_list_domains]{list_domains}} successive times with the
\code{NextToken} provided by the operation returns up to \code{MaxNumberOfDomains}
more domain names with each successive operation call.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_domains(
  MaxNumberOfDomains = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
