% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamquery_operations.R
\name{timestreamquery_delete_scheduled_query}
\alias{timestreamquery_delete_scheduled_query}
\title{Deletes a given scheduled query}
\usage{
timestreamquery_delete_scheduled_query(ScheduledQueryArn)
}
\arguments{
\item{ScheduledQueryArn}{[required] The ARN of the scheduled query.}
}
\description{
Deletes a given scheduled query. This is an irreversible operation.

See \url{https://www.paws-r-sdk.com/docs/timestreamquery_delete_scheduled_query/} for full documentation.
}
\keyword{internal}
