% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamquery_operations.R
\name{timestreamquery_describe_account_settings}
\alias{timestreamquery_describe_account_settings}
\title{Describes the settings for your account that include the query pricing
model and the configured maximum TCUs the service can use for your query
workload}
\usage{
timestreamquery_describe_account_settings()
}
\description{
Describes the settings for your account that include the query pricing model and the configured maximum TCUs the service can use for your query workload.

See \url{https://www.paws-r-sdk.com/docs/timestreamquery_describe_account_settings/} for full documentation.
}
\keyword{internal}
