% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamquery_operations.R
\name{timestreamquery_describe_endpoints}
\alias{timestreamquery_describe_endpoints}
\title{DescribeEndpoints returns a list of available endpoints to make
Timestream API calls against}
\usage{
timestreamquery_describe_endpoints()
}
\description{
DescribeEndpoints returns a list of available endpoints to make Timestream API calls against. This API is available through both Write and Query.

See \url{https://www.paws-r-sdk.com/docs/timestreamquery_describe_endpoints/} for full documentation.
}
\keyword{internal}
