% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamquery_operations.R
\name{timestreamquery_describe_scheduled_query}
\alias{timestreamquery_describe_scheduled_query}
\title{Provides detailed information about a scheduled query}
\usage{
timestreamquery_describe_scheduled_query(ScheduledQueryArn)
}
\arguments{
\item{ScheduledQueryArn}{[required] The ARN of the scheduled query.}
}
\description{
Provides detailed information about a scheduled query.

See \url{https://www.paws-r-sdk.com/docs/timestreamquery_describe_scheduled_query/} for full documentation.
}
\keyword{internal}
