% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamquery_operations.R
\name{timestreamquery_list_scheduled_queries}
\alias{timestreamquery_list_scheduled_queries}
\title{Gets a list of all scheduled queries in the caller's Amazon account and
Region}
\usage{
timestreamquery_list_scheduled_queries(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of items to return in the output. If the total number
of items available is more than the value specified, a \code{NextToken} is
provided in the output. To resume pagination, provide the \code{NextToken}
value as the argument to the subsequent call to
\code{ListScheduledQueriesRequest}.}

\item{NextToken}{A pagination token to resume pagination.}
}
\description{
Gets a list of all scheduled queries in the caller's Amazon account and Region. \code{\link[=timestreamquery_list_scheduled_queries]{list_scheduled_queries}} is eventually consistent.

See \url{https://paws-r.github.io/docs/timestreamquery/list_scheduled_queries.html} for full documentation.
}
\keyword{internal}
