% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamquery_operations.R
\name{timestreamquery_list_tags_for_resource}
\alias{timestreamquery_list_tags_for_resource}
\title{List all tags on a Timestream query resource}
\usage{
timestreamquery_list_tags_for_resource(
  ResourceARN,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ResourceARN}{[required] The Timestream resource with tags to be listed. This value is an Amazon
Resource Name (ARN).}

\item{MaxResults}{The maximum number of tags to return.}

\item{NextToken}{A pagination token to resume pagination.}
}
\description{
List all tags on a Timestream query resource.

See \url{https://www.paws-r-sdk.com/docs/timestreamquery_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
