% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamquery_operations.R
\name{timestreamquery_untag_resource}
\alias{timestreamquery_untag_resource}
\title{Removes the association of tags from a Timestream query resource}
\usage{
timestreamquery_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The Timestream resource that the tags will be removed from. This value
is an Amazon Resource Name (ARN).}

\item{TagKeys}{[required] A list of tags keys. Existing tags of the resource whose keys are
members of this list will be removed from the Timestream resource.}
}
\description{
Removes the association of tags from a Timestream query resource.

See \url{https://www.paws-r-sdk.com/docs/timestreamquery_untag_resource/} for full documentation.
}
\keyword{internal}
