% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamquery_operations.R
\name{timestreamquery_update_scheduled_query}
\alias{timestreamquery_update_scheduled_query}
\title{Update a scheduled query}
\usage{
timestreamquery_update_scheduled_query(ScheduledQueryArn, State)
}
\arguments{
\item{ScheduledQueryArn}{[required] ARN of the scheuled query.}

\item{State}{[required] State of the scheduled query.}
}
\description{
Update a scheduled query.

See \url{https://paws-r.github.io/docs/timestreamquery/update_scheduled_query.html} for full documentation.
}
\keyword{internal}
