% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamwrite_operations.R
\name{timestreamwrite_describe_database}
\alias{timestreamwrite_describe_database}
\title{Returns information about the database, including the database name,
time that the database was created, and the total number of tables found
within the database}
\usage{
timestreamwrite_describe_database(DatabaseName)
}
\arguments{
\item{DatabaseName}{[required] The name of the Timestream database.}
}
\description{
Returns information about the database, including the database name, time that the database was created, and the total number of tables found within the database. \href{https://docs.aws.amazon.com/timestream/latest/developerguide/ts-limits.html}{Service quotas apply}. See \href{https://docs.aws.amazon.com/timestream/latest/developerguide/code-samples.describe-db.html}{code sample} for details.

See \url{https://www.paws-r-sdk.com/docs/timestreamwrite_describe_database/} for full documentation.
}
\keyword{internal}
