% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamwrite_operations.R
\name{timestreamwrite_describe_endpoints}
\alias{timestreamwrite_describe_endpoints}
\title{DescribeEndpoints returns a list of available endpoints to make
Timestream API calls against}
\usage{
timestreamwrite_describe_endpoints()
}
\description{
DescribeEndpoints returns a list of available endpoints to make Timestream API calls against. This API is available through both Write and Query.

See \url{https://paws-r.github.io/docs/timestreamwrite/describe_endpoints.html} for full documentation.
}
\keyword{internal}
