% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamwrite_operations.R
\name{timestreamwrite_describe_table}
\alias{timestreamwrite_describe_table}
\title{Returns information about the table, including the table name, database
name, retention duration of the memory store and the magnetic store}
\usage{
timestreamwrite_describe_table(DatabaseName, TableName)
}
\arguments{
\item{DatabaseName}{[required] The name of the Timestream database.}

\item{TableName}{[required] The name of the Timestream table.}
}
\description{
Returns information about the table, including the table name, database name, retention duration of the memory store and the magnetic store. \href{https://docs.aws.amazon.com/timestream/latest/developerguide/ts-limits.html}{Service quotas apply}. See \href{https://docs.aws.amazon.com/timestream/latest/developerguide/code-samples.describe-table.html}{code sample} for details.

See \url{https://www.paws-r-sdk.com/docs/timestreamwrite_describe_table/} for full documentation.
}
\keyword{internal}
