% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamwrite_operations.R
\name{timestreamwrite_list_tags_for_resource}
\alias{timestreamwrite_list_tags_for_resource}
\title{Lists all tags on a Timestream resource}
\usage{
timestreamwrite_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The Timestream resource with tags to be listed. This value is an Amazon
Resource Name (ARN).}
}
\description{
Lists all tags on a Timestream resource.

See \url{https://paws-r.github.io/docs/timestreamwrite/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
