% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamwrite_operations.R
\name{timestreamwrite_untag_resource}
\alias{timestreamwrite_untag_resource}
\title{Removes the association of tags from a Timestream resource}
\usage{
timestreamwrite_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The Timestream resource that the tags will be removed from. This value
is an Amazon Resource Name (ARN).}

\item{TagKeys}{[required] A list of tags keys. Existing tags of the resource whose keys are
members of this list will be removed from the Timestream resource.}
}
\description{
Removes the association of tags from a Timestream resource.

See \url{https://paws-r.github.io/docs/timestreamwrite/untag_resource.html} for full documentation.
}
\keyword{internal}
