% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dax_operations.R
\name{dax_create_parameter_group}
\alias{dax_create_parameter_group}
\title{Creates a new parameter group}
\usage{
dax_create_parameter_group(ParameterGroupName, Description)
}
\arguments{
\item{ParameterGroupName}{[required] The name of the parameter group to apply to all of the clusters in this
replication group.}

\item{Description}{A description of the parameter group.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ParameterGroup = list(
    ParameterGroupName = "string",
    Description = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates a new parameter group. A parameter group is a collection of
parameters that you apply to all of the nodes in a DAX cluster.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_parameter_group(
  ParameterGroupName = "string",
  Description = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
