% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dax_operations.R
\name{dax_list_tags}
\alias{dax_list_tags}
\title{List all of the tags for a DAX cluster}
\usage{
dax_list_tags(ResourceName, NextToken)
}
\arguments{
\item{ResourceName}{[required] The name of the DAX resource to which the tags belong.}

\item{NextToken}{An optional token returned from a prior request. Use this token for
pagination of results from this action. If this parameter is specified,
the response includes only results beyond the token.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
List all of the tags for a DAX cluster. You can call
\code{\link[=dax_list_tags]{list_tags}} up to 10 times per second, per account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_tags(
  ResourceName = "string",
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
