% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dax_operations.R
\name{dax_reboot_node}
\alias{dax_reboot_node}
\title{Reboots a single node of a DAX cluster}
\usage{
dax_reboot_node(ClusterName, NodeId)
}
\arguments{
\item{ClusterName}{[required] The name of the DAX cluster containing the node to be rebooted.}

\item{NodeId}{[required] The system-assigned ID of the node to be rebooted.}
}
\description{
Reboots a single node of a DAX cluster. The reboot action takes place as
soon as possible. During the reboot, the node status is set to
REBOOTING.
}
\section{Request syntax}{
\preformatted{svc$reboot_node(
  ClusterName = "string",
  NodeId = "string"
)
}
}

\keyword{internal}
