% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dax_operations.R
\name{dax_update_subnet_group}
\alias{dax_update_subnet_group}
\title{Modifies an existing subnet group}
\usage{
dax_update_subnet_group(SubnetGroupName, Description, SubnetIds)
}
\arguments{
\item{SubnetGroupName}{[required] The name of the subnet group.}

\item{Description}{A description of the subnet group.}

\item{SubnetIds}{A list of subnet IDs in the subnet group.}
}
\description{
Modifies an existing subnet group.
}
\section{Request syntax}{
\preformatted{svc$update_subnet_group(
  SubnetGroupName = "string",
  Description = "string",
  SubnetIds = list(
    "string"
  )
)
}
}

\keyword{internal}
