% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dax_operations.R
\name{dax_update_subnet_group}
\alias{dax_update_subnet_group}
\title{Modifies an existing subnet group}
\usage{
dax_update_subnet_group(SubnetGroupName, Description = NULL, SubnetIds = NULL)
}
\arguments{
\item{SubnetGroupName}{[required] The name of the subnet group.}

\item{Description}{A description of the subnet group.}

\item{SubnetIds}{A list of subnet IDs in the subnet group.}
}
\description{
Modifies an existing subnet group.

See \url{https://www.paws-r-sdk.com/docs/dax_update_subnet_group/} for full documentation.
}
\keyword{internal}
