% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_create_db_subnet_group}
\alias{docdb_create_db_subnet_group}
\title{Creates a new subnet group}
\usage{
docdb_create_db_subnet_group(DBSubnetGroupName,
  DBSubnetGroupDescription, SubnetIds, Tags)
}
\arguments{
\item{DBSubnetGroupName}{[required] The name for the subnet group. This value is stored as a lowercase
string.

Constraints: Must contain no more than 255 letters, numbers, periods,
underscores, spaces, or hyphens. Must not be default.

Example: \code{mySubnetgroup}}

\item{DBSubnetGroupDescription}{[required] The description for the subnet group.}

\item{SubnetIds}{[required] The Amazon EC2 subnet IDs for the subnet group.}

\item{Tags}{The tags to be assigned to the subnet group.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DBSubnetGroup = list(
    DBSubnetGroupName = "string",
    DBSubnetGroupDescription = "string",
    VpcId = "string",
    SubnetGroupStatus = "string",
    Subnets = list(
      list(
        SubnetIdentifier = "string",
        SubnetAvailabilityZone = list(
          Name = "string"
        ),
        SubnetStatus = "string"
      )
    ),
    DBSubnetGroupArn = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates a new subnet group. subnet groups must contain at least one
subnet in at least two Availability Zones in the AWS Region.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_db_subnet_group(
  DBSubnetGroupName = "string",
  DBSubnetGroupDescription = "string",
  SubnetIds = list(
    "string"
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
