% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_delete_db_cluster_snapshot}
\alias{docdb_delete_db_cluster_snapshot}
\title{Deletes a cluster snapshot}
\usage{
docdb_delete_db_cluster_snapshot(DBClusterSnapshotIdentifier)
}
\arguments{
\item{DBClusterSnapshotIdentifier}{[required] The identifier of the cluster snapshot to delete.

Constraints: Must be the name of an existing cluster snapshot in the
\code{available} state.}
}
\description{
Deletes a cluster snapshot. If the snapshot is being copied, the copy operation is terminated.

See \url{https://www.paws-r-sdk.com/docs/docdb_delete_db_cluster_snapshot/} for full documentation.
}
\keyword{internal}
