% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_delete_db_subnet_group}
\alias{docdb_delete_db_subnet_group}
\title{Deletes a subnet group}
\usage{
docdb_delete_db_subnet_group(DBSubnetGroupName)
}
\arguments{
\item{DBSubnetGroupName}{[required] The name of the database subnet group to delete.

You can't delete the default subnet group.

Constraints:

Must match the name of an existing \code{DBSubnetGroup}. Must not be default.

Example: \code{mySubnetgroup}}
}
\value{
An empty list.
}
\description{
Deletes a subnet group.

The specified database subnet group must not be associated with any DB
instances.
}
\section{Request syntax}{
\preformatted{svc$delete_db_subnet_group(
  DBSubnetGroupName = "string"
)
}
}

\keyword{internal}
