% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_delete_event_subscription}
\alias{docdb_delete_event_subscription}
\title{Deletes an Amazon DocumentDB event notification subscription}
\usage{
docdb_delete_event_subscription(SubscriptionName)
}
\arguments{
\item{SubscriptionName}{[required] The name of the Amazon DocumentDB event notification subscription that
you want to delete.}
}
\description{
Deletes an Amazon DocumentDB event notification subscription.

See \url{https://www.paws-r-sdk.com/docs/docdb_delete_event_subscription/} for full documentation.
}
\keyword{internal}
