% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_describe_event_categories}
\alias{docdb_describe_event_categories}
\title{Displays a list of categories for all event source types, or, if
specified, for a specified source type}
\usage{
docdb_describe_event_categories(SourceType = NULL, Filters = NULL)
}
\arguments{
\item{SourceType}{The type of source that is generating the events.

Valid values: \code{db-instance}, \code{db-parameter-group}, \code{db-security-group}}

\item{Filters}{This parameter is not currently supported.}
}
\description{
Displays a list of categories for all event source types, or, if specified, for a specified source type.

See \url{https://paws-r.github.io/docs/docdb/describe_event_categories.html} for full documentation.
}
\keyword{internal}
