% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_failover_db_cluster}
\alias{docdb_failover_db_cluster}
\title{Forces a failover for a cluster}
\usage{
docdb_failover_db_cluster(
  DBClusterIdentifier = NULL,
  TargetDBInstanceIdentifier = NULL
)
}
\arguments{
\item{DBClusterIdentifier}{A cluster identifier to force a failover for. This parameter is not case
sensitive.

Constraints:
\itemize{
\item Must match the identifier of an existing \code{DBCluster}.
}}

\item{TargetDBInstanceIdentifier}{The name of the instance to promote to the primary instance.

You must specify the instance identifier for an Amazon DocumentDB
replica in the cluster. For example, \code{mydbcluster-replica1}.}
}
\description{
Forces a failover for a cluster.

See \url{https://paws-r.github.io/docs/docdb/failover_db_cluster.html} for full documentation.
}
\keyword{internal}
