% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_modify_db_instance}
\alias{docdb_modify_db_instance}
\title{Modifies settings for an instance}
\usage{
docdb_modify_db_instance(DBInstanceIdentifier, DBInstanceClass,
  ApplyImmediately, PreferredMaintenanceWindow, AutoMinorVersionUpgrade,
  NewDBInstanceIdentifier, CACertificateIdentifier, PromotionTier)
}
\arguments{
\item{DBInstanceIdentifier}{[required] The instance identifier. This value is stored as a lowercase string.

Constraints:
\itemize{
\item Must match the identifier of an existing \code{DBInstance}.
}}

\item{DBInstanceClass}{The new compute and memory capacity of the instance; for example,
\code{db.r5.large}. Not all instance classes are available in all AWS
Regions.

If you modify the instance class, an outage occurs during the change.
The change is applied during the next maintenance window, unless
\code{ApplyImmediately} is specified as \code{true} for this request.

Default: Uses existing setting.}

\item{ApplyImmediately}{Specifies whether the modifications in this request and any pending
modifications are asynchronously applied as soon as possible, regardless
of the \code{PreferredMaintenanceWindow} setting for the instance.

If this parameter is set to \code{false}, changes to the instance are applied
during the next maintenance window. Some parameter changes can cause an
outage and are applied on the next reboot.

Default: \code{false}}

\item{PreferredMaintenanceWindow}{The weekly time range (in UTC) during which system maintenance can
occur, which might result in an outage. Changing this parameter doesn't
result in an outage except in the following situation, and the change is
asynchronously applied as soon as possible. If there are pending actions
that cause a reboot, and the maintenance window is changed to include
the current time, changing this parameter causes a reboot of the
instance. If you are moving this window to the current time, there must
be at least 30 minutes between the current time and end of the window to
ensure that pending changes are applied.

Default: Uses existing setting.

Format: \code{ddd:hh24:mi-ddd:hh24:mi}

Valid days: Mon, Tue, Wed, Thu, Fri, Sat, Sun

Constraints: Must be at least 30 minutes.}

\item{AutoMinorVersionUpgrade}{Indicates that minor version upgrades are applied automatically to the
instance during the maintenance window. Changing this parameter doesn't
result in an outage except in the following case, and the change is
asynchronously applied as soon as possible. An outage results if this
parameter is set to \code{true} during the maintenance window, and a newer
minor version is available, and Amazon DocumentDB has enabled automatic
patching for that engine version.}

\item{NewDBInstanceIdentifier}{The new instance identifier for the instance when renaming an instance.
When you change the instance identifier, an instance reboot occurs
immediately if you set \verb{Apply Immediately} to \code{true}. It occurs during
the next maintenance window if you set \verb{Apply Immediately} to \code{false}.
This value is stored as a lowercase string.

Constraints:
\itemize{
\item Must contain from 1 to 63 letters, numbers, or hyphens.
\item The first character must be a letter.
\item Cannot end with a hyphen or contain two consecutive hyphens.
}

Example: \code{mydbinstance}}

\item{CACertificateIdentifier}{Indicates the certificate that needs to be associated with the instance.}

\item{PromotionTier}{A value that specifies the order in which an Amazon DocumentDB replica
is promoted to the primary instance after a failure of the existing
primary instance.

Default: 1

Valid values: 0-15}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DBInstance = list(
    DBInstanceIdentifier = "string",
    DBInstanceClass = "string",
    Engine = "string",
    DBInstanceStatus = "string",
    Endpoint = list(
      Address = "string",
      Port = 123,
      HostedZoneId = "string"
    ),
    InstanceCreateTime = as.POSIXct(
      "2015-01-01"
    ),
    PreferredBackupWindow = "string",
    BackupRetentionPeriod = 123,
    VpcSecurityGroups = list(
      list(
        VpcSecurityGroupId = "string",
        Status = "string"
      )
    ),
    AvailabilityZone = "string",
    DBSubnetGroup = list(
      DBSubnetGroupName = "string",
      DBSubnetGroupDescription = "string",
      VpcId = "string",
      SubnetGroupStatus = "string",
      Subnets = list(
        list(
          SubnetIdentifier = "string",
          SubnetAvailabilityZone = list(
            Name = "string"
          ),
          SubnetStatus = "string"
        )
      ),
      DBSubnetGroupArn = "string"
    ),
    PreferredMaintenanceWindow = "string",
    PendingModifiedValues = list(
      DBInstanceClass = "string",
      AllocatedStorage = 123,
      MasterUserPassword = "string",
      Port = 123,
      BackupRetentionPeriod = 123,
      MultiAZ = TRUE|FALSE,
      EngineVersion = "string",
      LicenseModel = "string",
      Iops = 123,
      DBInstanceIdentifier = "string",
      StorageType = "string",
      CACertificateIdentifier = "string",
      DBSubnetGroupName = "string",
      PendingCloudwatchLogsExports = list(
        LogTypesToEnable = list(
          "string"
        ),
        LogTypesToDisable = list(
          "string"
        )
      )
    ),
    LatestRestorableTime = as.POSIXct(
      "2015-01-01"
    ),
    EngineVersion = "string",
    AutoMinorVersionUpgrade = TRUE|FALSE,
    PubliclyAccessible = TRUE|FALSE,
    StatusInfos = list(
      list(
        StatusType = "string",
        Normal = TRUE|FALSE,
        Status = "string",
        Message = "string"
      )
    ),
    DBClusterIdentifier = "string",
    StorageEncrypted = TRUE|FALSE,
    KmsKeyId = "string",
    DbiResourceId = "string",
    CACertificateIdentifier = "string",
    PromotionTier = 123,
    DBInstanceArn = "string",
    EnabledCloudwatchLogsExports = list(
      "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Modifies settings for an instance. You can change one or more database
configuration parameters by specifying these parameters and the new
values in the request.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$modify_db_instance(
  DBInstanceIdentifier = "string",
  DBInstanceClass = "string",
  ApplyImmediately = TRUE|FALSE,
  PreferredMaintenanceWindow = "string",
  AutoMinorVersionUpgrade = TRUE|FALSE,
  NewDBInstanceIdentifier = "string",
  CACertificateIdentifier = "string",
  PromotionTier = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
