% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_reset_db_cluster_parameter_group}
\alias{docdb_reset_db_cluster_parameter_group}
\title{Modifies the parameters of a cluster parameter group to the default
value}
\usage{
docdb_reset_db_cluster_parameter_group(DBClusterParameterGroupName,
  ResetAllParameters, Parameters)
}
\arguments{
\item{DBClusterParameterGroupName}{[required] The name of the cluster parameter group to reset.}

\item{ResetAllParameters}{A value that is set to \code{true} to reset all parameters in the cluster
parameter group to their default values, and \code{false} otherwise. You
can't use this parameter if there is a list of parameter names specified
for the \code{Parameters} parameter.}

\item{Parameters}{A list of parameter names in the cluster parameter group to reset to the
default values. You can't use this parameter if the \code{ResetAllParameters}
parameter is set to \code{true}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DBClusterParameterGroupName = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Modifies the parameters of a cluster parameter group to the default
value. To reset specific parameters, submit a list of the following:
\code{ParameterName} and \code{ApplyMethod}. To reset the entire cluster parameter
group, specify the \code{DBClusterParameterGroupName} and
\code{ResetAllParameters} parameters.

When you reset the entire group, dynamic parameters are updated
immediately and static parameters are set to \code{pending-reboot} to take
effect on the next DB instance reboot.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$reset_db_cluster_parameter_group(
  DBClusterParameterGroupName = "string",
  ResetAllParameters = TRUE|FALSE,
  Parameters = list(
    list(
      ParameterName = "string",
      ParameterValue = "string",
      Description = "string",
      Source = "string",
      ApplyType = "string",
      DataType = "string",
      AllowedValues = "string",
      IsModifiable = TRUE|FALSE,
      MinimumEngineVersion = "string",
      ApplyMethod = "immediate"|"pending-reboot"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
