% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdbelastic_operations.R
\name{docdbelastic_tag_resource}
\alias{docdbelastic_tag_resource}
\title{Adds metadata tags to an elastic cluster resource}
\usage{
docdbelastic_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The ARN identifier of the elastic cluster resource.}

\item{tags}{[required] The tags that are assigned to the elastic cluster resource.}
}
\description{
Adds metadata tags to an elastic cluster resource

See \url{https://www.paws-r-sdk.com/docs/docdbelastic_tag_resource/} for full documentation.
}
\keyword{internal}
