% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_batch_execute_statement}
\alias{dynamodb_batch_execute_statement}
\title{This operation allows you to perform batch reads or writes on data
stored in DynamoDB, using PartiQL}
\usage{
dynamodb_batch_execute_statement(Statements, ReturnConsumedCapacity = NULL)
}
\arguments{
\item{Statements}{[required] The list of PartiQL statements representing the batch to run.}

\item{ReturnConsumedCapacity}{}
}
\description{
This operation allows you to perform batch reads or writes on data stored in DynamoDB, using PartiQL. Each read statement in a \code{\link[=dynamodb_batch_execute_statement]{batch_execute_statement}} must specify an equality condition on all key attributes. This enforces that each \code{SELECT} statement in a batch returns at most a single item.

See \url{https://www.paws-r-sdk.com/docs/dynamodb_batch_execute_statement/} for full documentation.
}
\keyword{internal}
