% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_create_backup}
\alias{dynamodb_create_backup}
\title{Creates a backup for an existing table}
\usage{
dynamodb_create_backup(TableName, BackupName)
}
\arguments{
\item{TableName}{[required] The name of the table.}

\item{BackupName}{[required] Specified name for the backup.}
}
\description{
Creates a backup for an existing table.

See \url{https://www.paws-r-sdk.com/docs/dynamodb_create_backup/} for full documentation.
}
\keyword{internal}
