% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_describe_backup}
\alias{dynamodb_describe_backup}
\title{Describes an existing backup of a table}
\usage{
dynamodb_describe_backup(BackupArn)
}
\arguments{
\item{BackupArn}{[required] The Amazon Resource Name (ARN) associated with the backup.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  BackupDescription = list(
    BackupDetails = list(
      BackupArn = "string",
      BackupName = "string",
      BackupSizeBytes = 123,
      BackupStatus = "CREATING"|"DELETED"|"AVAILABLE",
      BackupType = "USER"|"SYSTEM"|"AWS_BACKUP",
      BackupCreationDateTime = as.POSIXct(
        "2015-01-01"
      ),
      BackupExpiryDateTime = as.POSIXct(
        "2015-01-01"
      )
    ),
    SourceTableDetails = list(
      TableName = "string",
      TableId = "string",
      TableArn = "string",
      TableSizeBytes = 123,
      KeySchema = list(
        list(
          AttributeName = "string",
          KeyType = "HASH"|"RANGE"
        )
      ),
      TableCreationDateTime = as.POSIXct(
        "2015-01-01"
      ),
      ProvisionedThroughput = list(
        ReadCapacityUnits = 123,
        WriteCapacityUnits = 123
      ),
      ItemCount = 123,
      BillingMode = "PROVISIONED"|"PAY_PER_REQUEST"
    ),
    SourceTableFeatureDetails = list(
      LocalSecondaryIndexes = list(
        list(
          IndexName = "string",
          KeySchema = list(
            list(
              AttributeName = "string",
              KeyType = "HASH"|"RANGE"
            )
          ),
          Projection = list(
            ProjectionType = "ALL"|"KEYS_ONLY"|"INCLUDE",
            NonKeyAttributes = list(
              "string"
            )
          )
        )
      ),
      GlobalSecondaryIndexes = list(
        list(
          IndexName = "string",
          KeySchema = list(
            list(
              AttributeName = "string",
              KeyType = "HASH"|"RANGE"
            )
          ),
          Projection = list(
            ProjectionType = "ALL"|"KEYS_ONLY"|"INCLUDE",
            NonKeyAttributes = list(
              "string"
            )
          ),
          ProvisionedThroughput = list(
            ReadCapacityUnits = 123,
            WriteCapacityUnits = 123
          )
        )
      ),
      StreamDescription = list(
        StreamEnabled = TRUE|FALSE,
        StreamViewType = "NEW_IMAGE"|"OLD_IMAGE"|"NEW_AND_OLD_IMAGES"|"KEYS_ONLY"
      ),
      TimeToLiveDescription = list(
        TimeToLiveStatus = "ENABLING"|"DISABLING"|"ENABLED"|"DISABLED",
        AttributeName = "string"
      ),
      SSEDescription = list(
        Status = "ENABLING"|"ENABLED"|"DISABLING"|"DISABLED"|"UPDATING",
        SSEType = "AES256"|"KMS",
        KMSMasterKeyArn = "string",
        InaccessibleEncryptionDateTime = as.POSIXct(
          "2015-01-01"
        )
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes an existing backup of a table.

You can call \code{\link[=dynamodb_describe_backup]{describe_backup}} at a maximum
rate of 10 times per second.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_backup(
  BackupArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
