% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_describe_backup}
\alias{dynamodb_describe_backup}
\title{Describes an existing backup of a table}
\usage{
dynamodb_describe_backup(BackupArn)
}
\arguments{
\item{BackupArn}{[required] The Amazon Resource Name (ARN) associated with the backup.}
}
\description{
Describes an existing backup of a table.

See \url{https://www.paws-r-sdk.com/docs/dynamodb_describe_backup/} for full documentation.
}
\keyword{internal}
