% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_describe_endpoints}
\alias{dynamodb_describe_endpoints}
\title{Returns the regional endpoint information}
\usage{
dynamodb_describe_endpoints()
}
\description{
Returns the regional endpoint information. This action must be included in your VPC endpoint policies, or access to the DescribeEndpoints API will be denied. For more information on policy permissions, please see \href{https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/inter-network-traffic-privacy.html#inter-network-traffic-DescribeEndpoints}{Internetwork traffic privacy}.

See \url{https://www.paws-r-sdk.com/docs/dynamodb_describe_endpoints/} for full documentation.
}
\keyword{internal}
