% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_describe_endpoints}
\alias{dynamodb_describe_endpoints}
\title{Returns the regional endpoint information}
\usage{
dynamodb_describe_endpoints()
}
\value{
A list with the following syntax:\preformatted{list(
  Endpoints = list(
    list(
      Address = "string",
      CachePeriodInMinutes = 123
    )
  )
)
}
}
\description{
Returns the regional endpoint information.
}
\section{Request syntax}{
\preformatted{svc$describe_endpoints()
}
}

\keyword{internal}
