% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_describe_export}
\alias{dynamodb_describe_export}
\title{Describes an existing table export}
\usage{
dynamodb_describe_export(ExportArn)
}
\arguments{
\item{ExportArn}{[required] The Amazon Resource Name (ARN) associated with the export.}
}
\description{
Describes an existing table export.

See \url{https://www.paws-r-sdk.com/docs/dynamodb_describe_export/} for full documentation.
}
\keyword{internal}
