% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_describe_global_table_settings}
\alias{dynamodb_describe_global_table_settings}
\title{Describes Region-specific settings for a global table}
\usage{
dynamodb_describe_global_table_settings(GlobalTableName)
}
\arguments{
\item{GlobalTableName}{[required] The name of the global table to describe.}
}
\description{
Describes Region-specific settings for a global table.

See \url{https://www.paws-r-sdk.com/docs/dynamodb_describe_global_table_settings/} for full documentation.
}
\keyword{internal}
