% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_describe_kinesis_streaming_destination}
\alias{dynamodb_describe_kinesis_streaming_destination}
\title{Returns information about the status of Kinesis streaming}
\usage{
dynamodb_describe_kinesis_streaming_destination(TableName)
}
\arguments{
\item{TableName}{[required] The name of the table being described. You can also provide the Amazon
Resource Name (ARN) of the table in this parameter.}
}
\description{
Returns information about the status of Kinesis streaming.

See \url{https://www.paws-r-sdk.com/docs/dynamodb_describe_kinesis_streaming_destination/} for full documentation.
}
\keyword{internal}
