% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_describe_limits}
\alias{dynamodb_describe_limits}
\title{Returns the current provisioned-capacity quotas for your Amazon Web
Services account in a Region, both for the Region as a whole and for any
one DynamoDB table that you create there}
\usage{
dynamodb_describe_limits()
}
\description{
Returns the current provisioned-capacity quotas for your Amazon Web Services account in a Region, both for the Region as a whole and for any one DynamoDB table that you create there.

See \url{https://paws-r.github.io/docs/dynamodb/describe_limits.html} for full documentation.
}
\keyword{internal}
