% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_describe_table_replica_auto_scaling}
\alias{dynamodb_describe_table_replica_auto_scaling}
\title{Describes auto scaling settings across replicas of the global table at
once}
\usage{
dynamodb_describe_table_replica_auto_scaling(TableName)
}
\arguments{
\item{TableName}{[required] The name of the table.}
}
\description{
Describes auto scaling settings across replicas of the global table at once.

See \url{https://www.paws-r-sdk.com/docs/dynamodb_describe_table_replica_auto_scaling/} for full documentation.
}
\keyword{internal}
