% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_describe_table_replica_auto_scaling}
\alias{dynamodb_describe_table_replica_auto_scaling}
\title{Describes auto scaling settings across replicas of the global table at
once}
\usage{
dynamodb_describe_table_replica_auto_scaling(TableName)
}
\arguments{
\item{TableName}{[required] The name of the table.}
}
\description{
Describes auto scaling settings across replicas of the global table at
once.
}
\details{
This method only applies to \href{https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/globaltables.V2.html}{Version 2019.11.21}
of global tables.
}
\section{Request syntax}{
\preformatted{svc$describe_table_replica_auto_scaling(
  TableName = "string"
)
}
}

\keyword{internal}
