% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_describe_time_to_live}
\alias{dynamodb_describe_time_to_live}
\title{Gives a description of the Time to Live (TTL) status on the specified
table}
\usage{
dynamodb_describe_time_to_live(TableName)
}
\arguments{
\item{TableName}{[required] The name of the table to be described. You can also provide the Amazon
Resource Name (ARN) of the table in this parameter.}
}
\description{
Gives a description of the Time to Live (TTL) status on the specified table.

See \url{https://www.paws-r-sdk.com/docs/dynamodb_describe_time_to_live/} for full documentation.
}
\keyword{internal}
