% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_disable_kinesis_streaming_destination}
\alias{dynamodb_disable_kinesis_streaming_destination}
\title{Stops replication from the DynamoDB table to the Kinesis data stream}
\usage{
dynamodb_disable_kinesis_streaming_destination(
  TableName,
  StreamArn,
  EnableKinesisStreamingConfiguration = NULL
)
}
\arguments{
\item{TableName}{[required] The name of the DynamoDB table. You can also provide the Amazon Resource
Name (ARN) of the table in this parameter.}

\item{StreamArn}{[required] The ARN for a Kinesis data stream.}

\item{EnableKinesisStreamingConfiguration}{The source for the Kinesis streaming information that is being enabled.}
}
\description{
Stops replication from the DynamoDB table to the Kinesis data stream. This is done without deleting either of the resources.

See \url{https://www.paws-r-sdk.com/docs/dynamodb_disable_kinesis_streaming_destination/} for full documentation.
}
\keyword{internal}
