% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_disable_kinesis_streaming_destination}
\alias{dynamodb_disable_kinesis_streaming_destination}
\title{Stops replication from the DynamoDB table to the Kinesis data stream}
\usage{
dynamodb_disable_kinesis_streaming_destination(TableName, StreamArn)
}
\arguments{
\item{TableName}{[required] The name of the DynamoDB table.}

\item{StreamArn}{[required] The ARN for a Kinesis data stream.}
}
\description{
Stops replication from the DynamoDB table to the Kinesis data stream. This is done without deleting either of the resources.

See \url{https://paws-r.github.io/docs/dynamodb/disable_kinesis_streaming_destination.html} for full documentation.
}
\keyword{internal}
