% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_list_contributor_insights}
\alias{dynamodb_list_contributor_insights}
\title{Returns a list of ContributorInsightsSummary for a table and all its
global secondary indexes}
\usage{
dynamodb_list_contributor_insights(
  TableName = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{TableName}{The name of the table.}

\item{NextToken}{A token to for the desired page, if there is one.}

\item{MaxResults}{Maximum number of results to return per page.}
}
\description{
Returns a list of ContributorInsightsSummary for a table and all its global secondary indexes.

See \url{https://paws-r.github.io/docs/dynamodb/list_contributor_insights.html} for full documentation.
}
\keyword{internal}
