% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_list_contributor_insights}
\alias{dynamodb_list_contributor_insights}
\title{Returns a list of ContributorInsightsSummary for a table and all its
global secondary indexes}
\usage{
dynamodb_list_contributor_insights(TableName, NextToken, MaxResults)
}
\arguments{
\item{TableName}{The name of the table.}

\item{NextToken}{A token to for the desired page, if there is one.}

\item{MaxResults}{Maximum number of results to return per page.}
}
\value{
A list with the following syntax:\preformatted{list(
  ContributorInsightsSummaries = list(
    list(
      TableName = "string",
      IndexName = "string",
      ContributorInsightsStatus = "ENABLING"|"ENABLED"|"DISABLING"|"DISABLED"|"FAILED"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Returns a list of ContributorInsightsSummary for a table and all its
global secondary indexes.
}
\section{Request syntax}{
\preformatted{svc$list_contributor_insights(
  TableName = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
