% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_list_tags_of_resource}
\alias{dynamodb_list_tags_of_resource}
\title{List all tags on an Amazon DynamoDB resource}
\usage{
dynamodb_list_tags_of_resource(ResourceArn, NextToken)
}
\arguments{
\item{ResourceArn}{[required] The Amazon DynamoDB resource with tags to be listed. This value is an
Amazon Resource Name (ARN).}

\item{NextToken}{An optional string that, if supplied, must be copied from the output of
a previous call to ListTagOfResource. When provided in this manner, this
API fetches the next page of results.}
}
\description{
List all tags on an Amazon DynamoDB resource. You can call
ListTagsOfResource up to 10 times per second, per account.
}
\details{
For an overview on tagging DynamoDB resources, see \href{https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/Tagging.html}{Tagging for DynamoDB}
in the \emph{Amazon DynamoDB Developer Guide}.
}
\section{Request syntax}{
\preformatted{svc$list_tags_of_resource(
  ResourceArn = "string",
  NextToken = "string"
)
}
}

\keyword{internal}
