% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_update_global_table}
\alias{dynamodb_update_global_table}
\title{Adds or removes replicas in the specified global table}
\usage{
dynamodb_update_global_table(GlobalTableName, ReplicaUpdates)
}
\arguments{
\item{GlobalTableName}{[required] The global table name.}

\item{ReplicaUpdates}{[required] A list of Regions that should be added or removed from the global table.}
}
\description{
Adds or removes replicas in the specified global table. The global table must already exist to be able to use this operation. Any replica to be added must be empty, have the same name as the global table, have the same key schema, have DynamoDB Streams enabled, and have the same provisioned and maximum write capacity units.

See \url{https://www.paws-r-sdk.com/docs/dynamodb_update_global_table/} for full documentation.
}
\keyword{internal}
