% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_operations.R
\name{dynamodb_update_table_replica_auto_scaling}
\alias{dynamodb_update_table_replica_auto_scaling}
\title{Updates auto scaling settings on your global tables at once}
\usage{
dynamodb_update_table_replica_auto_scaling(
  GlobalSecondaryIndexUpdates = NULL,
  TableName,
  ProvisionedWriteCapacityAutoScalingUpdate = NULL,
  ReplicaUpdates = NULL
)
}
\arguments{
\item{GlobalSecondaryIndexUpdates}{Represents the auto scaling settings of the global secondary indexes of
the replica to be updated.}

\item{TableName}{[required] The name of the global table to be updated. You can also provide the
Amazon Resource Name (ARN) of the table in this parameter.}

\item{ProvisionedWriteCapacityAutoScalingUpdate}{}

\item{ReplicaUpdates}{Represents the auto scaling settings of replicas of the table that will
be modified.}
}
\description{
Updates auto scaling settings on your global tables at once.

See \url{https://www.paws-r-sdk.com/docs/dynamodb_update_table_replica_auto_scaling/} for full documentation.
}
\keyword{internal}
