% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodbstreams_operations.R
\name{dynamodbstreams_describe_stream}
\alias{dynamodbstreams_describe_stream}
\title{Returns information about a stream, including the current status of the
stream, its Amazon Resource Name (ARN), the composition of its shards,
and its corresponding DynamoDB table}
\usage{
dynamodbstreams_describe_stream(
  StreamArn,
  Limit = NULL,
  ExclusiveStartShardId = NULL,
  ShardFilter = NULL
)
}
\arguments{
\item{StreamArn}{[required] The Amazon Resource Name (ARN) for the stream.}

\item{Limit}{The maximum number of shard objects to return. The upper limit is 100.}

\item{ExclusiveStartShardId}{The shard ID of the first item that this operation will evaluate. Use
the value that was returned for \code{LastEvaluatedShardId} in the previous
operation.}

\item{ShardFilter}{This optional field contains the filter definition for the
\code{\link[=dynamodbstreams_describe_stream]{describe_stream}} API.}
}
\description{
Returns information about a stream, including the current status of the stream, its Amazon Resource Name (ARN), the composition of its shards, and its corresponding DynamoDB table.

See \url{https://www.paws-r-sdk.com/docs/dynamodbstreams_describe_stream/} for full documentation.
}
\keyword{internal}
