% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_batch_apply_update_action}
\alias{elasticache_batch_apply_update_action}
\title{Apply the service update}
\usage{
elasticache_batch_apply_update_action(
  ReplicationGroupIds = NULL,
  CacheClusterIds = NULL,
  ServiceUpdateName
)
}
\arguments{
\item{ReplicationGroupIds}{The replication group IDs}

\item{CacheClusterIds}{The cache cluster IDs}

\item{ServiceUpdateName}{[required] The unique ID of the service update}
}
\description{
Apply the service update. For more information on service updates and applying them, see \href{https://docs.aws.amazon.com/AmazonElastiCache/latest/red-ug/}{Applying Service Updates}.

See \url{https://www.paws-r-sdk.com/docs/elasticache_batch_apply_update_action/} for full documentation.
}
\keyword{internal}
