% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_create_snapshot}
\alias{elasticache_create_snapshot}
\title{Creates a copy of an entire cluster or replication group at a specific
moment in time}
\usage{
elasticache_create_snapshot(
  ReplicationGroupId = NULL,
  CacheClusterId = NULL,
  SnapshotName,
  KmsKeyId = NULL,
  Tags = NULL
)
}
\arguments{
\item{ReplicationGroupId}{The identifier of an existing replication group. The snapshot is created
from this replication group.}

\item{CacheClusterId}{The identifier of an existing cluster. The snapshot is created from this
cluster.}

\item{SnapshotName}{[required] A name for the snapshot being created.}

\item{KmsKeyId}{The ID of the KMS key used to encrypt the snapshot.}

\item{Tags}{A list of tags to be added to this resource. A tag is a key-value pair.
A tag key must be accompanied by a tag value, although null is accepted.}
}
\description{
Creates a copy of an entire cluster or replication group at a specific moment in time.

See \url{https://paws-r.github.io/docs/elasticache/create_snapshot.html} for full documentation.
}
\keyword{internal}
