% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_create_user_group}
\alias{elasticache_create_user_group}
\title{For Redis engine version 6}
\usage{
elasticache_create_user_group(UserGroupId, Engine, UserIds)
}
\arguments{
\item{UserGroupId}{[required] The ID of the user group.}

\item{Engine}{[required] The current supported value is Redis.}

\item{UserIds}{The list of user IDs that belong to the user group.}
}
\description{
For Redis engine version 6.x onwards: Creates a Redis user group. For
more information, see \href{https://docs.aws.amazon.com/AmazonElastiCache/latest/red-ug/Clusters.RBAC.html}{Using Role Based Access Control (RBAC)}
}
\section{Request syntax}{
\preformatted{svc$create_user_group(
  UserGroupId = "string",
  Engine = "string",
  UserIds = list(
    "string"
  )
)
}
}

\keyword{internal}
