% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_delete_cache_security_group}
\alias{elasticache_delete_cache_security_group}
\title{Deletes a cache security group}
\usage{
elasticache_delete_cache_security_group(CacheSecurityGroupName)
}
\arguments{
\item{CacheSecurityGroupName}{[required] The name of the cache security group to delete.

You cannot delete the default security group.}
}
\description{
Deletes a cache security group.
}
\details{
You cannot delete a cache security group if it is associated with any
clusters.
}
\section{Request syntax}{
\preformatted{svc$delete_cache_security_group(
  CacheSecurityGroupName = "string"
)
}
}

\keyword{internal}
