% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_delete_user}
\alias{elasticache_delete_user}
\title{For Redis engine version 6}
\usage{
elasticache_delete_user(UserId)
}
\arguments{
\item{UserId}{[required] The ID of the user.}
}
\description{
For Redis engine version 6.0 onwards: Deletes a user. The user will be removed from all user groups and in turn removed from all replication groups. For more information, see \href{https://docs.aws.amazon.com/AmazonElastiCache/latest/red-ug/Clusters.RBAC.html}{Using Role Based Access Control (RBAC)}.

See \url{https://paws-r.github.io/docs/elasticache/delete_user.html} for full documentation.
}
\keyword{internal}
