% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_delete_user_group}
\alias{elasticache_delete_user_group}
\title{For Redis OSS engine version 6}
\usage{
elasticache_delete_user_group(UserGroupId)
}
\arguments{
\item{UserGroupId}{[required] The ID of the user group.}
}
\description{
For Redis OSS engine version 6.0 onwards: Deletes a user group. The user group must first be disassociated from the replication group before it can be deleted. For more information, see \href{https://docs.aws.amazon.com/AmazonElastiCache/latest/red-ug/Clusters.RBAC.html}{Using Role Based Access Control (RBAC)}.

See \url{https://www.paws-r-sdk.com/docs/elasticache_delete_user_group/} for full documentation.
}
\keyword{internal}
