% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_describe_cache_parameter_groups}
\alias{elasticache_describe_cache_parameter_groups}
\title{Returns a list of cache parameter group descriptions}
\usage{
elasticache_describe_cache_parameter_groups(CacheParameterGroupName,
  MaxRecords, Marker)
}
\arguments{
\item{CacheParameterGroupName}{The name of a specific cache parameter group to return details for.}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a marker is
included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.}

\item{Marker}{An optional marker returned from a prior request. Use this marker for
pagination of results from this operation. If this parameter is
specified, the response includes only records beyond the marker, up to
the value specified by \code{MaxRecords}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Marker = "string",
  CacheParameterGroups = list(
    list(
      CacheParameterGroupName = "string",
      CacheParameterGroupFamily = "string",
      Description = "string",
      IsGlobal = TRUE|FALSE,
      ARN = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns a list of cache parameter group descriptions. If a cache
parameter group name is specified, the list contains only the
descriptions for that group.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_cache_parameter_groups(
  CacheParameterGroupName = "string",
  MaxRecords = 123,
  Marker = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
