% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_describe_update_actions}
\alias{elasticache_describe_update_actions}
\title{Returns details of the update actions}
\usage{
elasticache_describe_update_actions(
  ServiceUpdateName = NULL,
  ReplicationGroupIds = NULL,
  CacheClusterIds = NULL,
  Engine = NULL,
  ServiceUpdateStatus = NULL,
  ServiceUpdateTimeRange = NULL,
  UpdateActionStatus = NULL,
  ShowNodeLevelUpdateStatus = NULL,
  MaxRecords = NULL,
  Marker = NULL
)
}
\arguments{
\item{ServiceUpdateName}{The unique ID of the service update}

\item{ReplicationGroupIds}{The replication group IDs}

\item{CacheClusterIds}{The cache cluster IDs}

\item{Engine}{The Elasticache engine to which the update applies. Either Redis OSS or
Memcached.}

\item{ServiceUpdateStatus}{The status of the service update}

\item{ServiceUpdateTimeRange}{The range of time specified to search for service updates that are in
available status}

\item{UpdateActionStatus}{The status of the update action.}

\item{ShowNodeLevelUpdateStatus}{Dictates whether to include node level update status in the response}

\item{MaxRecords}{The maximum number of records to include in the response}

\item{Marker}{An optional marker returned from a prior request. Use this marker for
pagination of results from this operation. If this parameter is
specified, the response includes only records beyond the marker, up to
the value specified by \code{MaxRecords}.}
}
\description{
Returns details of the update actions

See \url{https://www.paws-r-sdk.com/docs/elasticache_describe_update_actions/} for full documentation.
}
\keyword{internal}
