% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_describe_user_groups}
\alias{elasticache_describe_user_groups}
\title{Returns a list of user groups}
\usage{
elasticache_describe_user_groups(
  UserGroupId = NULL,
  MaxRecords = NULL,
  Marker = NULL
)
}
\arguments{
\item{UserGroupId}{The ID of the user group.}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified MaxRecords value, a marker is included
in the response so that the remaining results can be retrieved.}

\item{Marker}{An optional marker returned from a prior request. Use this marker for
pagination of results from this operation. If this parameter is
specified, the response includes only records beyond the marker, up to
the value specified by MaxRecords. \>}
}
\description{
Returns a list of user groups.

See \url{https://paws-r.github.io/docs/elasticache/describe_user_groups.html} for full documentation.
}
\keyword{internal}
