% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_describe_users}
\alias{elasticache_describe_users}
\title{Returns a list of users}
\usage{
elasticache_describe_users(
  Engine = NULL,
  UserId = NULL,
  Filters = NULL,
  MaxRecords = NULL,
  Marker = NULL
)
}
\arguments{
\item{Engine}{The engine.}

\item{UserId}{The ID of the user.}

\item{Filters}{Filter to determine the list of User IDs to return.}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified MaxRecords value, a marker is included
in the response so that the remaining results can be retrieved.}

\item{Marker}{An optional marker returned from a prior request. Use this marker for
pagination of results from this operation. If this parameter is
specified, the response includes only records beyond the marker, up to
the value specified by MaxRecords. \>}
}
\description{
Returns a list of users.

See \url{https://www.paws-r-sdk.com/docs/elasticache_describe_users/} for full documentation.
}
\keyword{internal}
