% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_failover_global_replication_group}
\alias{elasticache_failover_global_replication_group}
\title{Used to failover the primary region to a secondary region}
\usage{
elasticache_failover_global_replication_group(
  GlobalReplicationGroupId,
  PrimaryRegion,
  PrimaryReplicationGroupId
)
}
\arguments{
\item{GlobalReplicationGroupId}{[required] The name of the Global datastore}

\item{PrimaryRegion}{[required] The Amazon region of the primary cluster of the Global datastore}

\item{PrimaryReplicationGroupId}{[required] The name of the primary replication group}
}
\description{
Used to failover the primary region to a secondary region. The secondary region will become primary, and all other clusters will become secondary.

See \url{https://paws-r.github.io/docs/elasticache/failover_global_replication_group.html} for full documentation.
}
\keyword{internal}
