% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_failover_global_replication_group}
\alias{elasticache_failover_global_replication_group}
\title{Used to failover the primary region to a selected secondary region}
\usage{
elasticache_failover_global_replication_group(GlobalReplicationGroupId,
  PrimaryRegion, PrimaryReplicationGroupId)
}
\arguments{
\item{GlobalReplicationGroupId}{[required] The name of the Global Datastore}

\item{PrimaryRegion}{[required] The AWS region of the primary cluster of the Global Datastore}

\item{PrimaryReplicationGroupId}{[required] The name of the primary replication group}
}
\description{
Used to failover the primary region to a selected secondary region. The
selected secondary region will become primary, and all other clusters
will become secondary.
}
\section{Request syntax}{
\preformatted{svc$failover_global_replication_group(
  GlobalReplicationGroupId = "string",
  PrimaryRegion = "string",
  PrimaryReplicationGroupId = "string"
)
}
}

\keyword{internal}
