% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_list_tags_for_resource}
\alias{elasticache_list_tags_for_resource}
\title{Lists all tags currently on a named resource}
\usage{
elasticache_list_tags_for_resource(ResourceName)
}
\arguments{
\item{ResourceName}{[required] The Amazon Resource Name (ARN) of the resource for which you want the
list of tags, for example
\code{arn:aws:elasticache:us-west-2:0123456789:cluster:myCluster} or
\code{arn:aws:elasticache:us-west-2:0123456789:snapshot:mySnapshot}.

For more information about ARNs, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html}{Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces}.}
}
\description{
Lists all tags currently on a named resource.

See \url{https://paws-r.github.io/docs/elasticache/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
