% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_list_tags_for_resource}
\alias{elasticache_list_tags_for_resource}
\title{Lists all tags currently on a named resource}
\usage{
elasticache_list_tags_for_resource(ResourceName)
}
\arguments{
\item{ResourceName}{[required] The Amazon Resource Name (ARN) of the resource for which you want the
list of tags, for example
\code{arn:aws:elasticache:us-west-2:0123456789:cluster:myCluster} or
\code{arn:aws:elasticache:us-west-2:0123456789:snapshot:mySnapshot}.

For more information about ARNs, see \href{https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{Amazon Resource Names (ARNs) and AWS Service Namespaces}.}
}
\value{
A list with the following syntax:\preformatted{list(
  TagList = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}
\description{
Lists all tags currently on a named resource.

A tag is a key-value pair where the key and value are case-sensitive.
You can use tags to categorize and track all your ElastiCache resources,
with the exception of global replication group. When you add or remove
tags on replication groups, those actions will be replicated to all
nodes in the replication group. For more information, see
\href{https://docs.aws.amazon.com/AmazonElastiCache/latest/red-ug/IAM.ResourceLevelPermissions.html}{Resource-level permissions}.

If the cluster is not in the \emph{available} state,
\code{\link[=elasticache_list_tags_for_resource]{list_tags_for_resource}} returns
an error.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceName = "string"
)
}
}

\keyword{internal}
