% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_modify_replication_group}
\alias{elasticache_modify_replication_group}
\title{Modifies the settings for a replication group}
\usage{
elasticache_modify_replication_group(ReplicationGroupId,
  ReplicationGroupDescription, PrimaryClusterId, SnapshottingClusterId,
  AutomaticFailoverEnabled, MultiAZEnabled, NodeGroupId,
  CacheSecurityGroupNames, SecurityGroupIds, PreferredMaintenanceWindow,
  NotificationTopicArn, CacheParameterGroupName, NotificationTopicStatus,
  ApplyImmediately, EngineVersion, AutoMinorVersionUpgrade,
  SnapshotRetentionLimit, SnapshotWindow, CacheNodeType, AuthToken,
  AuthTokenUpdateStrategy)
}
\arguments{
\item{ReplicationGroupId}{[required] The identifier of the replication group to modify.}

\item{ReplicationGroupDescription}{A description for the replication group. Maximum length is 255
characters.}

\item{PrimaryClusterId}{For replication groups with a single primary, if this parameter is
specified, ElastiCache promotes the specified cluster in the specified
replication group to the primary role. The nodes of all other clusters
in the replication group are read replicas.}

\item{SnapshottingClusterId}{The cluster ID that is used as the daily snapshot source for the
replication group. This parameter cannot be set for Redis (cluster mode
enabled) replication groups.}

\item{AutomaticFailoverEnabled}{Determines whether a read replica is automatically promoted to
read/write primary if the existing primary encounters a failure.

Valid values: \code{true} | \code{false}}

\item{MultiAZEnabled}{A flag indicating if you have Multi-AZ enabled to enhance fault
tolerance. For more information, see \href{http://docs.aws.amazon.com/AmazonElastiCache/latest/red-ug/AutoFailover.html}{Minimizing Downtime: Multi-AZ}.}

\item{NodeGroupId}{Deprecated. This parameter is not used.}

\item{CacheSecurityGroupNames}{A list of cache security group names to authorize for the clusters in
this replication group. This change is asynchronously applied as soon as
possible.

This parameter can be used only with replication group containing
clusters running outside of an Amazon Virtual Private Cloud (Amazon
VPC).

Constraints: Must contain no more than 255 alphanumeric characters. Must
not be \code{Default}.}

\item{SecurityGroupIds}{Specifies the VPC Security Groups associated with the clusters in the
replication group.

This parameter can be used only with replication group containing
clusters running in an Amazon Virtual Private Cloud (Amazon VPC).}

\item{PreferredMaintenanceWindow}{Specifies the weekly time range during which maintenance on the cluster
is performed. It is specified as a range in the format
ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window
is a 60 minute period.

Valid values for \code{ddd} are:
\itemize{
\item \code{sun}
\item \code{mon}
\item \code{tue}
\item \code{wed}
\item \code{thu}
\item \code{fri}
\item \code{sat}
}

Example: \code{sun:23:00-mon:01:30}}

\item{NotificationTopicArn}{The Amazon Resource Name (ARN) of the Amazon SNS topic to which
notifications are sent.

The Amazon SNS topic owner must be same as the replication group owner.}

\item{CacheParameterGroupName}{The name of the cache parameter group to apply to all of the clusters in
this replication group. This change is asynchronously applied as soon as
possible for parameters when the \code{ApplyImmediately} parameter is
specified as \code{true} for this request.}

\item{NotificationTopicStatus}{The status of the Amazon SNS notification topic for the replication
group. Notifications are sent only if the status is \code{active}.

Valid values: \code{active} | \code{inactive}}

\item{ApplyImmediately}{If \code{true}, this parameter causes the modifications in this request and
any pending modifications to be applied, asynchronously and as soon as
possible, regardless of the \code{PreferredMaintenanceWindow} setting for the
replication group.

If \code{false}, changes to the nodes in the replication group are applied on
the next maintenance reboot, or the next failure reboot, whichever
occurs first.

Valid values: \code{true} | \code{false}

Default: \code{false}}

\item{EngineVersion}{The upgraded version of the cache engine to be run on the clusters in
the replication group.

\strong{Important:} You can upgrade to a newer engine version (see \href{https://docs.aws.amazon.com/AmazonElastiCache/latest/red-ug/SelectEngine.html#VersionManagement}{Selecting a Cache Engine and Version}),
but you cannot downgrade to an earlier engine version. If you want to
use an earlier engine version, you must delete the existing replication
group and create it anew with the earlier engine version.}

\item{AutoMinorVersionUpgrade}{This parameter is currently disabled.}

\item{SnapshotRetentionLimit}{The number of days for which ElastiCache retains automatic node group
(shard) snapshots before deleting them. For example, if you set
\code{SnapshotRetentionLimit} to 5, a snapshot that was taken today is
retained for 5 days before being deleted.

\strong{Important} If the value of SnapshotRetentionLimit is set to zero (0),
backups are turned off.}

\item{SnapshotWindow}{The daily time range (in UTC) during which ElastiCache begins taking a
daily snapshot of the node group (shard) specified by
\code{SnapshottingClusterId}.

Example: \code{05:00-09:00}

If you do not specify this parameter, ElastiCache automatically chooses
an appropriate time range.}

\item{CacheNodeType}{A valid cache node type that you want to scale this replication group
to.}

\item{AuthToken}{Reserved parameter. The password used to access a password protected
server. This parameter must be specified with the
\code{auth-token-update-strategy } parameter. Password constraints:
\itemize{
\item Must be only printable ASCII characters
\item Must be at least 16 characters and no more than 128 characters in
length
\item Cannot contain any of the following characters: '/', '"', or '@',
'\%'
}

For more information, see AUTH password at
\href{http://redis.io/commands/AUTH}{AUTH}.}

\item{AuthTokenUpdateStrategy}{Specifies the strategy to use to update the AUTH token. This parameter
must be specified with the \code{auth-token} parameter. Possible values:
\itemize{
\item Rotate
\item Set
}

For more information, see \href{https://docs.aws.amazon.com/AmazonElastiCache/latest/red-ug/auth.html}{Authenticating Users with Redis AUTH}}
}
\description{
Modifies the settings for a replication group.
}
\details{
\itemize{
\item \href{https://docs.aws.amazon.com/AmazonElastiCache/latest/red-ug/scaling-redis-cluster-mode-enabled.html}{Scaling for Amazon ElastiCache for Redis (cluster mode enabled)}
in the ElastiCache User Guide
\item \href{https://docs.aws.amazon.com/AmazonElastiCache/latest/APIReference/API_ModifyReplicationGroupShardConfiguration.html}{ModifyReplicationGroupShardConfiguration}
in the ElastiCache API Reference
}

This operation is valid for Redis only.
}
\section{Request syntax}{
\preformatted{svc$modify_replication_group(
  ReplicationGroupId = "string",
  ReplicationGroupDescription = "string",
  PrimaryClusterId = "string",
  SnapshottingClusterId = "string",
  AutomaticFailoverEnabled = TRUE|FALSE,
  MultiAZEnabled = TRUE|FALSE,
  NodeGroupId = "string",
  CacheSecurityGroupNames = list(
    "string"
  ),
  SecurityGroupIds = list(
    "string"
  ),
  PreferredMaintenanceWindow = "string",
  NotificationTopicArn = "string",
  CacheParameterGroupName = "string",
  NotificationTopicStatus = "string",
  ApplyImmediately = TRUE|FALSE,
  EngineVersion = "string",
  AutoMinorVersionUpgrade = TRUE|FALSE,
  SnapshotRetentionLimit = 123,
  SnapshotWindow = "string",
  CacheNodeType = "string",
  AuthToken = "string",
  AuthTokenUpdateStrategy = "SET"|"ROTATE"
)
}
}

\keyword{internal}
