% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_modify_replication_group_shard_configuration}
\alias{elasticache_modify_replication_group_shard_configuration}
\title{Modifies a replication group's shards (node groups) by allowing you to
add shards, remove shards, or rebalance the keyspaces among existing
shards}
\usage{
elasticache_modify_replication_group_shard_configuration(
  ReplicationGroupId,
  NodeGroupCount,
  ApplyImmediately,
  ReshardingConfiguration = NULL,
  NodeGroupsToRemove = NULL,
  NodeGroupsToRetain = NULL
)
}
\arguments{
\item{ReplicationGroupId}{[required] The name of the Valkey or Redis OSS (cluster mode enabled) cluster
(replication group) on which the shards are to be configured.}

\item{NodeGroupCount}{[required] The number of node groups (shards) that results from the modification of
the shard configuration.}

\item{ApplyImmediately}{[required] Indicates that the shard reconfiguration process begins immediately. At
present, the only permitted value for this parameter is \code{true}.

Value: true}

\item{ReshardingConfiguration}{Specifies the preferred availability zones for each node group in the
cluster. If the value of \code{NodeGroupCount} is greater than the current
number of node groups (shards), you can use this parameter to specify
the preferred availability zones of the cluster's shards. If you omit
this parameter ElastiCache selects availability zones for you.

You can specify this parameter only if the value of \code{NodeGroupCount} is
greater than the current number of node groups (shards).}

\item{NodeGroupsToRemove}{If the value of \code{NodeGroupCount} is less than the current number of node
groups (shards), then either \code{NodeGroupsToRemove} or
\code{NodeGroupsToRetain} is required. \code{NodeGroupsToRemove} is a list of
\code{NodeGroupId}s to remove from the cluster.

ElastiCache will attempt to remove all node groups listed by
\code{NodeGroupsToRemove} from the cluster.}

\item{NodeGroupsToRetain}{If the value of \code{NodeGroupCount} is less than the current number of node
groups (shards), then either \code{NodeGroupsToRemove} or
\code{NodeGroupsToRetain} is required. \code{NodeGroupsToRetain} is a list of
\code{NodeGroupId}s to retain in the cluster.

ElastiCache will attempt to remove all node groups except those listed
by \code{NodeGroupsToRetain} from the cluster.}
}
\description{
Modifies a replication group's shards (node groups) by allowing you to add shards, remove shards, or rebalance the keyspaces among existing shards.

See \url{https://www.paws-r-sdk.com/docs/elasticache_modify_replication_group_shard_configuration/} for full documentation.
}
\keyword{internal}
