% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_modify_user}
\alias{elasticache_modify_user}
\title{Changes user password(s) and/or access string}
\usage{
elasticache_modify_user(
  UserId,
  AccessString = NULL,
  AppendAccessString = NULL,
  Passwords = NULL,
  NoPasswordRequired = NULL,
  AuthenticationMode = NULL
)
}
\arguments{
\item{UserId}{[required] The ID of the user.}

\item{AccessString}{Access permissions string used for this user.}

\item{AppendAccessString}{Adds additional user permissions to the access string.}

\item{Passwords}{The passwords belonging to the user. You are allowed up to two.}

\item{NoPasswordRequired}{Indicates no password is required for the user.}

\item{AuthenticationMode}{Specifies how to authenticate the user.}
}
\description{
Changes user password(s) and/or access string.

See \url{https://www.paws-r-sdk.com/docs/elasticache_modify_user/} for full documentation.
}
\keyword{internal}
