% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_modify_user_group}
\alias{elasticache_modify_user_group}
\title{Changes the list of users that belong to the user group}
\usage{
elasticache_modify_user_group(
  UserGroupId,
  UserIdsToAdd = NULL,
  UserIdsToRemove = NULL
)
}
\arguments{
\item{UserGroupId}{[required] The ID of the user group.}

\item{UserIdsToAdd}{The list of user IDs to add to the user group.}

\item{UserIdsToRemove}{The list of user IDs to remove from the user group.}
}
\description{
Changes the list of users that belong to the user group.

See \url{https://paws-r.github.io/docs/elasticache/modify_user_group.html} for full documentation.
}
\keyword{internal}
