% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_purchase_reserved_cache_nodes_offering}
\alias{elasticache_purchase_reserved_cache_nodes_offering}
\title{Allows you to purchase a reserved cache node offering}
\usage{
elasticache_purchase_reserved_cache_nodes_offering(
  ReservedCacheNodesOfferingId, ReservedCacheNodeId, CacheNodeCount, Tags)
}
\arguments{
\item{ReservedCacheNodesOfferingId}{[required] The ID of the reserved cache node offering to purchase.

Example: \verb{438012d3-4052-4cc7-b2e3-8d3372e0e706}}

\item{ReservedCacheNodeId}{A customer-specified identifier to track this reservation.

The Reserved Cache Node ID is an unique customer-specified identifier to
track this reservation. If this parameter is not specified, ElastiCache
automatically generates an identifier for the reservation.

Example: myreservationID}

\item{CacheNodeCount}{The number of cache node instances to reserve.

Default: \code{1}}

\item{Tags}{A list of tags to be added to this resource. A tag is a key-value pair.
A tag key must be accompanied by a tag value, although null is accepted.}
}
\value{
A list with the following syntax:\preformatted{list(
  ReservedCacheNode = list(
    ReservedCacheNodeId = "string",
    ReservedCacheNodesOfferingId = "string",
    CacheNodeType = "string",
    StartTime = as.POSIXct(
      "2015-01-01"
    ),
    Duration = 123,
    FixedPrice = 123.0,
    UsagePrice = 123.0,
    CacheNodeCount = 123,
    ProductDescription = "string",
    OfferingType = "string",
    State = "string",
    RecurringCharges = list(
      list(
        RecurringChargeAmount = 123.0,
        RecurringChargeFrequency = "string"
      )
    ),
    ReservationARN = "string"
  )
)
}
}
\description{
Allows you to purchase a reserved cache node offering. Reserved nodes
are not eligible for cancellation and are non-refundable. For more
information, see \href{https://docs.aws.amazon.com/AmazonElastiCache/latest/red-ug/}{Managing Costs with Reserved Nodes} for
Redis or \href{https://docs.aws.amazon.com/AmazonElastiCache/latest/mem-ug/}{Managing Costs with Reserved Nodes} for
Memcached.
}
\section{Request syntax}{
\preformatted{svc$purchase_reserved_cache_nodes_offering(
  ReservedCacheNodesOfferingId = "string",
  ReservedCacheNodeId = "string",
  CacheNodeCount = 123,
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
