% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_rebalance_slots_in_global_replication_group}
\alias{elasticache_rebalance_slots_in_global_replication_group}
\title{Redistribute slots to ensure uniform distribution across existing shards
in the cluster}
\usage{
elasticache_rebalance_slots_in_global_replication_group(
  GlobalReplicationGroupId,
  ApplyImmediately
)
}
\arguments{
\item{GlobalReplicationGroupId}{[required] The name of the Global datastore}

\item{ApplyImmediately}{[required] If \code{True}, redistribution is applied immediately.}
}
\description{
Redistribute slots to ensure uniform distribution across existing shards in the cluster.

See \url{https://paws-r.github.io/docs/elasticache/rebalance_slots_in_global_replication_group.html} for full documentation.
}
\keyword{internal}
