% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyspaces_operations.R
\name{keyspaces_get_keyspace}
\alias{keyspaces_get_keyspace}
\title{Returns the name and the Amazon Resource Name (ARN) of the specified
table}
\usage{
keyspaces_get_keyspace(keyspaceName)
}
\arguments{
\item{keyspaceName}{[required] The name of the keyspace.}
}
\description{
Returns the name and the Amazon Resource Name (ARN) of the specified table.

See \url{https://paws-r.github.io/docs/keyspaces/get_keyspace.html} for full documentation.
}
\keyword{internal}
