% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_create_lake_formation_opt_in}
\alias{lakeformation_create_lake_formation_opt_in}
\title{Enforce Lake Formation permissions for the given databases, tables, and
principals}
\usage{
lakeformation_create_lake_formation_opt_in(
  Principal,
  Resource,
  Condition = NULL
)
}
\arguments{
\item{Principal}{[required]}

\item{Resource}{[required]}

\item{Condition}{}
}
\description{
Enforce Lake Formation permissions for the given databases, tables, and principals.

See \url{https://www.paws-r-sdk.com/docs/lakeformation_create_lake_formation_opt_in/} for full documentation.
}
\keyword{internal}
