% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_delete_lake_formation_opt_in}
\alias{lakeformation_delete_lake_formation_opt_in}
\title{Remove the Lake Formation permissions enforcement of the given
databases, tables, and principals}
\usage{
lakeformation_delete_lake_formation_opt_in(
  Principal,
  Resource,
  Condition = NULL
)
}
\arguments{
\item{Principal}{[required]}

\item{Resource}{[required]}

\item{Condition}{}
}
\description{
Remove the Lake Formation permissions enforcement of the given databases, tables, and principals.

See \url{https://www.paws-r-sdk.com/docs/lakeformation_delete_lake_formation_opt_in/} for full documentation.
}
\keyword{internal}
