% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_delete_lf_tag}
\alias{lakeformation_delete_lf_tag}
\title{Deletes an LF-tag by its key name}
\usage{
lakeformation_delete_lf_tag(CatalogId = NULL, TagKey)
}
\arguments{
\item{CatalogId}{The identifier for the Data Catalog. By default, the account ID. The
Data Catalog is the persistent metadata store. It contains database
definitions, table definitions, and other control information to manage
your Lake Formation environment.}

\item{TagKey}{[required] The key-name for the LF-tag to delete.}
}
\description{
Deletes an LF-tag by its key name. The operation fails if the specified tag key doesn't exist. When you delete an LF-Tag:

See \url{https://www.paws-r-sdk.com/docs/lakeformation_delete_lf_tag/} for full documentation.
}
\keyword{internal}
