% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeformation_operations.R
\name{lakeformation_update_data_cells_filter}
\alias{lakeformation_update_data_cells_filter}
\title{Updates a data cell filter}
\usage{
lakeformation_update_data_cells_filter(TableData)
}
\arguments{
\item{TableData}{[required] A \code{DataCellsFilter} structure containing information about the data
cells filter.}
}
\description{
Updates a data cell filter.

See \url{https://www.paws-r-sdk.com/docs/lakeformation_update_data_cells_filter/} for full documentation.
}
\keyword{internal}
