% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_failover_shard}
\alias{memorydb_failover_shard}
\title{Used to failover a shard}
\usage{
memorydb_failover_shard(ClusterName, ShardName)
}
\arguments{
\item{ClusterName}{[required] The cluster being failed over}

\item{ShardName}{[required] The name of the shard}
}
\description{
Used to failover a shard. This API is designed for testing the behavior of your application in case of MemoryDB failover. It is not designed to be used as a production-level tool for initiating a failover to overcome a problem you may have with the cluster. Moreover, in certain conditions such as large scale operational events, Amazon may block this API.

See \url{https://www.paws-r-sdk.com/docs/memorydb_failover_shard/} for full documentation.
}
\keyword{internal}
