% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorydb_operations.R
\name{memorydb_tag_resource}
\alias{memorydb_tag_resource}
\title{A tag is a key-value pair where the key and value are case-sensitive}
\usage{
memorydb_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource to which the tags are to
be added}

\item{Tags}{[required] A list of tags to be added to this resource. A tag is a key-value pair.
A tag key must be accompanied by a tag value, although null is accepted.}
}
\description{
A tag is a key-value pair where the key and value are case-sensitive. You can use tags to categorize and track all your MemoryDB resources. When you add or remove tags on clusters, those actions will be replicated to all nodes in the cluster. For more information, see \href{https://docs.aws.amazon.com/memorydb/latest/devguide/iam.resourcelevelpermissions.html}{Resource-level permissions}.

See \url{https://www.paws-r-sdk.com/docs/memorydb_tag_resource/} for full documentation.
}
\keyword{internal}
