% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_operations.R
\name{neptune_create_db_cluster_parameter_group}
\alias{neptune_create_db_cluster_parameter_group}
\title{Creates a new DB cluster parameter group}
\usage{
neptune_create_db_cluster_parameter_group(
  DBClusterParameterGroupName,
  DBParameterGroupFamily,
  Description,
  Tags = NULL
)
}
\arguments{
\item{DBClusterParameterGroupName}{[required] The name of the DB cluster parameter group.

Constraints:
\itemize{
\item Must match the name of an existing DBClusterParameterGroup.
}

This value is stored as a lowercase string.}

\item{DBParameterGroupFamily}{[required] The DB cluster parameter group family name. A DB cluster parameter group
can be associated with one and only one DB cluster parameter group
family, and can be applied only to a DB cluster running a database
engine and engine version compatible with that DB cluster parameter
group family.}

\item{Description}{[required] The description for the DB cluster parameter group.}

\item{Tags}{The tags to be assigned to the new DB cluster parameter group.}
}
\description{
Creates a new DB cluster parameter group.

See \url{https://www.paws-r-sdk.com/docs/neptune_create_db_cluster_parameter_group/} for full documentation.
}
\keyword{internal}
