% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_operations.R
\name{neptune_describe_event_categories}
\alias{neptune_describe_event_categories}
\title{Displays a list of categories for all event source types, or, if
specified, for a specified source type}
\usage{
neptune_describe_event_categories(SourceType, Filters)
}
\arguments{
\item{SourceType}{The type of source that is generating the events.

Valid values: db-instance | db-parameter-group | db-security-group |
db-snapshot}

\item{Filters}{This parameter is not currently supported.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  EventCategoriesMapList = list(
    list(
      SourceType = "string",
      EventCategories = list(
        "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Displays a list of categories for all event source types, or, if
specified, for a specified source type.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_event_categories(
  SourceType = "string",
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
