% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_operations.R
\name{neptune_failover_db_cluster}
\alias{neptune_failover_db_cluster}
\title{Forces a failover for a DB cluster}
\usage{
neptune_failover_db_cluster(
  DBClusterIdentifier = NULL,
  TargetDBInstanceIdentifier = NULL
)
}
\arguments{
\item{DBClusterIdentifier}{A DB cluster identifier to force a failover for. This parameter is not
case-sensitive.

Constraints:
\itemize{
\item Must match the identifier of an existing DBCluster.
}}

\item{TargetDBInstanceIdentifier}{The name of the instance to promote to the primary instance.

You must specify the instance identifier for an Read Replica in the DB
cluster. For example, \code{mydbcluster-replica1}.}
}
\description{
Forces a failover for a DB cluster.

See \url{https://www.paws-r-sdk.com/docs/neptune_failover_db_cluster/} for full documentation.
}
\keyword{internal}
