% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_operations.R
\name{neptune_list_tags_for_resource}
\alias{neptune_list_tags_for_resource}
\title{Lists all tags on an Amazon Neptune resource}
\usage{
neptune_list_tags_for_resource(ResourceName, Filters = NULL)
}
\arguments{
\item{ResourceName}{[required] The Amazon Neptune resource with tags to be listed. This value is an
Amazon Resource Name (ARN). For information about creating an ARN, see
\href{https://docs.aws.amazon.com/neptune/latest/userguide/#tagging.ARN.Constructing}{Constructing an Amazon Resource Name (ARN)}.}

\item{Filters}{This parameter is not currently supported.}
}
\description{
Lists all tags on an Amazon Neptune resource.

See \url{https://www.paws-r-sdk.com/docs/neptune_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
