% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_operations.R
\name{neptune_start_db_cluster}
\alias{neptune_start_db_cluster}
\title{Starts an Amazon Neptune DB cluster that was stopped using the AWS
console, the AWS CLI stop-db-cluster command, or the StopDBCluster API}
\usage{
neptune_start_db_cluster(DBClusterIdentifier)
}
\arguments{
\item{DBClusterIdentifier}{[required] The DB cluster identifier of the Neptune DB cluster to be started. This
parameter is stored as a lowercase string.}
}
\description{
Starts an Amazon Neptune DB cluster that was stopped using the AWS
console, the AWS CLI stop-db-cluster command, or the StopDBCluster API.
}
\section{Request syntax}{
\preformatted{svc$start_db_cluster(
  DBClusterIdentifier = "string"
)
}
}

\keyword{internal}
