% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qldb_operations.R
\name{qldb_describe_ledger}
\alias{qldb_describe_ledger}
\title{Returns information about a ledger, including its state, permissions
mode, encryption at rest settings, and when it was created}
\usage{
qldb_describe_ledger(Name)
}
\arguments{
\item{Name}{[required] The name of the ledger that you want to describe.}
}
\description{
Returns information about a ledger, including its state, permissions mode, encryption at rest settings, and when it was created.

See \url{https://www.paws-r-sdk.com/docs/qldb_describe_ledger/} for full documentation.
}
\keyword{internal}
