% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qldb_operations.R
\name{qldb_list_tags_for_resource}
\alias{qldb_list_tags_for_resource}
\title{Returns all tags for a specified Amazon QLDB resource}
\usage{
qldb_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for which to list the tags. For example:

\code{arn:aws:qldb:us-east-1:123456789012:ledger/exampleLedger}}
}
\description{
Returns all tags for a specified Amazon QLDB resource.

See \url{https://www.paws-r-sdk.com/docs/qldb_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
