% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_cancel_export_task}
\alias{rds_cancel_export_task}
\title{Cancels an export task in progress that is exporting a snapshot or
cluster to Amazon S3}
\usage{
rds_cancel_export_task(ExportTaskIdentifier)
}
\arguments{
\item{ExportTaskIdentifier}{[required] The identifier of the snapshot or cluster export task to cancel.}
}
\description{
Cancels an export task in progress that is exporting a snapshot or cluster to Amazon S3. Any data that has already been written to the S3 bucket isn't removed.

See \url{https://www.paws-r-sdk.com/docs/rds_cancel_export_task/} for full documentation.
}
\keyword{internal}
