% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_create_db_security_group}
\alias{rds_create_db_security_group}
\title{Creates a new DB security group}
\usage{
rds_create_db_security_group(
  DBSecurityGroupName,
  DBSecurityGroupDescription,
  Tags = NULL
)
}
\arguments{
\item{DBSecurityGroupName}{[required] The name for the DB security group. This value is stored as a lowercase
string.

Constraints:
\itemize{
\item Must be 1 to 255 letters, numbers, or hyphens.
\item First character must be a letter
\item Can't end with a hyphen or contain two consecutive hyphens
\item Must not be "Default"
}

Example: \code{mysecuritygroup}}

\item{DBSecurityGroupDescription}{[required] The description for the DB security group.}

\item{Tags}{Tags to assign to the DB security group.}
}
\description{
Creates a new DB security group. DB security groups control access to a DB instance.

See \url{https://www.paws-r-sdk.com/docs/rds_create_db_security_group/} for full documentation.
}
\keyword{internal}
