% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_delete_blue_green_deployment}
\alias{rds_delete_blue_green_deployment}
\title{Deletes a blue/green deployment}
\usage{
rds_delete_blue_green_deployment(
  BlueGreenDeploymentIdentifier,
  DeleteTarget = NULL
)
}
\arguments{
\item{BlueGreenDeploymentIdentifier}{[required] The blue/green deployment identifier of the deployment to be deleted.
This parameter isn't case-sensitive.

Constraints:
\itemize{
\item Must match an existing blue/green deployment identifier.
}}

\item{DeleteTarget}{A value that indicates whether to delete the resources in the green
environment. You can't specify this option if the blue/green deployment
\href{https://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_BlueGreenDeployment.html}{status}
is \code{SWITCHOVER_COMPLETED}.}
}
\description{
Deletes a blue/green deployment.

See \url{https://paws-r.github.io/docs/rds/delete_blue_green_deployment.html} for full documentation.
}
\keyword{internal}
