% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_delete_custom_db_engine_version}
\alias{rds_delete_custom_db_engine_version}
\title{Deletes a custom engine version}
\usage{
rds_delete_custom_db_engine_version(Engine, EngineVersion)
}
\arguments{
\item{Engine}{[required] The database engine. The only supported engines are \code{custom-oracle-ee}
and \code{custom-oracle-ee-cdb}.}

\item{EngineVersion}{[required] The custom engine version (CEV) for your DB instance. This option is
required for RDS Custom, but optional for Amazon RDS. The combination of
\code{Engine} and \code{EngineVersion} is unique per customer per Amazon Web
Services Region.}
}
\description{
Deletes a custom engine version. To run this command, make sure you meet the following prerequisites:

See \url{https://paws-r.github.io/docs/rds/delete_custom_db_engine_version.html} for full documentation.
}
\keyword{internal}
