% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_delete_db_proxy}
\alias{rds_delete_db_proxy}
\title{This is prerelease documentation for the RDS Database Proxy feature in
preview release}
\usage{
rds_delete_db_proxy(DBProxyName)
}
\arguments{
\item{DBProxyName}{[required] The name of the DB proxy to delete.}
}
\description{
This is prerelease documentation for the RDS Database Proxy feature in
preview release. It is subject to change.
}
\details{
Deletes an existing proxy.
}
\section{Request syntax}{
\preformatted{svc$delete_db_proxy(
  DBProxyName = "string"
)
}
}

\keyword{internal}
