% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_delete_db_proxy}
\alias{rds_delete_db_proxy}
\title{Deletes an existing DB proxy}
\usage{
rds_delete_db_proxy(DBProxyName)
}
\arguments{
\item{DBProxyName}{[required] The name of the DB proxy to delete.}
}
\description{
Deletes an existing DB proxy.

See \url{https://paws-r.github.io/docs/rds/delete_db_proxy.html} for full documentation.
}
\keyword{internal}
