% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_delete_db_proxy_endpoint}
\alias{rds_delete_db_proxy_endpoint}
\title{Deletes a DBProxyEndpoint}
\usage{
rds_delete_db_proxy_endpoint(DBProxyEndpointName)
}
\arguments{
\item{DBProxyEndpointName}{[required] The name of the DB proxy endpoint to delete.}
}
\description{
Deletes a \code{DBProxyEndpoint}. Doing so removes the ability to access the DB proxy using the endpoint that you defined. The endpoint that you delete might have provided capabilities such as read/write or read-only operations, or using a different VPC than the DB proxy's default VPC.

See \url{https://www.paws-r-sdk.com/docs/rds_delete_db_proxy_endpoint/} for full documentation.
}
\keyword{internal}
