% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_delete_event_subscription}
\alias{rds_delete_event_subscription}
\title{Deletes an RDS event notification subscription}
\usage{
rds_delete_event_subscription(SubscriptionName)
}
\arguments{
\item{SubscriptionName}{[required] The name of the RDS event notification subscription you want to delete.}
}
\description{
Deletes an RDS event notification subscription.

See \url{https://paws-r.github.io/docs/rds/delete_event_subscription.html} for full documentation.
}
\keyword{internal}
