% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_delete_option_group}
\alias{rds_delete_option_group}
\title{Deletes an existing option group}
\usage{
rds_delete_option_group(OptionGroupName)
}
\arguments{
\item{OptionGroupName}{[required] The name of the option group to be deleted.

You can't delete default option groups.}
}
\description{
Deletes an existing option group.

See \url{https://www.paws-r-sdk.com/docs/rds_delete_option_group/} for full documentation.
}
\keyword{internal}
