% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_delete_option_group}
\alias{rds_delete_option_group}
\title{Deletes an existing option group}
\usage{
rds_delete_option_group(OptionGroupName)
}
\arguments{
\item{OptionGroupName}{[required] The name of the option group to be deleted.

You can't delete default option groups.}
}
\value{
An empty list.
}
\description{
Deletes an existing option group.
}
\section{Request syntax}{
\preformatted{svc$delete_option_group(
  OptionGroupName = "string"
)
}
}

\keyword{internal}
