% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_deregister_db_proxy_targets}
\alias{rds_deregister_db_proxy_targets}
\title{Remove the association between one or more DBProxyTarget data structures
and a DBProxyTargetGroup}
\usage{
rds_deregister_db_proxy_targets(DBProxyName, TargetGroupName,
  DBInstanceIdentifiers, DBClusterIdentifiers)
}
\arguments{
\item{DBProxyName}{[required] The identifier of the \code{DBProxy} that is associated with the
\code{DBProxyTargetGroup}.}

\item{TargetGroupName}{The identifier of the \code{DBProxyTargetGroup}.}

\item{DBInstanceIdentifiers}{One or more DB instance identifiers.}

\item{DBClusterIdentifiers}{One or more DB cluster identifiers.}
}
\value{
An empty list.
}
\description{
Remove the association between one or more \code{DBProxyTarget} data
structures and a \code{DBProxyTargetGroup}.
}
\section{Request syntax}{
\preformatted{svc$deregister_db_proxy_targets(
  DBProxyName = "string",
  TargetGroupName = "string",
  DBInstanceIdentifiers = list(
    "string"
  ),
  DBClusterIdentifiers = list(
    "string"
  )
)
}
}

\keyword{internal}
