% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_describe_db_engine_versions}
\alias{rds_describe_db_engine_versions}
\title{Returns a list of the available DB engines}
\usage{
rds_describe_db_engine_versions(
  Engine = NULL,
  EngineVersion = NULL,
  DBParameterGroupFamily = NULL,
  Filters = NULL,
  MaxRecords = NULL,
  Marker = NULL,
  DefaultOnly = NULL,
  ListSupportedCharacterSets = NULL,
  ListSupportedTimezones = NULL,
  IncludeAll = NULL
)
}
\arguments{
\item{Engine}{The database engine to return.

Valid Values:
\itemize{
\item \code{aurora} (for MySQL 5.6-compatible Aurora)
\item \code{aurora-mysql} (for MySQL 5.7-compatible and MySQL 8.0-compatible
Aurora)
\item \code{aurora-postgresql}
\item \code{mariadb}
\item \code{mysql}
\item \code{oracle-ee}
\item \code{oracle-ee-cdb}
\item \code{oracle-se2}
\item \code{oracle-se2-cdb}
\item \code{postgres}
\item \code{sqlserver-ee}
\item \code{sqlserver-se}
\item \code{sqlserver-ex}
\item \code{sqlserver-web}
}}

\item{EngineVersion}{The database engine version to return.

Example: \verb{5.1.49}}

\item{DBParameterGroupFamily}{The name of a specific DB parameter group family to return details for.

Constraints:
\itemize{
\item If supplied, must match an existing DBParameterGroupFamily.
}}

\item{Filters}{A filter that specifies one or more DB engine versions to describe.

Supported filters:
\itemize{
\item \code{db-parameter-group-family} - Accepts parameter groups family names.
The results list only includes information about the DB engine
versions for these parameter group families.
\item \code{engine} - Accepts engine names. The results list only includes
information about the DB engine versions for these engines.
\item \code{engine-mode} - Accepts DB engine modes. The results list only
includes information about the DB engine versions for these engine
modes. Valid DB engine modes are the following:
\itemize{
\item \code{global}
\item \code{multimaster}
\item \code{parallelquery}
\item \code{provisioned}
\item \code{serverless}
}
\item \code{engine-version} - Accepts engine versions. The results list only
includes information about the DB engine versions for these engine
versions.
\item \code{status} - Accepts engine version statuses. The results list only
includes information about the DB engine versions for these
statuses. Valid statuses are the following:
\itemize{
\item \code{available}
\item \code{deprecated}
}
}}

\item{MaxRecords}{The maximum number of records to include in the response. If more than
the \code{MaxRecords} value is available, a pagination token called a marker
is included in the response so you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.}

\item{Marker}{An optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}

\item{DefaultOnly}{A value that indicates whether only the default version of the specified
engine or engine and major version combination is returned.}

\item{ListSupportedCharacterSets}{A value that indicates whether to list the supported character sets for
each engine version.

If this parameter is enabled and the requested engine supports the
\code{CharacterSetName} parameter for
\code{\link[=rds_create_db_instance]{create_db_instance}}, the response includes a
list of supported character sets for each engine version.

For RDS Custom, the default is not to list supported character sets. If
you set \code{ListSupportedCharacterSets} to \code{true}, RDS Custom returns no
results.}

\item{ListSupportedTimezones}{A value that indicates whether to list the supported time zones for each
engine version.

If this parameter is enabled and the requested engine supports the
\code{TimeZone} parameter for \code{\link[=rds_create_db_instance]{create_db_instance}},
the response includes a list of supported time zones for each engine
version.

For RDS Custom, the default is not to list supported time zones. If you
set \code{ListSupportedTimezones} to \code{true}, RDS Custom returns no results.}

\item{IncludeAll}{A value that indicates whether to include engine versions that aren't
available in the list. The default is to list only available engine
versions.}
}
\description{
Returns a list of the available DB engines.

See \url{https://paws-r.github.io/docs/rds/describe_db_engine_versions.html} for full documentation.
}
\keyword{internal}
