% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_describe_db_major_engine_versions}
\alias{rds_describe_db_major_engine_versions}
\title{Describes the properties of specific major versions of DB engines}
\usage{
rds_describe_db_major_engine_versions(
  Engine = NULL,
  MajorEngineVersion = NULL,
  Marker = NULL,
  MaxRecords = NULL
)
}
\arguments{
\item{Engine}{The database engine to return major version details for.

Valid Values:
\itemize{
\item \code{aurora-mysql}
\item \code{aurora-postgresql}
\item \code{custom-sqlserver-ee}
\item \code{custom-sqlserver-se}
\item \code{custom-sqlserver-web}
\item \code{db2-ae}
\item \code{db2-se}
\item \code{mariadb}
\item \code{mysql}
\item \code{oracle-ee}
\item \code{oracle-ee-cdb}
\item \code{oracle-se2}
\item \code{oracle-se2-cdb}
\item \code{postgres}
\item \code{sqlserver-ee}
\item \code{sqlserver-se}
\item \code{sqlserver-ex}
\item \code{sqlserver-web}
}}

\item{MajorEngineVersion}{A specific database major engine version to return details for.

Example: \code{8.4}}

\item{Marker}{An optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}

\item{MaxRecords}{The maximum number of records to include in the response. If more than
the \code{MaxRecords} value is available, a pagination token called a marker
is included in the response so you can retrieve the remaining results.

Default: 100}
}
\description{
Describes the properties of specific major versions of DB engines.

See \url{https://www.paws-r-sdk.com/docs/rds_describe_db_major_engine_versions/} for full documentation.
}
\keyword{internal}
