% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_describe_db_proxies}
\alias{rds_describe_db_proxies}
\title{Returns information about DB proxies}
\usage{
rds_describe_db_proxies(DBProxyName, Filters, Marker, MaxRecords)
}
\arguments{
\item{DBProxyName}{The name of the DB proxy.}

\item{Filters}{This parameter is not currently supported.}

\item{Marker}{An optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
called a marker is included in the response so that the remaining
results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}
}
\value{
A list with the following syntax:\preformatted{list(
  DBProxies = list(
    list(
      DBProxyName = "string",
      DBProxyArn = "string",
      Status = "available"|"modifying"|"incompatible-network"|"insufficient-resource-limits"|"creating"|"deleting"|"suspended"|"suspending"|"reactivating",
      EngineFamily = "string",
      VpcSecurityGroupIds = list(
        "string"
      ),
      VpcSubnetIds = list(
        "string"
      ),
      Auth = list(
        list(
          Description = "string",
          UserName = "string",
          AuthScheme = "SECRETS",
          SecretArn = "string",
          IAMAuth = "DISABLED"|"REQUIRED"
        )
      ),
      RoleArn = "string",
      Endpoint = "string",
      RequireTLS = TRUE|FALSE,
      IdleClientTimeout = 123,
      DebugLogging = TRUE|FALSE,
      CreatedDate = as.POSIXct(
        "2015-01-01"
      ),
      UpdatedDate = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  Marker = "string"
)
}
}
\description{
Returns information about DB proxies.
}
\section{Request syntax}{
\preformatted{svc$describe_db_proxies(
  DBProxyName = "string",
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  Marker = "string",
  MaxRecords = 123
)
}
}

\keyword{internal}
