% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_describe_db_proxy_targets}
\alias{rds_describe_db_proxy_targets}
\title{Returns information about DBProxyTarget objects}
\usage{
rds_describe_db_proxy_targets(
  DBProxyName,
  TargetGroupName = NULL,
  Filters = NULL,
  Marker = NULL,
  MaxRecords = NULL
)
}
\arguments{
\item{DBProxyName}{[required] The identifier of the \code{DBProxyTarget} to describe.}

\item{TargetGroupName}{The identifier of the \code{DBProxyTargetGroup} to describe.}

\item{Filters}{This parameter is not currently supported.}

\item{Marker}{An optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
called a marker is included in the response so that the remaining
results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}
}
\description{
Returns information about \code{DBProxyTarget} objects. This API supports pagination.

See \url{https://paws-r.github.io/docs/rds/describe_db_proxy_targets.html} for full documentation.
}
\keyword{internal}
