% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_describe_source_regions}
\alias{rds_describe_source_regions}
\title{Returns a list of the source AWS Regions where the current AWS Region
can create a read replica, copy a DB snapshot from, or replicate
automated backups from}
\usage{
rds_describe_source_regions(RegionName, MaxRecords, Marker, Filters)
}
\arguments{
\item{RegionName}{The source AWS Region name. For example, \code{us-east-1}.

Constraints:
\itemize{
\item Must specify a valid AWS Region name.
}}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
called a marker is included in the response so you can retrieve the
remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.}

\item{Marker}{An optional pagination token provided by a previous
\code{\link[=rds_describe_source_regions]{describe_source_regions}} request. If
this parameter is specified, the response includes only records beyond
the marker, up to the value specified by \code{MaxRecords}.}

\item{Filters}{This parameter isn't currently supported.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Marker = "string",
  SourceRegions = list(
    list(
      RegionName = "string",
      Endpoint = "string",
      Status = "string",
      SupportsDBInstanceAutomatedBackupsReplication = TRUE|FALSE
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns a list of the source AWS Regions where the current AWS Region
can create a read replica, copy a DB snapshot from, or replicate
automated backups from. This API action supports pagination.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_source_regions(
  RegionName = "string",
  MaxRecords = 123,
  Marker = "string",
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
