% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_describe_valid_db_instance_modifications}
\alias{rds_describe_valid_db_instance_modifications}
\title{You can call DescribeValidDBInstanceModifications to learn what
modifications you can make to your DB instance}
\usage{
rds_describe_valid_db_instance_modifications(DBInstanceIdentifier)
}
\arguments{
\item{DBInstanceIdentifier}{[required] The customer identifier or the ARN of your DB instance.}
}
\description{
You can call \code{DescribeValidDBInstanceModifications} to learn what
modifications you can make to your DB instance. You can use this
information when you call \code{ModifyDBInstance}.
}
\section{Request syntax}{
\preformatted{svc$describe_valid_db_instance_modifications(
  DBInstanceIdentifier = "string"
)
}
}

\keyword{internal}
