% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_modify_db_cluster_snapshot_attribute}
\alias{rds_modify_db_cluster_snapshot_attribute}
\title{Adds an attribute and values to, or removes an attribute and values
from, a manual DB cluster snapshot}
\usage{
rds_modify_db_cluster_snapshot_attribute(
  DBClusterSnapshotIdentifier,
  AttributeName,
  ValuesToAdd = NULL,
  ValuesToRemove = NULL
)
}
\arguments{
\item{DBClusterSnapshotIdentifier}{[required] The identifier for the DB cluster snapshot to modify the attributes for.}

\item{AttributeName}{[required] The name of the DB cluster snapshot attribute to modify.

To manage authorization for other Amazon Web Services accounts to copy
or restore a manual DB cluster snapshot, set this value to \code{restore}.

To view the list of attributes available to modify, use the
\code{\link[=rds_describe_db_cluster_snapshot_attributes]{describe_db_cluster_snapshot_attributes}}
API operation.}

\item{ValuesToAdd}{A list of DB cluster snapshot attributes to add to the attribute
specified by \code{AttributeName}.

To authorize other Amazon Web Services accounts to copy or restore a
manual DB cluster snapshot, set this list to include one or more Amazon
Web Services account IDs, or \code{all} to make the manual DB cluster
snapshot restorable by any Amazon Web Services account. Do not add the
\code{all} value for any manual DB cluster snapshots that contain private
information that you don't want available to all Amazon Web Services
accounts.}

\item{ValuesToRemove}{A list of DB cluster snapshot attributes to remove from the attribute
specified by \code{AttributeName}.

To remove authorization for other Amazon Web Services accounts to copy
or restore a manual DB cluster snapshot, set this list to include one or
more Amazon Web Services account identifiers, or \code{all} to remove
authorization for any Amazon Web Services account to copy or restore the
DB cluster snapshot. If you specify \code{all}, an Amazon Web Services
account whose account ID is explicitly added to the \code{restore} attribute
can still copy or restore a manual DB cluster snapshot.}
}
\description{
Adds an attribute and values to, or removes an attribute and values from, a manual DB cluster snapshot.

See \url{https://paws-r.github.io/docs/rds/modify_db_cluster_snapshot_attribute.html} for full documentation.
}
\keyword{internal}
