% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_modify_db_proxy_endpoint}
\alias{rds_modify_db_proxy_endpoint}
\title{Changes the settings for an existing DB proxy endpoint}
\usage{
rds_modify_db_proxy_endpoint(DBProxyEndpointName,
  NewDBProxyEndpointName, VpcSecurityGroupIds)
}
\arguments{
\item{DBProxyEndpointName}{[required] The name of the DB proxy sociated with the DB proxy endpoint that you
want to modify.}

\item{NewDBProxyEndpointName}{The new identifier for the \code{DBProxyEndpoint}. An identifier must begin
with a letter and must contain only ASCII letters, digits, and hyphens;
it can't end with a hyphen or contain two consecutive hyphens.}

\item{VpcSecurityGroupIds}{The VPC security group IDs for the DB proxy endpoint. When the DB proxy
endpoint uses a different VPC than the original proxy, you also specify
a different set of security group IDs than for the original proxy.}
}
\value{
A list with the following syntax:\preformatted{list(
  DBProxyEndpoint = list(
    DBProxyEndpointName = "string",
    DBProxyEndpointArn = "string",
    DBProxyName = "string",
    Status = "available"|"modifying"|"incompatible-network"|"insufficient-resource-limits"|"creating"|"deleting",
    VpcId = "string",
    VpcSecurityGroupIds = list(
      "string"
    ),
    VpcSubnetIds = list(
      "string"
    ),
    Endpoint = "string",
    CreatedDate = as.POSIXct(
      "2015-01-01"
    ),
    TargetRole = "READ_WRITE"|"READ_ONLY",
    IsDefault = TRUE|FALSE
  )
)
}
}
\description{
Changes the settings for an existing DB proxy endpoint.
}
\section{Request syntax}{
\preformatted{svc$modify_db_proxy_endpoint(
  DBProxyEndpointName = "string",
  NewDBProxyEndpointName = "string",
  VpcSecurityGroupIds = list(
    "string"
  )
)
}
}

\keyword{internal}
