% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_modify_db_shard_group}
\alias{rds_modify_db_shard_group}
\title{Modifies the settings of an Aurora Limitless Database DB shard group}
\usage{
rds_modify_db_shard_group(DBShardGroupIdentifier, MaxACU = NULL)
}
\arguments{
\item{DBShardGroupIdentifier}{[required] The name of the DB shard group to modify.}

\item{MaxACU}{The maximum capacity of the DB shard group in Aurora capacity units
(ACUs).}
}
\description{
Modifies the settings of an Aurora Limitless Database DB shard group. You can change one or more settings by specifying these parameters and the new values in the request.

See \url{https://www.paws-r-sdk.com/docs/rds_modify_db_shard_group/} for full documentation.
}
\keyword{internal}
