% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_reboot_db_instance}
\alias{rds_reboot_db_instance}
\title{You might need to reboot your DB instance, usually for maintenance
reasons}
\usage{
rds_reboot_db_instance(DBInstanceIdentifier, ForceFailover = NULL)
}
\arguments{
\item{DBInstanceIdentifier}{[required] The DB instance identifier. This parameter is stored as a lowercase
string.

Constraints:
\itemize{
\item Must match the identifier of an existing DBInstance.
}}

\item{ForceFailover}{A value that indicates whether the reboot is conducted through a
Multi-AZ failover.

Constraint: You can't enable force failover if the instance isn't
configured for Multi-AZ.}
}
\description{
You might need to reboot your DB instance, usually for maintenance reasons. For example, if you make certain modifications, or if you change the DB parameter group associated with the DB instance, you must reboot the instance for the changes to take effect.

See \url{https://www.paws-r-sdk.com/docs/rds_reboot_db_instance/} for full documentation.
}
\keyword{internal}
