% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_remove_source_identifier_from_subscription}
\alias{rds_remove_source_identifier_from_subscription}
\title{Removes a source identifier from an existing RDS event notification
subscription}
\usage{
rds_remove_source_identifier_from_subscription(
  SubscriptionName,
  SourceIdentifier
)
}
\arguments{
\item{SubscriptionName}{[required] The name of the RDS event notification subscription you want to remove a
source identifier from.}

\item{SourceIdentifier}{[required] The source identifier to be removed from the subscription, such as the
\strong{DB instance identifier} for a DB instance or the name of a security
group.}
}
\description{
Removes a source identifier from an existing RDS event notification subscription.

See \url{https://www.paws-r-sdk.com/docs/rds_remove_source_identifier_from_subscription/} for full documentation.
}
\keyword{internal}
