% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_start_db_instance}
\alias{rds_start_db_instance}
\title{Starts an Amazon RDS DB instance that was stopped using the Amazon Web
Services console, the stop-db-instance CLI command, or the
StopDBInstance action}
\usage{
rds_start_db_instance(DBInstanceIdentifier)
}
\arguments{
\item{DBInstanceIdentifier}{[required] The user-supplied instance identifier.}
}
\description{
Starts an Amazon RDS DB instance that was stopped using the Amazon Web Services console, the stop-db-instance CLI command, or the StopDBInstance action.

See \url{https://paws-r.github.io/docs/rds/start_db_instance.html} for full documentation.
}
\keyword{internal}
