% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_start_db_instance}
\alias{rds_start_db_instance}
\title{Starts an Amazon RDS DB instance that was stopped using the AWS console,
the stop-db-instance AWS CLI command, or the StopDBInstance action}
\usage{
rds_start_db_instance(DBInstanceIdentifier)
}
\arguments{
\item{DBInstanceIdentifier}{[required] The user-supplied instance identifier.}
}
\description{
Starts an Amazon RDS DB instance that was stopped using the AWS console,
the stop-db-instance AWS CLI command, or the StopDBInstance action.

For more information, see \href{https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/USER_StartInstance.html}{Starting an Amazon RDS DB instance That Was Previously Stopped}
in the \emph{Amazon RDS User Guide.}

This command doesn't apply to Aurora MySQL and Aurora PostgreSQL. For
Aurora DB clusters, use \code{StartDBCluster} instead.
}
\section{Request syntax}{
\preformatted{svc$start_db_instance(
  DBInstanceIdentifier = "string"
)
}
}

\keyword{internal}
