% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_stop_db_instance_automated_backups_replication}
\alias{rds_stop_db_instance_automated_backups_replication}
\title{Stops automated backup replication for a DB instance}
\usage{
rds_stop_db_instance_automated_backups_replication(SourceDBInstanceArn)
}
\arguments{
\item{SourceDBInstanceArn}{[required] The Amazon Resource Name (ARN) of the source DB instance for which to
stop replicating automated backups, for example,
\code{arn:aws:rds:us-west-2:123456789012:db:mydatabase}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DBInstanceAutomatedBackup = list(
    DBInstanceArn = "string",
    DbiResourceId = "string",
    Region = "string",
    DBInstanceIdentifier = "string",
    RestoreWindow = list(
      EarliestTime = as.POSIXct(
        "2015-01-01"
      ),
      LatestTime = as.POSIXct(
        "2015-01-01"
      )
    ),
    AllocatedStorage = 123,
    Status = "string",
    Port = 123,
    AvailabilityZone = "string",
    VpcId = "string",
    InstanceCreateTime = as.POSIXct(
      "2015-01-01"
    ),
    MasterUsername = "string",
    Engine = "string",
    EngineVersion = "string",
    LicenseModel = "string",
    Iops = 123,
    OptionGroupName = "string",
    TdeCredentialArn = "string",
    Encrypted = TRUE|FALSE,
    StorageType = "string",
    KmsKeyId = "string",
    Timezone = "string",
    IAMDatabaseAuthenticationEnabled = TRUE|FALSE,
    BackupRetentionPeriod = 123,
    DBInstanceAutomatedBackupsArn = "string",
    DBInstanceAutomatedBackupsReplications = list(
      list(
        DBInstanceAutomatedBackupsArn = "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Stops automated backup replication for a DB instance.

For more information, see \href{https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/USER_ReplicateBackups.html}{Replicating Automated Backups to Another AWS Region}
in the \emph{Amazon RDS User Guide.}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$stop_db_instance_automated_backups_replication(
  SourceDBInstanceArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
