% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_operations.R
\name{rds_stop_db_instance_automated_backups_replication}
\alias{rds_stop_db_instance_automated_backups_replication}
\title{Stops automated backup replication for a DB instance}
\usage{
rds_stop_db_instance_automated_backups_replication(SourceDBInstanceArn)
}
\arguments{
\item{SourceDBInstanceArn}{[required] The Amazon Resource Name (ARN) of the source DB instance for which to
stop replicating automated backups, for example,
\code{arn:aws:rds:us-west-2:123456789012:db:mydatabase}.}
}
\description{
Stops automated backup replication for a DB instance.

See \url{https://paws-r.github.io/docs/rds/stop_db_instance_automated_backups_replication.html} for full documentation.
}
\keyword{internal}
