% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsdataservice_operations.R
\name{rdsdataservice_begin_transaction}
\alias{rdsdataservice_begin_transaction}
\title{Starts a SQL transaction}
\usage{
rdsdataservice_begin_transaction(
  resourceArn,
  secretArn,
  database = NULL,
  schema = NULL
)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the Aurora Serverless DB cluster.}

\item{secretArn}{[required] The name or ARN of the secret that enables access to the DB cluster.}

\item{database}{The name of the database.}

\item{schema}{The name of the database schema.}
}
\description{
Starts a SQL transaction.

See \url{https://paws-r.github.io/docs/rdsdataservice/begin_transaction.html} for full documentation.
}
\keyword{internal}
