% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsdataservice_operations.R
\name{rdsdataservice_rollback_transaction}
\alias{rdsdataservice_rollback_transaction}
\title{Performs a rollback of a transaction}
\usage{
rdsdataservice_rollback_transaction(resourceArn, secretArn, transactionId)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the Aurora Serverless DB cluster.}

\item{secretArn}{[required] The name or ARN of the secret that enables access to the DB cluster.}

\item{transactionId}{[required] The identifier of the transaction to roll back.}
}
\description{
Performs a rollback of a transaction. Rolling back a transaction cancels its changes.

See \url{https://www.paws-r-sdk.com/docs/rdsdataservice_rollback_transaction/} for full documentation.
}
\keyword{internal}
