% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_accept_reserved_node_exchange}
\alias{redshift_accept_reserved_node_exchange}
\title{Exchanges a DC1 Reserved Node for a DC2 Reserved Node with no changes to
the configuration (term, payment type, or number of nodes) and no
additional costs}
\usage{
redshift_accept_reserved_node_exchange(ReservedNodeId,
  TargetReservedNodeOfferingId)
}
\arguments{
\item{ReservedNodeId}{[required] A string representing the node identifier of the DC1 Reserved Node to be
exchanged.}

\item{TargetReservedNodeOfferingId}{[required] The unique identifier of the DC2 Reserved Node offering to be used for
the exchange. You can obtain the value for the parameter by calling
\code{\link[=redshift_get_reserved_node_exchange_offerings]{get_reserved_node_exchange_offerings}}}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ExchangedReservedNode = list(
    ReservedNodeId = "string",
    ReservedNodeOfferingId = "string",
    NodeType = "string",
    StartTime = as.POSIXct(
      "2015-01-01"
    ),
    Duration = 123,
    FixedPrice = 123.0,
    UsagePrice = 123.0,
    CurrencyCode = "string",
    NodeCount = 123,
    State = "string",
    OfferingType = "string",
    RecurringCharges = list(
      list(
        RecurringChargeAmount = 123.0,
        RecurringChargeFrequency = "string"
      )
    ),
    ReservedNodeOfferingType = "Regular"|"Upgradable"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Exchanges a DC1 Reserved Node for a DC2 Reserved Node with no changes to
the configuration (term, payment type, or number of nodes) and no
additional costs.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$accept_reserved_node_exchange(
  ReservedNodeId = "string",
  TargetReservedNodeOfferingId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
