% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_add_partner}
\alias{redshift_add_partner}
\title{Adds a partner integration to a cluster}
\usage{
redshift_add_partner(AccountId, ClusterIdentifier, DatabaseName, PartnerName)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID that owns the cluster.}

\item{ClusterIdentifier}{[required] The cluster identifier of the cluster that receives data from the
partner.}

\item{DatabaseName}{[required] The name of the database that receives data from the partner.}

\item{PartnerName}{[required] The name of the partner that is authorized to send data.}
}
\description{
Adds a partner integration to a cluster. This operation authorizes a partner to push status updates for the specified database. To complete the integration, you also set up the integration on the partner website.

See \url{https://www.paws-r-sdk.com/docs/redshift_add_partner/} for full documentation.
}
\keyword{internal}
