% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_authorize_endpoint_access}
\alias{redshift_authorize_endpoint_access}
\title{Grants access to a cluster}
\usage{
redshift_authorize_endpoint_access(ClusterIdentifier, Account, VpcIds)
}
\arguments{
\item{ClusterIdentifier}{The cluster identifier of the cluster to grant access to.}

\item{Account}{[required] The AWS account ID to grant access to.}

\item{VpcIds}{The virtual private cloud (VPC) identifiers to grant access to.}
}
\value{
A list with the following syntax:\preformatted{list(
  Grantor = "string",
  Grantee = "string",
  ClusterIdentifier = "string",
  AuthorizeTime = as.POSIXct(
    "2015-01-01"
  ),
  ClusterStatus = "string",
  Status = "Authorized"|"Revoking",
  AllowedAllVPCs = TRUE|FALSE,
  AllowedVPCs = list(
    "string"
  ),
  EndpointCount = 123
)
}
}
\description{
Grants access to a cluster.
}
\section{Request syntax}{
\preformatted{svc$authorize_endpoint_access(
  ClusterIdentifier = "string",
  Account = "string",
  VpcIds = list(
    "string"
  )
)
}
}

\keyword{internal}
