% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_batch_delete_cluster_snapshots}
\alias{redshift_batch_delete_cluster_snapshots}
\title{Deletes a set of cluster snapshots}
\usage{
redshift_batch_delete_cluster_snapshots(Identifiers)
}
\arguments{
\item{Identifiers}{[required] A list of identifiers for the snapshots that you want to delete.}
}
\description{
Deletes a set of cluster snapshots.

See \url{https://paws-r.github.io/docs/redshift/batch_delete_cluster_snapshots.html} for full documentation.
}
\keyword{internal}
