% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_delete_authentication_profile}
\alias{redshift_delete_authentication_profile}
\title{Deletes an authentication profile}
\usage{
redshift_delete_authentication_profile(AuthenticationProfileName)
}
\arguments{
\item{AuthenticationProfileName}{[required] The name of the authentication profile to delete.}
}
\description{
Deletes an authentication profile.

See \url{https://www.paws-r-sdk.com/docs/redshift_delete_authentication_profile/} for full documentation.
}
\keyword{internal}
