% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_delete_cluster_security_group}
\alias{redshift_delete_cluster_security_group}
\title{Deletes an Amazon Redshift security group}
\usage{
redshift_delete_cluster_security_group(ClusterSecurityGroupName)
}
\arguments{
\item{ClusterSecurityGroupName}{[required] The name of the cluster security group to be deleted.}
}
\description{
Deletes an Amazon Redshift security group.

See \url{https://paws-r.github.io/docs/redshift/delete_cluster_security_group.html} for full documentation.
}
\keyword{internal}
