% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_delete_cluster_snapshot}
\alias{redshift_delete_cluster_snapshot}
\title{Deletes the specified manual snapshot}
\usage{
redshift_delete_cluster_snapshot(
  SnapshotIdentifier,
  SnapshotClusterIdentifier = NULL
)
}
\arguments{
\item{SnapshotIdentifier}{[required] The unique identifier of the manual snapshot to be deleted.

Constraints: Must be the name of an existing snapshot that is in the
\code{available}, \code{failed}, or \code{cancelled} state.}

\item{SnapshotClusterIdentifier}{The unique identifier of the cluster the snapshot was created from. This
parameter is required if your IAM user has a policy containing a
snapshot resource element that specifies anything other than * for the
cluster name.

Constraints: Must be the name of valid cluster.}
}
\description{
Deletes the specified manual snapshot. The snapshot must be in the \code{available} state, with no other users authorized to access the snapshot.

See \url{https://www.paws-r-sdk.com/docs/redshift_delete_cluster_snapshot/} for full documentation.
}
\keyword{internal}
