% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_delete_cluster_subnet_group}
\alias{redshift_delete_cluster_subnet_group}
\title{Deletes the specified cluster subnet group}
\usage{
redshift_delete_cluster_subnet_group(ClusterSubnetGroupName)
}
\arguments{
\item{ClusterSubnetGroupName}{[required] The name of the cluster subnet group name to be deleted.}
}
\value{
An empty list.
}
\description{
Deletes the specified cluster subnet group.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_cluster_subnet_group(
  ClusterSubnetGroupName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
