% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_delete_custom_domain_association}
\alias{redshift_delete_custom_domain_association}
\title{Contains information about deleting a custom domain association for a
cluster}
\usage{
redshift_delete_custom_domain_association(ClusterIdentifier)
}
\arguments{
\item{ClusterIdentifier}{[required] The identifier of the cluster to delete a custom domain association for.}
}
\description{
Contains information about deleting a custom domain association for a cluster.

See \url{https://www.paws-r-sdk.com/docs/redshift_delete_custom_domain_association/} for full documentation.
}
\keyword{internal}
