% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_delete_endpoint_access}
\alias{redshift_delete_endpoint_access}
\title{Deletes a Redshift-managed VPC endpoint}
\usage{
redshift_delete_endpoint_access(EndpointName)
}
\arguments{
\item{EndpointName}{[required] The Redshift-managed VPC endpoint to delete.}
}
\description{
Deletes a Redshift-managed VPC endpoint.

See \url{https://paws-r.github.io/docs/redshift/delete_endpoint_access.html} for full documentation.
}
\keyword{internal}
