% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_delete_hsm_client_certificate}
\alias{redshift_delete_hsm_client_certificate}
\title{Deletes the specified HSM client certificate}
\usage{
redshift_delete_hsm_client_certificate(HsmClientCertificateIdentifier)
}
\arguments{
\item{HsmClientCertificateIdentifier}{[required] The identifier of the HSM client certificate to be deleted.}
}
\description{
Deletes the specified HSM client certificate.

See \url{https://paws-r.github.io/docs/redshift/delete_hsm_client_certificate.html} for full documentation.
}
\keyword{internal}
