% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_delete_partner}
\alias{redshift_delete_partner}
\title{Deletes a partner integration from a cluster}
\usage{
redshift_delete_partner(
  AccountId,
  ClusterIdentifier,
  DatabaseName,
  PartnerName
)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID that owns the cluster.}

\item{ClusterIdentifier}{[required] The cluster identifier of the cluster that receives data from the
partner.}

\item{DatabaseName}{[required] The name of the database that receives data from the partner.}

\item{PartnerName}{[required] The name of the partner that is authorized to send data.}
}
\description{
Deletes a partner integration from a cluster. Data can still flow to the cluster until the integration is deleted at the partner's website.

See \url{https://www.paws-r-sdk.com/docs/redshift_delete_partner/} for full documentation.
}
\keyword{internal}
