% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_delete_scheduled_action}
\alias{redshift_delete_scheduled_action}
\title{Deletes a scheduled action}
\usage{
redshift_delete_scheduled_action(ScheduledActionName)
}
\arguments{
\item{ScheduledActionName}{[required] The name of the scheduled action to delete.}
}
\value{
An empty list.
}
\description{
Deletes a scheduled action.
}
\section{Request syntax}{
\preformatted{svc$delete_scheduled_action(
  ScheduledActionName = "string"
)
}
}

\keyword{internal}
