% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_delete_snapshot_copy_grant}
\alias{redshift_delete_snapshot_copy_grant}
\title{Deletes the specified snapshot copy grant}
\usage{
redshift_delete_snapshot_copy_grant(SnapshotCopyGrantName)
}
\arguments{
\item{SnapshotCopyGrantName}{[required] The name of the snapshot copy grant to delete.}
}
\description{
Deletes the specified snapshot copy grant.

See \url{https://www.paws-r-sdk.com/docs/redshift_delete_snapshot_copy_grant/} for full documentation.
}
\keyword{internal}
