% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_delete_snapshot_schedule}
\alias{redshift_delete_snapshot_schedule}
\title{Deletes a snapshot schedule}
\usage{
redshift_delete_snapshot_schedule(ScheduleIdentifier)
}
\arguments{
\item{ScheduleIdentifier}{[required] A unique identifier of the snapshot schedule to delete.}
}
\description{
Deletes a snapshot schedule.

See \url{https://www.paws-r-sdk.com/docs/redshift_delete_snapshot_schedule/} for full documentation.
}
\keyword{internal}
