% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_delete_snapshot_schedule}
\alias{redshift_delete_snapshot_schedule}
\title{Deletes a snapshot schedule}
\usage{
redshift_delete_snapshot_schedule(ScheduleIdentifier)
}
\arguments{
\item{ScheduleIdentifier}{[required] A unique identifier of the snapshot schedule to delete.}
}
\value{
An empty list.
}
\description{
Deletes a snapshot schedule.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_snapshot_schedule(
  ScheduleIdentifier = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
