% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_describe_account_attributes}
\alias{redshift_describe_account_attributes}
\title{Returns a list of attributes attached to an account}
\usage{
redshift_describe_account_attributes(AttributeNames)
}
\arguments{
\item{AttributeNames}{A list of attribute names.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AccountAttributes = list(
    list(
      AttributeName = "string",
      AttributeValues = list(
        list(
          AttributeValue = "string"
        )
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns a list of attributes attached to an account
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_account_attributes(
  AttributeNames = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
