% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_describe_account_attributes}
\alias{redshift_describe_account_attributes}
\title{Returns a list of attributes attached to an account}
\usage{
redshift_describe_account_attributes(AttributeNames)
}
\arguments{
\item{AttributeNames}{A list of attribute names.}
}
\description{
Returns a list of attributes attached to an account
}
\section{Request syntax}{
\preformatted{svc$describe_account_attributes(
  AttributeNames = list(
    "string"
  )
)
}
}

\keyword{internal}
