% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_describe_logging_status}
\alias{redshift_describe_logging_status}
\title{Describes whether information, such as queries and connection attempts,
is being logged for the specified Amazon Redshift cluster}
\usage{
redshift_describe_logging_status(ClusterIdentifier)
}
\arguments{
\item{ClusterIdentifier}{[required] The identifier of the cluster from which to get the logging status.

Example: \code{examplecluster}}
}
\description{
Describes whether information, such as queries and connection attempts, is being logged for the specified Amazon Redshift cluster.

See \url{https://www.paws-r-sdk.com/docs/redshift_describe_logging_status/} for full documentation.
}
\keyword{internal}
