% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_describe_partners}
\alias{redshift_describe_partners}
\title{Returns information about the partner integrations defined for a cluster}
\usage{
redshift_describe_partners(
  AccountId,
  ClusterIdentifier,
  DatabaseName = NULL,
  PartnerName = NULL
)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account ID that owns the cluster.}

\item{ClusterIdentifier}{[required] The cluster identifier of the cluster whose partner integration is being
described.}

\item{DatabaseName}{The name of the database whose partner integration is being described.
If database name is not specified, then all databases in the cluster are
described.}

\item{PartnerName}{The name of the partner that is being described. If partner name is not
specified, then all partner integrations are described.}
}
\description{
Returns information about the partner integrations defined for a cluster.

See \url{https://paws-r.github.io/docs/redshift/describe_partners.html} for full documentation.
}
\keyword{internal}
