% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_describe_storage}
\alias{redshift_describe_storage}
\title{Returns account level backups storage size and provisional storage}
\usage{
redshift_describe_storage()
}
\description{
Returns account level backups storage size and provisional storage.
}
\keyword{internal}
