% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_disable_logging}
\alias{redshift_disable_logging}
\title{Stops logging information, such as queries and connection attempts, for
the specified Amazon Redshift cluster}
\usage{
redshift_disable_logging(ClusterIdentifier)
}
\arguments{
\item{ClusterIdentifier}{[required] The identifier of the cluster on which logging is to be stopped.

Example: \code{examplecluster}}
}
\description{
Stops logging information, such as queries and connection attempts, for the specified Amazon Redshift cluster.

See \url{https://www.paws-r-sdk.com/docs/redshift_disable_logging/} for full documentation.
}
\keyword{internal}
