% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_disable_snapshot_copy}
\alias{redshift_disable_snapshot_copy}
\title{Disables the automatic copying of snapshots from one region to another
region for a specified cluster}
\usage{
redshift_disable_snapshot_copy(ClusterIdentifier)
}
\arguments{
\item{ClusterIdentifier}{[required] The unique identifier of the source cluster that you want to disable
copying of snapshots to a destination region.

Constraints: Must be the valid name of an existing cluster that has
cross-region snapshot copy enabled.}
}
\description{
Disables the automatic copying of snapshots from one region to another region for a specified cluster.

See \url{https://paws-r.github.io/docs/redshift/disable_snapshot_copy.html} for full documentation.
}
\keyword{internal}
