% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_modify_aqua_configuration}
\alias{redshift_modify_aqua_configuration}
\title{This operation is retired}
\usage{
redshift_modify_aqua_configuration(
  ClusterIdentifier,
  AquaConfigurationStatus = NULL
)
}
\arguments{
\item{ClusterIdentifier}{[required] The identifier of the cluster to be modified.}

\item{AquaConfigurationStatus}{This parameter is retired. Amazon Redshift automatically determines
whether to use AQUA (Advanced Query Accelerator).}
}
\description{
This operation is retired. Calling this operation does not change AQUA configuration. Amazon Redshift automatically determines whether to use AQUA (Advanced Query Accelerator).

See \url{https://paws-r.github.io/docs/redshift/modify_aqua_configuration.html} for full documentation.
}
\keyword{internal}
