% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_modify_cluster_snapshot_schedule}
\alias{redshift_modify_cluster_snapshot_schedule}
\title{Modifies a snapshot schedule for a cluster}
\usage{
redshift_modify_cluster_snapshot_schedule(
  ClusterIdentifier,
  ScheduleIdentifier = NULL,
  DisassociateSchedule = NULL
)
}
\arguments{
\item{ClusterIdentifier}{[required] A unique identifier for the cluster whose snapshot schedule you want to
modify.}

\item{ScheduleIdentifier}{A unique alphanumeric identifier for the schedule that you want to
associate with the cluster.}

\item{DisassociateSchedule}{A boolean to indicate whether to remove the assoiciation between the
cluster and the schedule.}
}
\description{
Modifies a snapshot schedule for a cluster.

See \url{https://www.paws-r-sdk.com/docs/redshift_modify_cluster_snapshot_schedule/} for full documentation.
}
\keyword{internal}
