% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_modify_cluster_subnet_group}
\alias{redshift_modify_cluster_subnet_group}
\title{Modifies a cluster subnet group to include the specified list of VPC
subnets}
\usage{
redshift_modify_cluster_subnet_group(
  ClusterSubnetGroupName,
  Description = NULL,
  SubnetIds
)
}
\arguments{
\item{ClusterSubnetGroupName}{[required] The name of the subnet group to be modified.}

\item{Description}{A text description of the subnet group to be modified.}

\item{SubnetIds}{[required] An array of VPC subnet IDs. A maximum of 20 subnets can be modified in a
single request.}
}
\description{
Modifies a cluster subnet group to include the specified list of VPC subnets. The operation replaces the existing list of subnets with the new list of subnets.

See \url{https://www.paws-r-sdk.com/docs/redshift_modify_cluster_subnet_group/} for full documentation.
}
\keyword{internal}
