% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_modify_endpoint_access}
\alias{redshift_modify_endpoint_access}
\title{Modifies a Redshift-managed VPC endpoint}
\usage{
redshift_modify_endpoint_access(EndpointName, VpcSecurityGroupIds)
}
\arguments{
\item{EndpointName}{[required] The endpoint to be modified.}

\item{VpcSecurityGroupIds}{The complete list of VPC security groups associated with the endpoint
after the endpoint is modified.}
}
\value{
A list with the following syntax:\preformatted{list(
  ClusterIdentifier = "string",
  ResourceOwner = "string",
  SubnetGroupName = "string",
  EndpointStatus = "string",
  EndpointName = "string",
  EndpointCreateTime = as.POSIXct(
    "2015-01-01"
  ),
  Port = 123,
  Address = "string",
  VpcSecurityGroups = list(
    list(
      VpcSecurityGroupId = "string",
      Status = "string"
    )
  ),
  VpcEndpoint = list(
    VpcEndpointId = "string",
    VpcId = "string",
    NetworkInterfaces = list(
      list(
        NetworkInterfaceId = "string",
        SubnetId = "string",
        PrivateIpAddress = "string",
        AvailabilityZone = "string"
      )
    )
  )
)
}
}
\description{
Modifies a Redshift-managed VPC endpoint.
}
\section{Request syntax}{
\preformatted{svc$modify_endpoint_access(
  EndpointName = "string",
  VpcSecurityGroupIds = list(
    "string"
  )
)
}
}

\keyword{internal}
