% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_revoke_endpoint_access}
\alias{redshift_revoke_endpoint_access}
\title{Revokes access to a cluster}
\usage{
redshift_revoke_endpoint_access(ClusterIdentifier, Account, VpcIds,
  Force)
}
\arguments{
\item{ClusterIdentifier}{The cluster to revoke access from.}

\item{Account}{The AWS account ID whose access is to be revoked.}

\item{VpcIds}{The virtual private cloud (VPC) identifiers for which access is to be
revoked.}

\item{Force}{Indicates whether to force the revoke action. If true, the
Redshift-managed VPC endpoints associated with the endpoint
authorization are also deleted.}
}
\value{
A list with the following syntax:\preformatted{list(
  Grantor = "string",
  Grantee = "string",
  ClusterIdentifier = "string",
  AuthorizeTime = as.POSIXct(
    "2015-01-01"
  ),
  ClusterStatus = "string",
  Status = "Authorized"|"Revoking",
  AllowedAllVPCs = TRUE|FALSE,
  AllowedVPCs = list(
    "string"
  ),
  EndpointCount = 123
)
}
}
\description{
Revokes access to a cluster.
}
\section{Request syntax}{
\preformatted{svc$revoke_endpoint_access(
  ClusterIdentifier = "string",
  Account = "string",
  VpcIds = list(
    "string"
  ),
  Force = TRUE|FALSE
)
}
}

\keyword{internal}
