% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_convert_recovery_point_to_snapshot}
\alias{redshiftserverless_convert_recovery_point_to_snapshot}
\title{Converts a recovery point to a snapshot}
\usage{
redshiftserverless_convert_recovery_point_to_snapshot(
  recoveryPointId,
  retentionPeriod = NULL,
  snapshotName,
  tags = NULL
)
}
\arguments{
\item{recoveryPointId}{[required] The unique identifier of the recovery point.}

\item{retentionPeriod}{How long to retain the snapshot.}

\item{snapshotName}{[required] The name of the snapshot.}

\item{tags}{An array of \href{https://docs.aws.amazon.com/redshift-serverless/latest/APIReference/API_Tag.html}{Tag objects}
to associate with the created snapshot.}
}
\description{
Converts a recovery point to a snapshot. For more information about recovery points and snapshots, see \href{https://docs.aws.amazon.com/redshift/latest/mgmt/serverless-snapshots-recovery.html}{Working with snapshots and recovery points}.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_convert_recovery_point_to_snapshot/} for full documentation.
}
\keyword{internal}
