% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_delete_endpoint_access}
\alias{redshiftserverless_delete_endpoint_access}
\title{Deletes an Amazon Redshift Serverless managed VPC endpoint}
\usage{
redshiftserverless_delete_endpoint_access(endpointName)
}
\arguments{
\item{endpointName}{[required] The name of the VPC endpoint to delete.}
}
\description{
Deletes an Amazon Redshift Serverless managed VPC endpoint.

See \url{https://paws-r.github.io/docs/redshiftserverless/delete_endpoint_access.html} for full documentation.
}
\keyword{internal}
