% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_get_recovery_point}
\alias{redshiftserverless_get_recovery_point}
\title{Returns information about a recovery point}
\usage{
redshiftserverless_get_recovery_point(recoveryPointId)
}
\arguments{
\item{recoveryPointId}{[required] The unique identifier of the recovery point to return information for.}
}
\description{
Returns information about a recovery point.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_get_recovery_point/} for full documentation.
}
\keyword{internal}
