% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_list_tags_for_resource}
\alias{redshiftserverless_list_tags_for_resource}
\title{Lists the tags assigned to a resource}
\usage{
redshiftserverless_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource to list tags for.}
}
\description{
Lists the tags assigned to a resource.

See \url{https://paws-r.github.io/docs/redshiftserverless/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
