% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_list_workgroups}
\alias{redshiftserverless_list_workgroups}
\title{Returns information about a list of specified workgroups}
\usage{
redshiftserverless_list_workgroups(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{An optional parameter that specifies the maximum number of results to
return. You can use \code{nextToken} to get the next page of results.}

\item{nextToken}{If your initial ListWorkgroups operation returns a \code{nextToken}, you can
include the returned \code{nextToken} in subsequent ListNamespaces
operations, which returns results in the next page.}
}
\description{
Returns information about a list of specified workgroups.

See \url{https://paws-r.github.io/docs/redshiftserverless/list_workgroups.html} for full documentation.
}
\keyword{internal}
