% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_restore_from_recovery_point}
\alias{redshiftserverless_restore_from_recovery_point}
\title{Restore the data from a recovery point}
\usage{
redshiftserverless_restore_from_recovery_point(
  namespaceName,
  recoveryPointId,
  workgroupName
)
}
\arguments{
\item{namespaceName}{[required] The name of the namespace to restore data into.}

\item{recoveryPointId}{[required] The unique identifier of the recovery point to restore from.}

\item{workgroupName}{[required] The name of the workgroup used to restore data.}
}
\description{
Restore the data from a recovery point.

See \url{https://paws-r.github.io/docs/redshiftserverless/restore_from_recovery_point.html} for full documentation.
}
\keyword{internal}
