% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftserverless_operations.R
\name{redshiftserverless_update_snapshot}
\alias{redshiftserverless_update_snapshot}
\title{Updates a snapshot}
\usage{
redshiftserverless_update_snapshot(retentionPeriod = NULL, snapshotName)
}
\arguments{
\item{retentionPeriod}{The new retention period of the snapshot.}

\item{snapshotName}{[required] The name of the snapshot.}
}
\description{
Updates a snapshot.

See \url{https://www.paws-r-sdk.com/docs/redshiftserverless_update_snapshot/} for full documentation.
}
\keyword{internal}
