% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpledb_operations.R
\name{simpledb_domain_metadata}
\alias{simpledb_domain_metadata}
\title{Returns information about the domain, including when the domain was
created, the number of items and attributes in the domain, and the size
of the attribute names and values}
\usage{
simpledb_domain_metadata(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the domain for which to display the metadata of.}
}
\description{
Returns information about the domain, including when the domain was created, the number of items and attributes in the domain, and the size of the attribute names and values.

See \url{https://www.paws-r-sdk.com/docs/simpledb_domain_metadata/} for full documentation.
}
\keyword{internal}
