% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamquery_operations.R
\name{timestreamquery_execute_scheduled_query}
\alias{timestreamquery_execute_scheduled_query}
\title{You can use this API to run a scheduled query manually}
\usage{
timestreamquery_execute_scheduled_query(
  ScheduledQueryArn,
  InvocationTime,
  ClientToken = NULL
)
}
\arguments{
\item{ScheduledQueryArn}{[required] ARN of the scheduled query.}

\item{InvocationTime}{[required] The timestamp in UTC. Query will be run as if it was invoked at this
timestamp.}

\item{ClientToken}{Not used.}
}
\description{
You can use this API to run a scheduled query manually.

See \url{https://www.paws-r-sdk.com/docs/timestreamquery_execute_scheduled_query/} for full documentation.
}
\keyword{internal}
