% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamwrite_operations.R
\name{timestreamwrite_describe_endpoints}
\alias{timestreamwrite_describe_endpoints}
\title{Returns a list of available endpoints to make Timestream API calls
against}
\usage{
timestreamwrite_describe_endpoints()
}
\description{
Returns a list of available endpoints to make Timestream API calls against. This API operation is available through both the Write and Query APIs.

See \url{https://www.paws-r-sdk.com/docs/timestreamwrite_describe_endpoints/} for full documentation.
}
\keyword{internal}
