% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestreamwrite_operations.R
\name{timestreamwrite_tag_resource}
\alias{timestreamwrite_tag_resource}
\title{Associates a set of tags with a Timestream resource}
\usage{
timestreamwrite_tag_resource(ResourceARN, Tags)
}
\arguments{
\item{ResourceARN}{[required] Identifies the Timestream resource to which tags should be added. This
value is an Amazon Resource Name (ARN).}

\item{Tags}{[required] The tags to be assigned to the Timestream resource.}
}
\description{
Associates a set of tags with a Timestream resource. You can then activate these user-defined tags so that they appear on the Billing and Cost Management console for cost allocation tracking.

See \url{https://www.paws-r-sdk.com/docs/timestreamwrite_tag_resource/} for full documentation.
}
\keyword{internal}
